Calculates aggregated deuterium uptake difference for one time point

Function aggregates the differential deuterium uptake values from peptide level into single-amino resolution using `weighted approach` (defined in `vignette("datafiles"))`. For visualization use plot_aggregated_uptake

Usage

calculate_aggregated_diff_uptake(diff_uptake_dat, time_t)

Arguments

diff_uptake_dat

differential uptake data, product of e.q. create_diff_uptake_dataset

time_t

chosen time point

Value

a data.frame object

Examples

diff_uptake_dat <- create_diff_uptake_dataset(alpha_dat)
calculate_aggregated_diff_uptake(diff_uptake_dat, time_t = 5)
#>     position aa diff_frac_deut_uptake err_diff_frac_deut_uptake
#> 1          1  G             6.9909186                 1.2323610
#> 2          2  F             6.0759840                 0.9869864
#> 3          3  G             5.5274366                 0.7674040
#> 4          4  D             5.5274366                 0.7674040
#> 5          5  L             5.7528855                 0.3914090
#> 6          6  K             5.7528855                 0.3914090
#> 7          7  S             5.7528855                 0.3914090
#> 8          8  P             5.7528855                 0.3914090
#> 9          9  A             6.9376397                 0.5824853
#> 10        10  G             6.9376397                 0.5824853
#> 11        11  L             6.1447267                 0.7650383
#> 12        12  Q             6.5363571                 0.8066338
#> 13        13  V             6.5363571                 0.8066338
#> 14        14  L             6.5363571                 0.8066338
#> 15        15  N             7.1177847                 0.3962519
#> 16        16  D             7.1177847                 0.3962519
#> 17        17  Y             5.2498642                 0.3191644
#> 18        18  L             6.2590951                 0.3845800
#> 19        19  A             5.9178512                 0.3543794
#> 20        20  D             5.9178512                 0.3543794
#> 21        21  K             3.4312828                 0.1957034
#> 22        22  S             2.0477185                 0.2463916
#> 23        23  Y             2.0255660                 0.2800342
#> 24        24  I             1.9354199                 0.3747816
#> 25        25  E             1.9354199                 0.3747816
#> 26        26  G             1.9354199                 0.3747816
#> 27        27  Y             3.3125984                 0.3799765
#> 28        28  V             3.3125984                 0.3799765
#> 29        29  P             3.3125984                 0.3799765
#> 30        30  S             3.8357800                 0.3658725
#> 31        31  Q             3.8357800                 0.3658725
#> 32        32  A             3.8357800                 0.3658725
#> 33        33  D             3.9668778                 0.2771063
#> 34        34  V             2.5134958                 1.2720380
#> 35        35  A             2.5134958                 1.2720380
#> 36        36  V             2.5134958                 1.2720380
#> 37        37  F             1.0704636                 0.8310383
#> 38        38  E             2.0241975                 0.5833559
#> 39        39  A             1.9570643                 0.6142379
#> 40        40  V             1.6129451                 0.5635091
#> 41        41  S             1.6129451                 0.5635091
#> 42        42  S             1.6129451                 0.5635091
#> 43        43  P             1.6129451                 0.5635091
#> 44        44  P             1.6129451                 0.5635091
#> 45        45  P             1.6129451                 0.5635091
#> 46        46  A             1.6129451                 0.5635091
#> 47        47  D             1.6129451                 0.5635091
#> 48        48  L             2.0533111                 0.4918354
#> 49        49  C             3.9737234                 0.4926546
#> 50        50  H             3.9737234                 0.4926546
#> 51        51  A             6.1449142                 0.5969541
#> 52        52  L             7.2339028                 0.4263357
#> 53        53  R             7.2339028                 0.4263357
#> 54        54  W             7.2339028                 0.4263357
#> 55        55  Y             7.2339028                 0.4263357
#> 56        56  N             7.2339028                 0.4263357
#> 57        57  H             7.2339028                 0.4263357
#> 58        58  I             1.6814433                 1.0319907
#> 59        59  K             1.6814433                 1.0319907
#> 60        60  S             1.6814433                 1.0319907
#> 61        61  Y             1.6814433                 1.0319907
#> 62        62  E             1.6814433                 1.0319907
#> 63        63  K             1.6814433                 1.0319907
#> 64        64  E             1.2260523                 0.5499892
#> 65        65  K             1.2260523                 0.5499892
#> 66        66  A             1.2115643                 0.4606580
#> 67        67  S             1.2115643                 0.4606580
#> 68        68  L             1.2115643                 0.4606580
#> 69        69  P             1.2115643                 0.4606580
#> 70        70  G             1.2115643                 0.4606580
#> 71        71  V             1.2115643                 0.4606580
#> 72        72  K             1.2115643                 0.4606580
#> 73        73  K             1.2115643                 0.4606580
#> 74        74  A             1.4462670                 0.4585428
#> 75        75  L             1.4462670                 0.4585428
#> 76        76  G             1.4462670                 0.4585428
#> 77        77  K             1.4462670                 0.4585428
#> 78        78  Y             1.4462670                 0.4585428
#> 79        79  G             1.3429057                 0.4530602
#> 80        80  P             1.3429057                 0.4530602
#> 81        81  A             1.3429057                 0.4530602
#> 82        82  D             1.3429057                 0.4530602
#> 83        83  V             2.4147884                 0.5611923
#> 84        84  E             2.4147884                 0.5611923
#> 85        85  D             4.4107039                 1.0781856
#> 86        86  T             4.7053248                 0.8585749
#> 87        87  T             4.7053248                 0.8585749
#> 88        88  G             4.7053248                 0.8585749
#> 89        89  S             4.7053248                 0.8585749
#> 90        90  G             4.7053248                 0.8585749
#> 91        91  A             4.7053248                 0.8585749
#> 92        92  T             4.7053248                 0.8585749
#> 93        93  D             4.7053248                 0.8585749
#> 94        94  S             4.7053248                 0.8585749
#> 95        95  K             4.7053248                 0.8585749
#> 96        96  D             4.7053248                 0.8585749
#> 97        97  D             8.1809368                 1.0485531
#> 98        98  D             8.1809368                 1.0485531
#> 99        99  D             8.1809368                 1.0485531
#> 100      100  I             8.1809368                 1.0485531
#> 101      101  D             8.1809368                 1.0485531
#> 102      102  L             5.2062863                 1.0333211
#> 103      103  F             1.2161895                 1.3955704
#> 104      104  G             1.2161895                 1.3955704
#> 105      105  S             1.2161895                 1.3955704
#> 106      106  D             1.2161895                 1.3955704
#> 107      107  D             1.0548358                 1.2861063
#> 108      108  E            -0.3410352                 0.4393945
#> 109      109  E            -0.3410352                 0.4393945
#> 110      110  E            -0.3410352                 0.4393945
#> 111      111  S            -0.3410352                 0.4393945
#> 112      112  E            -0.3410352                 0.4393945
#> 113      113  E            -0.3410352                 0.4393945
#> 114      114  A            -0.3410352                 0.4393945
#> 115      115  K            -0.3410352                 0.4393945
#> 116      116  R            -0.3410352                 0.4393945
#> 117      117  L            -0.3410352                 0.4393945
#> 118      118  R            -0.3410352                 0.4393945
#> 119      119  E            -0.3410352                 0.4393945
#> 120      120  E            -0.1900336                 0.5297256
#> 121      121  R            -0.3355108                 0.6874587
#> 122      122  L            -0.3355108                 0.6874587
#> 123      123  A             0.6242733                 0.6408505
#> 124      124  Q             0.6242733                 0.6408505
#> 125      125  Y             3.1683518                 0.9391278
#> 126      126  E             3.1683518                 0.9391278
#> 127      127  S             3.1683518                 0.9391278
#> 128      128  K             3.1683518                 0.9391278
#> 129      129  K             3.1683518                 0.9391278
#> 130      130  A             3.1683518                 0.9391278
#> 131      131  K             3.1683518                 0.9391278
#> 132      132  K             3.1683518                 0.9391278
#> 133      133  P             3.1683518                 0.9391278
#> 134      134  A             3.1683518                 0.9391278
#> 135      135  L             3.1683518                 0.9391278
#> 136      136  V            13.9210141                 0.6842206
#> 137      137  A            13.1348738                 0.9333198
#> 138      138  K            13.1348738                 0.9333198
#> 139      139  S            13.1348738                 0.9333198
#> 140      140  S            13.1348738                 0.9333198
#> 141      141  I            13.1348738                 0.9333198
#> 142      142  L            13.1348738                 0.9333198
#> 143      143  L             9.3138443                 0.8185149
#> 144      144  D             7.4923600                 0.8640612
#> 145      145  V             7.4923600                 0.8640612
#> 146      146  K             7.4923600                 0.8640612
#> 147      147  P             7.4923600                 0.8640612
#> 148      148  W             7.4923600                 0.8640612
#> 149      149  D             7.4923600                 0.8640612
#> 150      150  D             7.4923600                 0.8640612
#> 151      151  E             6.9544018                 0.9081452
#> 152      152  T             6.9544018                 0.9081452
#> 153      153  D             6.9544018                 0.9081452
#> 154      154  M             6.5498833                 0.5015676
#> 155      155  A             8.8227406                 0.7253119
#> 156      156  K             8.8227406                 0.7253119
#> 157      157  L             8.8227406                 0.7253119
#> 158      158  E             8.8227406                 0.7253119
#> 159      159  E             8.8227406                 0.7253119
#> 160      160  C             8.6544869                 0.8698117
#> 161      161  V             7.8995109                 0.6587245
#> 162      162  R             7.8995109                 0.6587245
#> 163      163  S             6.1400545                 0.4959744
#> 164      164  I             6.1400545                 0.4959744
#> 165      165  Q             6.1400545                 0.4959744
#> 166      166  A             6.0353324                 0.4780480
#> 167      167  D             4.8446670                 0.3995575
#> 168      168  G             4.8446670                 0.3995575
#> 169      169  L             4.6114595                 0.3626169
#> 170      170  V             4.8090691                 0.4236054
#> 171      171  W             4.8090691                 0.4236054
#> 172      172  G             4.8090691                 0.4236054
#> 173      173  S             4.8090691                 0.4236054
#> 174      174  S             4.8090691                 0.4236054
#> 175      175  K             4.4446293                 0.4375109
#> 176      176  L             4.4446293                 0.4375109
#> 177      177  V             6.0011595                 0.4131336
#> 178      178  P             6.0011595                 0.4131336
#> 179      179  V             6.0011595                 0.4131336
#> 180      180  G             6.0011595                 0.4131336
#> 181      181  Y             6.0011595                 0.4131336
#> 182      182  G             6.8772121                 0.5115313
#> 183      183  I             7.2054625                 0.5639662
#> 184      184  K             7.2054625                 0.5639662
#> 185      185  K             8.7092399                 0.8827517
#> 186      186  L             9.3370866                 0.9776491
#> 187      187  Q             9.3370866                 0.9776491
#> 188      188  I             9.3370866                 0.9776491
#> 189      189  Q             9.3370866                 0.9776491
#> 190      190  C             8.4910587                 0.6900596
#> 191      191  V             7.3182540                 0.3636827
#> 192      192  V             7.3182540                 0.3636827
#> 193      193  E             7.3182540                 0.3636827
#> 194      194  D             6.6665448                 0.3997945
#> 195      195  D             6.4975161                 0.3453443
#> 196      196  K             6.4975161                 0.3453443
#> 197      197  V             6.3796253                 0.4630638
#> 198      198  G             6.3796253                 0.4630638
#> 199      199  T             6.3796253                 0.4630638
#> 200      200  D             6.3796253                 0.4630638
#> 201      201  M             6.3796253                 0.4630638
#> 202      202  L             8.6642687                 0.5105340
#> 203      203  E            12.2014488                 0.5924221
#> 204      204  E            11.6516832                 0.5614844
#> 205      205  Q            12.5469313                 0.5524433
#> 206      206  I            12.5469313                 0.5524433
#> 207      207  T            12.5469313                 0.5524433
#> 208      208  A            12.5567849                 0.6139466
#> 209      209  F            12.2430890                 0.6542324
#> 210      210  E            10.6311764                 0.6434289
#> 211      211  D             8.5329563                 0.5397469
#> 212      212  Y             8.3575435                 0.7699638
#> 213      213  V             8.3575435                 0.7699638
#> 214      214  Q             8.3575435                 0.7699638
#> 215      215  S             8.3575435                 0.7699638
#> 216      216  M             8.3575435                 0.7699638
#> 217      217  D             7.0696573                 0.8988115
#> 218      218  V             6.9960476                 0.2662618
#> 219      219  A             5.9133185                 0.3920682
#> 220      220  A             5.9133185                 0.3920682
#> 221      221  F             5.9133185                 0.3920682
#>     diff_deut_uptake err_diff_deut_uptake diff_theo_frac_deut_uptake
#> 1        0.394655763          0.057318502                  4.8370964
#> 2        0.306330808          0.039630642                  3.9393247
#> 3        0.268968579          0.029185483                  3.6852798
#> 4        0.268968579          0.029185483                  3.6852798
#> 5        0.213570566          0.014213174                  4.1165899
#> 6        0.213570566          0.014213174                  4.1165899
#> 7        0.213570566          0.014213174                  4.1165899
#> 8        0.213570566          0.014213174                  4.1165899
#> 9        0.252768361          0.019687996                  5.0970362
#> 10       0.252768361          0.019687996                  5.0970362
#> 11       0.250376852          0.025978700                  4.4297816
#> 12       0.189629410          0.025747445                  4.4645569
#> 13       0.189629410          0.025747445                  4.4645569
#> 14       0.189629410          0.025747445                  4.4645569
#> 15       0.197283250          0.010747723                  4.3612065
#> 16       0.197283250          0.010747723                  4.3612065
#> 17       0.164961212          0.009187371                  3.3534052
#> 18       0.197333848          0.011272242                  3.9961208
#> 19       0.198306669          0.011003254                  3.8007184
#> 20       0.198306669          0.011003254                  3.8007184
#> 21       0.135057912          0.007805785                  2.2769859
#> 22       0.086753706          0.006323913                  1.3828747
#> 23       0.097223015          0.008721448                  1.3803166
#> 24       0.092428166          0.012003785                  1.3548282
#> 25       0.092428166          0.012003785                  1.3548282
#> 26       0.092428166          0.012003785                  1.3548282
#> 27       0.147346328          0.013104198                  2.4935694
#> 28       0.147346328          0.013104198                  2.4935694
#> 29       0.147346328          0.013104198                  2.4935694
#> 30       0.178581807          0.015332810                  2.9253222
#> 31       0.178581807          0.015332810                  2.9253222
#> 32       0.178581807          0.015332810                  2.9253222
#> 33       0.184692979          0.010841865                  3.0422327
#> 34       0.066892333          0.033842324                  1.8446975
#> 35       0.066892333          0.033842324                  1.8446975
#> 36       0.066892333          0.033842324                  1.8446975
#> 37       0.034655812          0.019367325                  0.6330033
#> 38       0.097604559          0.014749190                  1.3410891
#> 39       0.104832032          0.013798172                  1.2784330
#> 40       0.089567007          0.010499214                  1.0681073
#> 41       0.089567007          0.010499214                  1.0681073
#> 42       0.089567007          0.010499214                  1.0681073
#> 43       0.089567007          0.010499214                  1.0681073
#> 44       0.089567007          0.010499214                  1.0681073
#> 45       0.089567007          0.010499214                  1.0681073
#> 46       0.089567007          0.010499214                  1.0681073
#> 47       0.089567007          0.010499214                  1.0681073
#> 48       0.116371857          0.011022389                  1.4489221
#> 49       0.221997295          0.020571955                  2.7624439
#> 50       0.221997295          0.020571955                  2.7624439
#> 51       0.276453156          0.029972474                  4.0130314
#> 52       0.213492666          0.012991009                  4.2214365
#> 53       0.213492666          0.012991009                  4.2214365
#> 54       0.213492666          0.012991009                  4.2214365
#> 55       0.213492666          0.012991009                  4.2214365
#> 56       0.213492666          0.012991009                  4.2214365
#> 57       0.213492666          0.012991009                  4.2214365
#> 58       0.200314499          0.122068001                  1.2649609
#> 59       0.200314499          0.122068001                  1.2649609
#> 60       0.200314499          0.122068001                  1.2649609
#> 61       0.200314499          0.122068001                  1.2649609
#> 62       0.200314499          0.122068001                  1.2649609
#> 63       0.200314499          0.122068001                  1.2649609
#> 64       0.143845036          0.049682276                  0.9921975
#> 65       0.143845036          0.049682276                  0.9921975
#> 66       0.124989123          0.041624910                  0.9173807
#> 67       0.124989123          0.041624910                  0.9173807
#> 68       0.124989123          0.041624910                  0.9173807
#> 69       0.124989123          0.041624910                  0.9173807
#> 70       0.124989123          0.041624910                  0.9173807
#> 71       0.124989123          0.041624910                  0.9173807
#> 72       0.124989123          0.041624910                  0.9173807
#> 73       0.124989123          0.041624910                  0.9173807
#> 74       0.131526002          0.038348947                  1.1259303
#> 75       0.131526002          0.038348947                  1.1259303
#> 76       0.131526002          0.038348947                  1.1259303
#> 77       0.131526002          0.038348947                  1.1259303
#> 78       0.131526002          0.038348947                  1.1259303
#> 79       0.088384343          0.023661401                  0.9291945
#> 80       0.088384343          0.023661401                  0.9291945
#> 81       0.088384343          0.023661401                  0.9291945
#> 82       0.088384343          0.023661401                  0.9291945
#> 83       0.192983147          0.027691169                  1.5666578
#> 84       0.192983147          0.027691169                  1.5666578
#> 85       0.372932852          0.047114052                  2.2916752
#> 86       0.378149169          0.034575924                  2.4280330
#> 87       0.378149169          0.034575924                  2.4280330
#> 88       0.378149169          0.034575924                  2.4280330
#> 89       0.378149169          0.034575924                  2.4280330
#> 90       0.378149169          0.034575924                  2.4280330
#> 91       0.378149169          0.034575924                  2.4280330
#> 92       0.378149169          0.034575924                  2.4280330
#> 93       0.378149169          0.034575924                  2.4280330
#> 94       0.378149169          0.034575924                  2.4280330
#> 95       0.378149169          0.034575924                  2.4280330
#> 96       0.378149169          0.034575924                  2.4280330
#> 97       0.304097053          0.018049063                  3.9575733
#> 98       0.304097053          0.018049063                  3.9575733
#> 99       0.304097053          0.018049063                  3.9575733
#> 100      0.304097053          0.018049063                  3.9575733
#> 101      0.304097053          0.018049063                  3.9575733
#> 102      0.195531866          0.015923348                  2.5211654
#> 103      0.020143823          0.012750099                  0.5815545
#> 104      0.020143823          0.012750099                  0.5815545
#> 105      0.020143823          0.012750099                  0.5815545
#> 106      0.020143823          0.012750099                  0.5815545
#> 107      0.006507185          0.019851043                  0.4366310
#> 108     -0.025176909          0.040310768                 -0.2652365
#> 109     -0.025176909          0.040310768                 -0.2652365
#> 110     -0.025176909          0.040310768                 -0.2652365
#> 111     -0.025176909          0.040310768                 -0.2652365
#> 112     -0.025176909          0.040310768                 -0.2652365
#> 113     -0.025176909          0.040310768                 -0.2652365
#> 114     -0.025176909          0.040310768                 -0.2652365
#> 115     -0.025176909          0.040310768                 -0.2652365
#> 116     -0.025176909          0.040310768                 -0.2652365
#> 117     -0.025176909          0.040310768                 -0.2652365
#> 118     -0.025176909          0.040310768                 -0.2652365
#> 119     -0.025176909          0.040310768                 -0.2652365
#> 120     -0.025171291          0.050191574                 -0.2898080
#> 121     -0.050201684          0.065128064                 -0.4440677
#> 122     -0.050201684          0.065128064                 -0.4440677
#> 123      0.024653075          0.056224166                  0.2548388
#> 124      0.024653075          0.056224166                  0.2548388
#> 125      0.183832668          0.036722337                  2.1218501
#> 126      0.183832668          0.036722337                  2.1218501
#> 127      0.183832668          0.036722337                  2.1218501
#> 128      0.183832668          0.036722337                  2.1218501
#> 129      0.183832668          0.036722337                  2.1218501
#> 130      0.183832668          0.036722337                  2.1218501
#> 131      0.183832668          0.036722337                  2.1218501
#> 132      0.183832668          0.036722337                  2.1218501
#> 133      0.183832668          0.036722337                  2.1218501
#> 134      0.183832668          0.036722337                  2.1218501
#> 135      0.183832668          0.036722337                  2.1218501
#> 136      0.524109778          0.020930016                  9.0578279
#> 137      0.476139977          0.027876196                  8.9827726
#> 138      0.476139977          0.027876196                  8.9827726
#> 139      0.476139977          0.027876196                  8.9827726
#> 140      0.476139977          0.027876196                  8.9827726
#> 141      0.476139977          0.027876196                  8.9827726
#> 142      0.476139977          0.027876196                  8.9827726
#> 143      0.367798252          0.028105468                  5.6169257
#> 144      0.289397698          0.029247571                  4.1333703
#> 145      0.289397698          0.029247571                  4.1333703
#> 146      0.289397698          0.029247571                  4.1333703
#> 147      0.289397698          0.029247571                  4.1333703
#> 148      0.289397698          0.029247571                  4.1333703
#> 149      0.289397698          0.029247571                  4.1333703
#> 150      0.289397698          0.029247571                  4.1333703
#> 151      0.309083560          0.036976693                  3.7882811
#> 152      0.309083560          0.036976693                  3.7882811
#> 153      0.309083560          0.036976693                  3.7882811
#> 154      0.231567117          0.015357718                  4.2915741
#> 155      0.243312863          0.018828470                  6.1446659
#> 156      0.243312863          0.018828470                  6.1446659
#> 157      0.243312863          0.018828470                  6.1446659
#> 158      0.243312863          0.018828470                  6.1446659
#> 159      0.243312863          0.018828470                  6.1446659
#> 160      0.458231771          0.042270686                  6.6308105
#> 161      0.450590091          0.031653181                  5.9535155
#> 162      0.450590091          0.031653181                  5.9535155
#> 163      0.388753040          0.023745647                  4.6064265
#> 164      0.388753040          0.023745647                  4.6064265
#> 165      0.388753040          0.023745647                  4.6064265
#> 166      0.391770438          0.023651493                  4.5006825
#> 167      0.347067110          0.020723009                  3.4196028
#> 168      0.347067110          0.020723009                  3.4196028
#> 169      0.341744190          0.019170106                  3.2285961
#> 170      0.394930144          0.024024105                  3.2981668
#> 171      0.394930144          0.024024105                  3.2981668
#> 172      0.394930144          0.024024105                  3.2981668
#> 173      0.394930144          0.024024105                  3.2981668
#> 174      0.394930144          0.024024105                  3.2981668
#> 175      0.443486267          0.026653631                  3.1688658
#> 176      0.443486267          0.026653631                  3.1688658
#> 177      0.593116976          0.031777109                  4.5120175
#> 178      0.593116976          0.031777109                  4.5120175
#> 179      0.593116976          0.031777109                  4.5120175
#> 180      0.593116976          0.031777109                  4.5120175
#> 181      0.593116976          0.031777109                  4.5120175
#> 182      0.726924996          0.042990013                  5.2539014
#> 183      0.782009360          0.048705381                  5.5519228
#> 184      0.782009360          0.048705381                  5.5519228
#> 185      0.598561598          0.039577316                  6.8581155
#> 186      0.620572174          0.040951928                  7.4436900
#> 187      0.620572174          0.040951928                  7.4436900
#> 188      0.620572174          0.040951928                  7.4436900
#> 189      0.620572174          0.040951928                  7.4436900
#> 190      0.559091753          0.029090530                  6.4129228
#> 191      0.429851959          0.017832099                  4.2630370
#> 192      0.429851959          0.017832099                  4.2630370
#> 193      0.429851959          0.017832099                  4.2630370
#> 194      0.359730596          0.013830332                  3.9878513
#> 195      0.314410989          0.012051723                  3.9745272
#> 196      0.314410989          0.012051723                  3.9745272
#> 197      0.291531814          0.014619397                  3.9530097
#> 198      0.291531814          0.014619397                  3.9530097
#> 199      0.291531814          0.014619397                  3.9530097
#> 200      0.291531814          0.014619397                  3.9530097
#> 201      0.291531814          0.014619397                  3.9530097
#> 202      0.370170017          0.018138522                  5.9472229
#> 203      0.496184658          0.023505286                  8.8020419
#> 204      0.449788832          0.021595053                  8.4642373
#> 205      0.441961000          0.019265557                  8.9188096
#> 206      0.441961000          0.019265557                  8.9188096
#> 207      0.441961000          0.019265557                  8.9188096
#> 208      0.446926351          0.021539685                  8.8972951
#> 209      0.440758741          0.023244887                  8.5202614
#> 210      0.466460393          0.027013356                  7.3764991
#> 211      0.336349565          0.019486165                  5.7604695
#> 212      0.314377532          0.022777337                  5.6154066
#> 213      0.314377532          0.022777337                  5.6154066
#> 214      0.314377532          0.022777337                  5.6154066
#> 215      0.314377532          0.022777337                  5.6154066
#> 216      0.314377532          0.022777337                  5.6154066
#> 217      0.239992285          0.024044352                  4.5097693
#> 218      0.230672444          0.003906933                  4.2586387
#> 219      0.117271000          0.005726051                  3.2339957
#> 220      0.117271000          0.005726051                  3.2339957
#> 221      0.117271000          0.005726051                  3.2339957
#>     err_diff_theo_frac_deut_uptake diff_theo_deut_uptake
#> 1                        0.7025239           0.394655763
#> 2                        0.5192063           0.306330808
#> 3                        0.4059179           0.268968579
#> 4                        0.4059179           0.268968579
#> 5                        0.2155591           0.213570566
#> 6                        0.2155591           0.213570566
#> 7                        0.2155591           0.213570566
#> 8                        0.2155591           0.213570566
#> 9                        0.3937032           0.252768361
#> 10                       0.3937032           0.252768361
#> 11                       0.5189953           0.250376852
#> 12                       0.5711172           0.189629410
#> 13                       0.5711172           0.189629410
#> 14                       0.5711172           0.189629410
#> 15                       0.2364717           0.197283250
#> 16                       0.2364717           0.197283250
#> 17                       0.1925369           0.164961212
#> 18                       0.2357431           0.197333848
#> 19                       0.2176136           0.198306669
#> 20                       0.2176136           0.198306669
#> 21                       0.1206626           0.135057912
#> 22                       0.1158861           0.086753706
#> 23                       0.1386991           0.097223015
#> 24                       0.2193912           0.092428166
#> 25                       0.2193912           0.092428166
#> 26                       0.2193912           0.092428166
#> 27                       0.2403175           0.147346328
#> 28                       0.2403175           0.147346328
#> 29                       0.2403175           0.147346328
#> 30                       0.2462175           0.178581807
#> 31                       0.2462175           0.178581807
#> 32                       0.2462175           0.178581807
#> 33                       0.1798370           0.184692979
#> 34                       0.9332736           0.066892333
#> 35                       0.9332736           0.066892333
#> 36                       0.9332736           0.066892333
#> 37                       0.3770572           0.034655812
#> 38                       0.2663872           0.098009416
#> 39                       0.1726326           0.105271337
#> 40                       0.1372217           0.089892401
#> 41                       0.1372217           0.089892401
#> 42                       0.1372217           0.089892401
#> 43                       0.1372217           0.089892401
#> 44                       0.1372217           0.089892401
#> 45                       0.1372217           0.089892401
#> 46                       0.1372217           0.089892401
#> 47                       0.1372217           0.089892401
#> 48                       0.1393939           0.116736666
#> 49                       0.2207342           0.222786932
#> 50                       0.2207342           0.222786932
#> 51                       0.3591245           0.277656413
#> 52                       0.2515752           0.213492666
#> 53                       0.2515752           0.213492666
#> 54                       0.2515752           0.213492666
#> 55                       0.2515752           0.213492666
#> 56                       0.2515752           0.213492666
#> 57                       0.2515752           0.213492666
#> 58                       0.7200605           0.197371990
#> 59                       0.7200605           0.197371990
#> 60                       0.7200605           0.197371990
#> 61                       0.7200605           0.197371990
#> 62                       0.7200605           0.197371990
#> 63                       0.7200605           0.197371990
#> 64                       0.2975538           0.142700824
#> 65                       0.2975538           0.142700824
#> 66                       0.2537146           0.130136783
#> 67                       0.2537146           0.130136783
#> 68                       0.2537146           0.130136783
#> 69                       0.2537146           0.130136783
#> 70                       0.2537146           0.130136783
#> 71                       0.2537146           0.130136783
#> 72                       0.2537146           0.130136783
#> 73                       0.2537146           0.130136783
#> 74                       0.2523694           0.136136321
#> 75                       0.2523694           0.136136321
#> 76                       0.2523694           0.136136321
#> 77                       0.2523694           0.136136321
#> 78                       0.2523694           0.136136321
#> 79                       0.2176280           0.096196097
#> 80                       0.2176280           0.096196097
#> 81                       0.2176280           0.096196097
#> 82                       0.2176280           0.096196097
#> 83                       0.2582920           0.192860643
#> 84                       0.2582920           0.192860643
#> 85                       0.2866142           0.372932852
#> 86                       0.2173342           0.378149169
#> 87                       0.2173342           0.378149169
#> 88                       0.2173342           0.378149169
#> 89                       0.2173342           0.378149169
#> 90                       0.2173342           0.378149169
#> 91                       0.2173342           0.378149169
#> 92                       0.2173342           0.378149169
#> 93                       0.2173342           0.378149169
#> 94                       0.2173342           0.378149169
#> 95                       0.2173342           0.378149169
#> 96                       0.2173342           0.378149169
#> 97                       0.2070468           0.304097053
#> 98                       0.2070468           0.304097053
#> 99                       0.2070468           0.304097053
#> 100                      0.2070468           0.304097053
#> 101                      0.2070468           0.304097053
#> 102                      0.2725770           0.195531866
#> 103                      0.2510838           0.017955789
#> 104                      0.2510838           0.017955789
#> 105                      0.2510838           0.017955789
#> 106                      0.2510838           0.017955789
#> 107                      0.2520393           0.001795709
#> 108                      0.2616046          -0.041262093
#> 109                      0.2616046          -0.041262093
#> 110                      0.2616046          -0.041262093
#> 111                      0.2616046          -0.041262093
#> 112                      0.2616046          -0.041262093
#> 113                      0.2616046          -0.041262093
#> 114                      0.2616046          -0.041262093
#> 115                      0.2616046          -0.041262093
#> 116                      0.2616046          -0.041262093
#> 117                      0.2616046          -0.041262093
#> 118                      0.2616046          -0.041262093
#> 119                      0.2616046          -0.041262093
#> 120                      0.3249333          -0.045575273
#> 121                      0.4215685          -0.069657142
#> 122                      0.4215685          -0.069657142
#> 123                      0.3977109           0.008706533
#> 124                      0.3977109           0.008706533
#> 125                      0.4235520           0.183832668
#> 126                      0.4235520           0.183832668
#> 127                      0.4235520           0.183832668
#> 128                      0.4235520           0.183832668
#> 129                      0.4235520           0.183832668
#> 130                      0.4235520           0.183832668
#> 131                      0.4235520           0.183832668
#> 132                      0.4235520           0.183832668
#> 133                      0.4235520           0.183832668
#> 134                      0.4235520           0.183832668
#> 135                      0.4235520           0.183832668
#> 136                      0.3657008           0.524109778
#> 137                      0.5906697           0.476139977
#> 138                      0.5906697           0.476139977
#> 139                      0.5906697           0.476139977
#> 140                      0.5906697           0.476139977
#> 141                      0.5906697           0.476139977
#> 142                      0.5906697           0.476139977
#> 143                      0.4390790           0.362529811
#> 144                      0.4530966           0.283699180
#> 145                      0.4530966           0.283699180
#> 146                      0.4530966           0.283699180
#> 147                      0.4530966           0.283699180
#> 148                      0.4530966           0.283699180
#> 149                      0.4530966           0.283699180
#> 150                      0.4530966           0.283699180
#> 151                      0.4532047           0.309083560
#> 152                      0.4532047           0.309083560
#> 153                      0.4532047           0.309083560
#> 154                      0.3125891           0.231567117
#> 155                      0.4953631           0.243312863
#> 156                      0.4953631           0.243312863
#> 157                      0.4953631           0.243312863
#> 158                      0.4953631           0.243312863
#> 159                      0.4953631           0.243312863
#> 160                      0.6069933           0.474643789
#> 161                      0.4491722           0.462213825
#> 162                      0.4491722           0.462213825
#> 163                      0.3268855           0.396885289
#> 164                      0.3268855           0.396885289
#> 165                      0.3268855           0.396885289
#> 166                      0.3120608           0.399448420
#> 167                      0.2416427           0.347067110
#> 168                      0.2416427           0.347067110
#> 169                      0.2090493           0.341744190
#> 170                      0.2336560           0.394930144
#> 171                      0.2336560           0.394930144
#> 172                      0.2336560           0.394930144
#> 173                      0.2336560           0.394930144
#> 174                      0.2336560           0.394930144
#> 175                      0.2464675           0.440433427
#> 176                      0.2464675           0.440433427
#> 177                      0.2531733           0.588601222
#> 178                      0.2531733           0.588601222
#> 179                      0.2531733           0.588601222
#> 180                      0.2531733           0.588601222
#> 181                      0.2531733           0.588601222
#> 182                      0.3257279           0.720042375
#> 183                      0.3646403           0.773150625
#> 184                      0.3646403           0.773150625
#> 185                      0.6696483           0.599950323
#> 186                      0.7552821           0.622157101
#> 187                      0.7552821           0.622157101
#> 188                      0.7552821           0.622157101
#> 189                      0.7552821           0.622157101
#> 190                      0.5271649           0.561722781
#> 191                      0.1713823           0.431791352
#> 192                      0.1713823           0.431791352
#> 193                      0.1713823           0.431791352
#> 194                      0.1574253           0.360944804
#> 195                      0.1634238           0.321528753
#> 196                      0.1634238           0.321528753
#> 197                      0.2705934           0.297488315
#> 198                      0.2705934           0.297488315
#> 199                      0.2705934           0.297488315
#> 200                      0.2705934           0.297488315
#> 201                      0.2705934           0.297488315
#> 202                      0.3402071           0.370456608
#> 203                      0.4237829           0.495551907
#> 204                      0.4051083           0.449216638
#> 205                      0.3886634           0.441486181
#> 206                      0.3886634           0.441486181
#> 207                      0.3886634           0.441486181
#> 208                      0.4314656           0.446376701
#> 209                      0.4533595           0.440091398
#> 210                      0.4448619           0.465302394
#> 211                      0.3581205           0.335276686
#> 212                      0.4655437           0.313524195
#> 213                      0.4655437           0.313524195
#> 214                      0.4655437           0.313524195
#> 215                      0.4655437           0.313524195
#> 216                      0.4655437           0.313524195
#> 217                      0.5183704           0.239992285
#> 218                      0.1058949           0.230672444
#> 219                      0.1579080           0.117271000
#> 220                      0.1579080           0.117271000
#> 221                      0.1579080           0.117271000
#>     err_diff_theo_deut_uptake Exposure
#> 1                 0.057318502        5
#> 2                 0.039630642        5
#> 3                 0.029185483        5
#> 4                 0.029185483        5
#> 5                 0.014213174        5
#> 6                 0.014213174        5
#> 7                 0.014213174        5
#> 8                 0.014213174        5
#> 9                 0.019687996        5
#> 10                0.019687996        5
#> 11                0.025978700        5
#> 12                0.025747445        5
#> 13                0.025747445        5
#> 14                0.025747445        5
#> 15                0.010747723        5
#> 16                0.010747723        5
#> 17                0.009187371        5
#> 18                0.011272242        5
#> 19                0.011003254        5
#> 20                0.011003254        5
#> 21                0.007805785        5
#> 22                0.006323913        5
#> 23                0.008721448        5
#> 24                0.012003785        5
#> 25                0.012003785        5
#> 26                0.012003785        5
#> 27                0.013104198        5
#> 28                0.013104198        5
#> 29                0.013104198        5
#> 30                0.015332810        5
#> 31                0.015332810        5
#> 32                0.015332810        5
#> 33                0.010841865        5
#> 34                0.033842324        5
#> 35                0.033842324        5
#> 36                0.033842324        5
#> 37                0.019367325        5
#> 38                0.014749190        5
#> 39                0.013798172        5
#> 40                0.010499214        5
#> 41                0.010499214        5
#> 42                0.010499214        5
#> 43                0.010499214        5
#> 44                0.010499214        5
#> 45                0.010499214        5
#> 46                0.010499214        5
#> 47                0.010499214        5
#> 48                0.011022389        5
#> 49                0.020571955        5
#> 50                0.020571955        5
#> 51                0.029972474        5
#> 52                0.012991009        5
#> 53                0.012991009        5
#> 54                0.012991009        5
#> 55                0.012991009        5
#> 56                0.012991009        5
#> 57                0.012991009        5
#> 58                0.122068001        5
#> 59                0.122068001        5
#> 60                0.122068001        5
#> 61                0.122068001        5
#> 62                0.122068001        5
#> 63                0.122068001        5
#> 64                0.049682276        5
#> 65                0.049682276        5
#> 66                0.041624910        5
#> 67                0.041624910        5
#> 68                0.041624910        5
#> 69                0.041624910        5
#> 70                0.041624910        5
#> 71                0.041624910        5
#> 72                0.041624910        5
#> 73                0.041624910        5
#> 74                0.038348947        5
#> 75                0.038348947        5
#> 76                0.038348947        5
#> 77                0.038348947        5
#> 78                0.038348947        5
#> 79                0.023661401        5
#> 80                0.023661401        5
#> 81                0.023661401        5
#> 82                0.023661401        5
#> 83                0.027691169        5
#> 84                0.027691169        5
#> 85                0.047114052        5
#> 86                0.034575924        5
#> 87                0.034575924        5
#> 88                0.034575924        5
#> 89                0.034575924        5
#> 90                0.034575924        5
#> 91                0.034575924        5
#> 92                0.034575924        5
#> 93                0.034575924        5
#> 94                0.034575924        5
#> 95                0.034575924        5
#> 96                0.034575924        5
#> 97                0.018049063        5
#> 98                0.018049063        5
#> 99                0.018049063        5
#> 100               0.018049063        5
#> 101               0.018049063        5
#> 102               0.015923348        5
#> 103               0.012750099        5
#> 104               0.012750099        5
#> 105               0.012750099        5
#> 106               0.012750099        5
#> 107               0.019851043        5
#> 108               0.040310768        5
#> 109               0.040310768        5
#> 110               0.040310768        5
#> 111               0.040310768        5
#> 112               0.040310768        5
#> 113               0.040310768        5
#> 114               0.040310768        5
#> 115               0.040310768        5
#> 116               0.040310768        5
#> 117               0.040310768        5
#> 118               0.040310768        5
#> 119               0.040310768        5
#> 120               0.050191574        5
#> 121               0.065128064        5
#> 122               0.065128064        5
#> 123               0.056224166        5
#> 124               0.056224166        5
#> 125               0.036722337        5
#> 126               0.036722337        5
#> 127               0.036722337        5
#> 128               0.036722337        5
#> 129               0.036722337        5
#> 130               0.036722337        5
#> 131               0.036722337        5
#> 132               0.036722337        5
#> 133               0.036722337        5
#> 134               0.036722337        5
#> 135               0.036722337        5
#> 136               0.020930016        5
#> 137               0.027876196        5
#> 138               0.027876196        5
#> 139               0.027876196        5
#> 140               0.027876196        5
#> 141               0.027876196        5
#> 142               0.027876196        5
#> 143               0.028105468        5
#> 144               0.029247571        5
#> 145               0.029247571        5
#> 146               0.029247571        5
#> 147               0.029247571        5
#> 148               0.029247571        5
#> 149               0.029247571        5
#> 150               0.029247571        5
#> 151               0.036976693        5
#> 152               0.036976693        5
#> 153               0.036976693        5
#> 154               0.015357718        5
#> 155               0.018828470        5
#> 156               0.018828470        5
#> 157               0.018828470        5
#> 158               0.018828470        5
#> 159               0.018828470        5
#> 160               0.042270686        5
#> 161               0.031653181        5
#> 162               0.031653181        5
#> 163               0.023745647        5
#> 164               0.023745647        5
#> 165               0.023745647        5
#> 166               0.023651493        5
#> 167               0.020723009        5
#> 168               0.020723009        5
#> 169               0.019170106        5
#> 170               0.024024105        5
#> 171               0.024024105        5
#> 172               0.024024105        5
#> 173               0.024024105        5
#> 174               0.024024105        5
#> 175               0.026653631        5
#> 176               0.026653631        5
#> 177               0.031777109        5
#> 178               0.031777109        5
#> 179               0.031777109        5
#> 180               0.031777109        5
#> 181               0.031777109        5
#> 182               0.042990013        5
#> 183               0.048705381        5
#> 184               0.048705381        5
#> 185               0.039577316        5
#> 186               0.040951928        5
#> 187               0.040951928        5
#> 188               0.040951928        5
#> 189               0.040951928        5
#> 190               0.029090530        5
#> 191               0.017832099        5
#> 192               0.017832099        5
#> 193               0.017832099        5
#> 194               0.013830332        5
#> 195               0.012051723        5
#> 196               0.012051723        5
#> 197               0.014619397        5
#> 198               0.014619397        5
#> 199               0.014619397        5
#> 200               0.014619397        5
#> 201               0.014619397        5
#> 202               0.018138522        5
#> 203               0.023505286        5
#> 204               0.021595053        5
#> 205               0.019265557        5
#> 206               0.019265557        5
#> 207               0.019265557        5
#> 208               0.021539685        5
#> 209               0.023244887        5
#> 210               0.027013356        5
#> 211               0.019486165        5
#> 212               0.022777337        5
#> 213               0.022777337        5
#> 214               0.022777337        5
#> 215               0.022777337        5
#> 216               0.022777337        5
#> 217               0.024044352        5
#> 218               0.003906933        5
#> 219               0.005726051        5
#> 220               0.005726051        5
#> 221               0.005726051        5