Calculate Area Under the Curve

Calculates area under the deuterium uptake curve

Usage

calculate_auc(
  uptake_dat,
  protein = uptake_dat[["Protein"]][1],
  state = uptake_dat[["State"]][1],
  preserve_values = FALSE
)

Arguments

uptake_dat

data with deuterium uptake values, calculated e.q. by create_uptake_dataset

protein

...

state

...

preserve_values

...

Value

a data.frame object

Details

The AUC is calculated on the data normalized to unit square by division by maximum values of exposure time and deuterium uptake, respectively.

See also

Examples

uptake_dat <- create_uptake_dataset(alpha_dat)
calculate_auc(uptake_dat)
#>       Protein                     Sequence Start End       State Modification
#> 1   db_eEF1Ba                  GFGDLKSPAGL     1  11 ALPHA_Gamma           NA
#> 2   db_eEF1Ba                   FGDLKSPAGL     2  11 ALPHA_Gamma           NA
#> 3   db_eEF1Ba                    GDLKSPAGL     3  11 ALPHA_Gamma           NA
#> 4   db_eEF1Ba                       LKSPAG     5  10 ALPHA_Gamma           NA
#> 5   db_eEF1Ba                      LKSPAGL     5  11 ALPHA_Gamma           NA
#> 6   db_eEF1Ba                       AGLQVL     9  14 ALPHA_Gamma           NA
#> 7   db_eEF1Ba                        QVLND    12  16 ALPHA_Gamma           NA
#> 8   db_eEF1Ba                       QVLNDY    12  17 ALPHA_Gamma           NA
#> 9   db_eEF1Ba                       NDYLAD    15  20 ALPHA_Gamma           NA
#> 10  db_eEF1Ba                      NDYLADK    15  21 ALPHA_Gamma           NA
#> 11  db_eEF1Ba                       YLADKS    17  22 ALPHA_Gamma           NA
#> 12  db_eEF1Ba                      YLADKSY    17  23 ALPHA_Gamma           NA
#> 13  db_eEF1Ba            YLADKSYIEGYVPSQAD    17  33 ALPHA_Gamma           NA
#> 14  db_eEF1Ba              ADKSYIEGYVPSQAD    19  33 ALPHA_Gamma           NA
#> 15  db_eEF1Ba                KSYIEGYVPSQAD    21  33 ALPHA_Gamma           NA
#> 16  db_eEF1Ba                        SYIEG    22  26 ALPHA_Gamma           NA
#> 17  db_eEF1Ba                 SYIEGYVPSQAD    22  33 ALPHA_Gamma           NA
#> 18  db_eEF1Ba                  YIEGYVPSQAD    23  33 ALPHA_Gamma           NA
#> 19  db_eEF1Ba                       IEGYVP    24  29 ALPHA_Gamma           NA
#> 20  db_eEF1Ba                    IEGYVPSQA    24  32 ALPHA_Gamma           NA
#> 21  db_eEF1Ba                   IEGYVPSQAD    24  33 ALPHA_Gamma           NA
#> 22  db_eEF1Ba                      YVPSQAD    27  33 ALPHA_Gamma           NA
#> 23  db_eEF1Ba                        VAVFE    34  38 ALPHA_Gamma           NA
#> 24  db_eEF1Ba                  FEAVSSPPPAD    37  47 ALPHA_Gamma           NA
#> 25  db_eEF1Ba                 FEAVSSPPPADL    37  48 ALPHA_Gamma           NA
#> 26  db_eEF1Ba               FEAVSSPPPADLCH    37  50 ALPHA_Gamma           NA
#> 27  db_eEF1Ba              FEAVSSPPPADLCHA    37  51 ALPHA_Gamma           NA
#> 28  db_eEF1Ba                  EAVSSPPPADL    38  48 ALPHA_Gamma           NA
#> 29  db_eEF1Ba                EAVSSPPPADLCH    38  50 ALPHA_Gamma           NA
#> 30  db_eEF1Ba               EAVSSPPPADLCHA    38  51 ALPHA_Gamma           NA
#> 31  db_eEF1Ba                   AVSSPPPADL    39  48 ALPHA_Gamma           NA
#> 32  db_eEF1Ba                    VSSPPPADL    40  48 ALPHA_Gamma           NA
#> 33  db_eEF1Ba                  VSSPPPADLCH    40  50 ALPHA_Gamma           NA
#> 34  db_eEF1Ba                      ALRWYNH    51  57 ALPHA_Gamma           NA
#> 35  db_eEF1Ba                       LRWYNH    52  57 ALPHA_Gamma           NA
#> 36  db_eEF1Ba             IKSYEKEKASLPGVKK    58  73 ALPHA_Gamma           NA
#> 37  db_eEF1Ba        IKSYEKEKASLPGVKKALGKY    58  78 ALPHA_Gamma           NA
#> 38  db_eEF1Ba              EKASLPGVKKALGKY    64  78 ALPHA_Gamma           NA
#> 39  db_eEF1Ba          EKASLPGVKKALGKYGPAD    64  82 ALPHA_Gamma           NA
#> 40  db_eEF1Ba        EKASLPGVKKALGKYGPADVE    64  84 ALPHA_Gamma           NA
#> 41  db_eEF1Ba            ASLPGVKKALGKYGPAD    66  82 ALPHA_Gamma           NA
#> 42  db_eEF1Ba                  ALGKYGPADVE    74  84 ALPHA_Gamma           NA
#> 43  db_eEF1Ba         VEDTTGSGATDSKDDDDIDL    83 102 ALPHA_Gamma           NA
#> 44  db_eEF1Ba           DTTGSGATDSKDDDDIDL    85 102 ALPHA_Gamma           NA
#> 45  db_eEF1Ba            TTGSGATDSKDDDDIDL    86 102 ALPHA_Gamma           NA
#> 46  db_eEF1Ba                       DDDIDL    97 102 ALPHA_Gamma           NA
#> 47  db_eEF1Ba                       LFGSDD   102 107 ALPHA_Gamma           NA
#> 48  db_eEF1Ba                        FGSDD   103 107 ALPHA_Gamma           NA
#> 49  db_eEF1Ba            FGSDDEEESEEAKRLRE   103 119 ALPHA_Gamma           NA
#> 50  db_eEF1Ba           FGSDDEEESEEAKRLREE   103 120 ALPHA_Gamma           NA
#> 51  db_eEF1Ba         FGSDDEEESEEAKRLREERL   103 122 ALPHA_Gamma           NA
#> 52  db_eEF1Ba           DEEESEEAKRLREERLAQ   107 124 ALPHA_Gamma           NA
#> 53  db_eEF1Ba            EEESEEAKRLREERLAQ   108 124 ALPHA_Gamma           NA
#> 54  db_eEF1Ba                AQYESKKAKKPAL   123 135 ALPHA_Gamma           NA
#> 55  db_eEF1Ba                  YESKKAKKPAL   125 135 ALPHA_Gamma           NA
#> 56  db_eEF1Ba                      VAKSSIL   136 142 ALPHA_Gamma           NA
#> 57  db_eEF1Ba                     VAKSSILL   136 143 ALPHA_Gamma           NA
#> 58  db_eEF1Ba                       AKSSIL   137 142 ALPHA_Gamma           NA
#> 59  db_eEF1Ba                     LDVKPWDD   143 150 ALPHA_Gamma           NA
#> 60  db_eEF1Ba                  LDVKPWDDETD   143 153 ALPHA_Gamma           NA
#> 61  db_eEF1Ba                  DVKPWDDETDM   144 154 ALPHA_Gamma           NA
#> 62  db_eEF1Ba                       MAKLEE   154 159 ALPHA_Gamma           NA
#> 63  db_eEF1Ba                        AKLEE   155 159 ALPHA_Gamma           NA
#> 64  db_eEF1Ba                      CVRSIQA   160 166 ALPHA_Gamma           NA
#> 65  db_eEF1Ba                    CVRSIQADG   160 168 ALPHA_Gamma           NA
#> 66  db_eEF1Ba                   CVRSIQADGL   160 169 ALPHA_Gamma           NA
#> 67  db_eEF1Ba            CVRSIQADGLVWGSSKL   160 176 ALPHA_Gamma           NA
#> 68  db_eEF1Ba                    VRSIQADGL   161 169 ALPHA_Gamma           NA
#> 69  db_eEF1Ba             VRSIQADGLVWGSSKL   161 176 ALPHA_Gamma           NA
#> 70  db_eEF1Ba                      SIQADGL   163 169 ALPHA_Gamma           NA
#> 71  db_eEF1Ba               SIQADGLVWGSSKL   163 176 ALPHA_Gamma           NA
#> 72  db_eEF1Ba SIQADGLVWGSSKLVPVGYGIKKLQIQC   163 190 ALPHA_Gamma           NA
#> 73  db_eEF1Ba         ADGLVWGSSKLVPVGYGIKK   166 185 ALPHA_Gamma           NA
#> 74  db_eEF1Ba                     DGLVWGSS   167 174 ALPHA_Gamma           NA
#> 75  db_eEF1Ba                   DGLVWGSSKL   167 176 ALPHA_Gamma           NA
#> 76  db_eEF1Ba              DGLVWGSSKLVPVGY   167 181 ALPHA_Gamma           NA
#> 77  db_eEF1Ba             DGLVWGSSKLVPVGYG   167 182 ALPHA_Gamma           NA
#> 78  db_eEF1Ba                LVWGSSKLVPVGY   169 181 ALPHA_Gamma           NA
#> 79  db_eEF1Ba             KLVPVGYGIKKLQIQC   175 190 ALPHA_Gamma           NA
#> 80  db_eEF1Ba               VPVGYGIKKLQIQC   177 190 ALPHA_Gamma           NA
#> 81  db_eEF1Ba                       KLQIQC   185 190 ALPHA_Gamma           NA
#> 82  db_eEF1Ba                 CVVEDDKVGTDM   190 201 ALPHA_Gamma           NA
#> 83  db_eEF1Ba                CVVEDDKVGTDML   190 202 ALPHA_Gamma           NA
#> 84  db_eEF1Ba                  VVEDDKVGTDM   191 201 ALPHA_Gamma           NA
#> 85  db_eEF1Ba                 VVEDDKVGTDML   191 202 ALPHA_Gamma           NA
#> 86  db_eEF1Ba                VVEDDKVGTDMLE   191 203 ALPHA_Gamma           NA
#> 87  db_eEF1Ba                     DDKVGTDM   194 201 ALPHA_Gamma           NA
#> 88  db_eEF1Ba                    DDKVGTDML   194 202 ALPHA_Gamma           NA
#> 89  db_eEF1Ba                      DKVGTDM   195 201 ALPHA_Gamma           NA
#> 90  db_eEF1Ba                     DKVGTDML   195 202 ALPHA_Gamma           NA
#> 91  db_eEF1Ba                       VGTDML   197 202 ALPHA_Gamma           NA
#> 92  db_eEF1Ba                       LEEQIT   202 207 ALPHA_Gamma           NA
#> 93  db_eEF1Ba                      LEEQITA   202 208 ALPHA_Gamma           NA
#> 94  db_eEF1Ba                     LEEQITAF   202 209 ALPHA_Gamma           NA
#> 95  db_eEF1Ba                       EEQITA   203 208 ALPHA_Gamma           NA
#> 96  db_eEF1Ba                      EEQITAF   203 209 ALPHA_Gamma           NA
#> 97  db_eEF1Ba                     EEQITAFE   203 210 ALPHA_Gamma           NA
#> 98  db_eEF1Ba                       EQITAF   204 209 ALPHA_Gamma           NA
#> 99  db_eEF1Ba                        QITAF   205 209 ALPHA_Gamma           NA
#> 100 db_eEF1Ba                     FEDYVQSM   209 216 ALPHA_Gamma           NA
#> 101 db_eEF1Ba                      EDYVQSM   210 216 ALPHA_Gamma           NA
#> 102 db_eEF1Ba                     EDYVQSMD   210 217 ALPHA_Gamma           NA
#> 103 db_eEF1Ba                    EDYVQSMDV   210 218 ALPHA_Gamma           NA
#> 104 db_eEF1Ba                       DYVQSM   211 216 ALPHA_Gamma           NA
#> 105 db_eEF1Ba                       YVQSMD   212 217 ALPHA_Gamma           NA
#> 106 db_eEF1Ba                        DVAAF   217 221 ALPHA_Gamma           NA
#>           auc
#> 1   0.9973726
#> 2   0.9993199
#> 3   1.0119067
#> 4   0.9991318
#> 5   0.9991318
#> 6   0.8587529
#> 7   0.4479167
#> 8   0.6116898
#> 9   0.6284722
#> 10  0.5833333
#> 11  0.7955825
#> 12  0.7955825
#> 13  0.7176388
#> 14  0.7477944
#> 15  0.7546626
#> 16  0.7239003
#> 17  0.7546626
#> 18  0.8057869
#> 19  0.9989003
#> 20  0.7790971
#> 21  0.7790392
#> 22  0.7238280
#> 23  0.4479167
#> 24  0.8618199
#> 25  0.7790392
#> 26  0.8155188
#> 27  0.7618137
#> 28  0.8618199
#> 29  0.8155670
#> 30  0.8141299
#> 31  0.8618199
#> 32  0.8158949
#> 33  0.8618199
#> 34  0.4479167
#> 35  0.4479167
#> 36  0.9987395
#> 37  0.9998599
#> 38  0.9998840
#> 39  0.9998840
#> 40  0.9998840
#> 41  0.9998840
#> 42  0.9998840
#> 43  0.9998840
#> 44  0.9998840
#> 45  1.0073162
#> 46  0.9998840
#> 47  0.9998840
#> 48  0.9998840
#> 49  0.9998840
#> 50  0.9998840
#> 51  0.9589166
#> 52  1.0040796
#> 53  0.9998840
#> 54  0.9998840
#> 55  0.9998840
#> 56  0.9963541
#> 57  0.9820746
#> 58  0.9963541
#> 59  0.9762345
#> 60  0.8769097
#> 61  0.9846497
#> 62  0.9755787
#> 63  0.7922454
#> 64  0.9820746
#> 65  0.9973726
#> 66  0.9877314
#> 67  0.9905034
#> 68  0.9974304
#> 69  0.9915315
#> 70  0.9846497
#> 71  0.9907334
#> 72  0.9599247
#> 73  0.9927372
#> 74  0.9842302
#> 75  0.9994327
#> 76  0.9904875
#> 77  0.9915315
#> 78  0.9966186
#> 79  0.9852719
#> 80  0.9894611
#> 81  0.9790799
#> 82  0.9893517
#> 83  0.9879546
#> 84  0.9973726
#> 85  0.9973726
#> 86  0.9891451
#> 87  0.9993199
#> 88  0.9873031
#> 89  1.1803432
#> 90  0.9972365
#> 91  0.9964505
#> 92  0.9755787
#> 93  0.9790799
#> 94  0.9828472
#> 95  0.9755787
#> 96  0.9790799
#> 97  0.9831828
#> 98  0.9755787
#> 99  0.9937500
#> 100 0.9816550
#> 101 0.8587529
#> 102 0.9815827
#> 103 0.9852430
#> 104 0.9761381
#> 105 0.9956983
#> 106 0.9947337