Calculate deuteration

Calculates deuteration uptake based on supplied parameters.

calculate_state_deuteration(
  dat,
  protein,
  state,
  time_in,
  time_chosen,
  time_out,
  deut_part = 1
)

Arguments

dat
data as imported by the read_hdx function

protein
protein included in calculations

state
state included in calculations

time_in
experimental `time_in`

time_chosen
chosen time point

time_out
experimental `time_out`

deut_part
percentage of deuterium the protein was exposed to, value in range [0, 1]

Value

a data.frame object

Details

The function calculate_state_deuteration calculates deuteration for peptides in given protein in given state based on supplied parameters: `time_in`, `time_out` and `time_chosen`. All four variants (combinations of theoretical & relative) are supplied (mean values and uncertainty). Manual correction of percentage of deuterium the protein was exposed to during the exchange in theoretical calculations is provided.

Methods of calculation and uncertainty are profoundly discussed in the vignette.

See also

Examples

# load example data
dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

# calculate deuteration for state "CD160"
calculate_state_deuteration(dat, protein = "db_CD160", state = "CD160",
                            time_in = 0, time_chosen = 5.000, time_out = 1440.000)

#> # A tibble: 41 × 14
#>    Protein  Sequence          Start   End State frac_exch_state err_frac_exch_s…
#>    <chr>    <chr>             <int> <int> <chr>           <dbl>            <dbl>
#>  1 db_CD160 INITSSASQEGTRLN       1    15 CD160            87.6            0.592
#>  2 db_CD160 LICTVW               16    21 CD160            41.7            0.581
#>  3 db_CD160 LICTVWHKKEEAEG       16    29 CD160            72.1            1.07 
#>  4 db_CD160 ICTVWHKKEEAEG        17    29 CD160            75.6            0.780
#>  5 db_CD160 FVVFL                30    34 CD160            40.1            1.85 
#>  6 db_CD160 LCKDRSGDCSPETSL      34    48 CD160            97.6            0.552
#>  7 db_CD160 LCKDRSGDCSPETSLK     34    49 CD160            97.2            0.497
#>  8 db_CD160 LCKDRSGDCSPETSLK…    34    51 CD160            95.3            0.359
#>  9 db_CD160 LCKDRSGDCSPETSLK…    34    69 CD160            93.6            0.336
#> 10 db_CD160 CKDRSGDCSPETSL       35    48 CD160            97.2            0.693
#> # … with 31 more rows, and 7 more variables: abs_frac_exch_state <dbl>,
#> #   err_abs_frac_exch_state <dbl>, avg_theo_in_time <dbl>,
#> #   err_avg_theo_in_time <dbl>, abs_avg_theo_in_time <dbl>,
#> #   err_abs_avg_theo_in_time <dbl>, Med_Sequence <dbl>