Create volcano dataset

Create volcano dataset

create_volcano_dataset(
  dat,
  protein = unique(dat[["Protein"]])[1],
  state_1 = unique(dat[["State"]])[1],
  state_2 = unique(dat[["State"]])[2],
  p_adjustment_method = "none",
  confidence_level = 0.98
)

Arguments

dat
data imported by the read_hdx function.

protein
chosen protein.

state_1
biological state for chosen protein. From this state values the second state values are subtracted to get the deuterium uptake difference.

state_2
biological state for chosen protein. This state values are subtracted from the first state values to get the deuterium uptake difference.

p_adjustment_method
method of adjustment P-values for multiple comparisons. Possible methods: "BH" (Benjamini & Hochberg correction), "bonferroni" (Bonferroni correction) and "none" (default).

confidence_level
confidence level for the t-test.

Value

a data.frame object with calculated deuterium uptake difference, uncertainty, P-value and -log(P-value) for the peptides from the provided data.

Details

The volcano plot shows the deuterium uptake difference in time for the peptides. This function prepares data for the plot. For peptides in all of the time points of measurement (except for minimal and maximal exchange control) the deuterium uptake difference between state_1 and state_2 is calculated, with its uncertainty (combined and propagated as described in `Data processing` article). For each peptide in time point the P-value is calculated using unpaired t-test - the deuterium uptake difference is calculated as the difference of measured masses in a given time point for two states. The tested hypothesis is that the mean masses for states from the replicates of the experiment are similar. The P-values indicates if the null hypothesis can be rejected - rejection of the hypothesis means that the difference between states is statistically significant at provided confidence level. The P-values can be adjusted using the provided method. This plot is visible in GUI.

References

Hageman, T. S. & Weis, D. D. Reliable Identification of Significant Differences in Differential Hydrogen Exchange-Mass Spectrometry Measurements Using a Hybrid Significance Testing Approach. Anal Chem 91, 8008–8016 (2019).

See also

Examples

dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

#> [1] ""

vol_dat <- create_volcano_dataset(dat)

#> Warning: długość dłuszego obiektu nie jest wielokrotnością długości krótszego obiektu

#> Warning: długość dłuszego obiektu nie jest wielokrotnością długości krótszego obiektu

#> Warning: długość dłuszego obiektu nie jest wielokrotnością długości krótszego obiektu

#> Warning: długość dłuszego obiektu nie jest wielokrotnością długości krótszego obiektu

#> Warning: długość dłuszego obiektu nie jest wielokrotnością długości krótszego obiektu

#> Warning: długość dłuszego obiektu nie jest wielokrotnością długości krótszego obiektu

#> Warning: długość dłuszego obiektu nie jest wielokrotnością długości krótszego obiektu

#> Warning: długość dłuszego obiektu nie jest wielokrotnością długości krótszego obiektu

head(vol_dat)

#>         Sequence Start End Exposure       D_diff Uncertainty log_p_value
#> 1 ARSQKSGIRLQGHF    88 101    0.167  0.288733032  0.04274585  6.37955170
#> 2 ARSQKSGIRLQGHF    88 101    1.000  0.179737975  0.06185471  2.90257691
#> 3 ARSQKSGIRLQGHF    88 101  120.000 -0.332431049  0.02093622 12.40741873
#> 4 ARSQKSGIRLQGHF    88 101 1440.000  0.001092857  0.03873885  0.02165359
#> 5 ARSQKSGIRLQGHF    88 101   25.000 -0.186360824  0.02384238  8.30704304
#> 6 ARSQKSGIRLQGHF    88 101    5.000 -0.397528180  0.10068439  2.89209817
#>        P_value
#> 1 1.695883e-03
#> 2 5.488161e-02
#> 3 4.088147e-06
#> 4 9.785792e-01
#> 5 2.467727e-04
#> 6 5.545973e-02