Get replicates sd

Get list of peptides with their standard deviation.

Usage

get_replicate_list_sd(
  dat,
  protein = dat[["Protein"]][1],
  state = dat[["State"]][1],
  time_t = unique(dat[["Exposure"]])[3]
)

Arguments

dat

data as imported by the read_hdx function.

protein

chosen protein.

state

biological state for chosen protein.

time_t

time point of the measurement.

Value

a data.frame object.

Details

Function gets the pepitde list in selected state with their standard deviation of measurement, calculated from the technical replicates. It is used for selection for measurement variability plot.

See also

Examples

get_replicate_list_sd(alpha_dat)
#>                         Sequence Start End Exposure          sd
#> 1                    GFGDLKSPAGL     1  11    0.000          NA
#> 2                    GFGDLKSPAGL     1  11    0.167 0.005298784
#> 3                    GFGDLKSPAGL     1  11    1.000 0.060740980
#> 4                    GFGDLKSPAGL     1  11    5.000 0.099278557
#> 5                    GFGDLKSPAGL     1  11   25.000 0.125437084
#> 6                    GFGDLKSPAGL     1  11  150.000 0.139498448
#> 7                    GFGDLKSPAGL     1  11 1440.000 0.065783930
#> 8                     FGDLKSPAGL     2  11    0.000          NA
#> 9                     FGDLKSPAGL     2  11    0.167 0.027535825
#> 10                    FGDLKSPAGL     2  11    1.000 0.071530970
#> 11                    FGDLKSPAGL     2  11    5.000 0.094991929
#> 12                    FGDLKSPAGL     2  11   25.000 0.018528284
#> 13                    FGDLKSPAGL     2  11  150.000 0.068856529
#> 14                    FGDLKSPAGL     2  11 1440.000 0.046383385
#> 15                     GDLKSPAGL     3  11    0.000          NA
#> 16                     GDLKSPAGL     3  11    0.167 0.031656859
#> 17                     GDLKSPAGL     3  11    1.000 0.061859649
#> 18                     GDLKSPAGL     3  11    5.000 0.071496673
#> 19                     GDLKSPAGL     3  11   25.000 0.057760087
#> 20                     GDLKSPAGL     3  11  150.000 0.108215457
#> 21                     GDLKSPAGL     3  11 1440.000 0.046225696
#> 22                        LKSPAG     5  10    0.000          NA
#> 23                        LKSPAG     5  10    0.167 0.014764882
#> 24                        LKSPAG     5  10    1.000 0.065872038
#> 25                        LKSPAG     5  10    5.000 0.006500575
#> 26                        LKSPAG     5  10   25.000 0.013421469
#> 27                        LKSPAG     5  10  150.000 0.013795561
#> 28                        LKSPAG     5  10 1440.000 0.008140351
#> 29                       LKSPAGL     5  11    0.000          NA
#> 30                       LKSPAGL     5  11    0.167 0.010678931
#> 31                       LKSPAGL     5  11    1.000 0.050497228
#> 32                       LKSPAGL     5  11    5.000 0.034215899
#> 33                       LKSPAGL     5  11   25.000 0.016767404
#> 34                       LKSPAGL     5  11  150.000 0.031292395
#> 35                       LKSPAGL     5  11 1440.000 0.004964169
#> 36                        AGLQVL     9  14    0.000          NA
#> 37                        AGLQVL     9  14    0.167 0.062760928
#> 38                        AGLQVL     9  14    1.000 0.083449116
#> 39                        AGLQVL     9  14    5.000 0.142717665
#> 40                        AGLQVL     9  14   25.000 0.037316330
#> 41                        AGLQVL     9  14  150.000 0.090183539
#> 42                        AGLQVL     9  14 1440.000 0.072036928
#> 43                         QVLND    12  16    0.000          NA
#> 44                         QVLND    12  16    0.167 0.007894753
#> 45                         QVLND    12  16    1.000 0.054409578
#> 46                         QVLND    12  16    5.000 0.016134513
#> 47                         QVLND    12  16   25.000 0.010093219
#> 48                         QVLND    12  16  150.000 0.072963510
#> 49                         QVLND    12  16 1440.000 0.055188475
#> 50                        QVLNDY    12  17    0.000          NA
#> 51                        QVLNDY    12  17    0.167 0.021275423
#> 52                        QVLNDY    12  17    1.000 0.074676640
#> 53                        QVLNDY    12  17    5.000 0.015216383
#> 54                        QVLNDY    12  17   25.000 0.116884924
#> 55                        QVLNDY    12  17  150.000 0.079154703
#> 56                        QVLNDY    12  17 1440.000 0.068917232
#> 57                        NDYLAD    15  20    0.000          NA
#> 58                        NDYLAD    15  20    0.167 0.067150310
#> 59                        NDYLAD    15  20    1.000 0.018584124
#> 60                        NDYLAD    15  20    5.000 0.065489344
#> 61                        NDYLAD    15  20   25.000 0.064829152
#> 62                        NDYLAD    15  20  150.000 0.011045494
#> 63                        NDYLAD    15  20 1440.000 0.084064545
#> 64                       NDYLADK    15  21    0.000          NA
#> 65                       NDYLADK    15  21    0.167 0.106581360
#> 66                       NDYLADK    15  21    1.000 0.054964632
#> 67                       NDYLADK    15  21    5.000 0.035005000
#> 68                       NDYLADK    15  21   25.000 0.040682619
#> 69                       NDYLADK    15  21  150.000 0.043741768
#> 70                       NDYLADK    15  21 1440.000 0.070151041
#> 71                        YLADKS    17  22    0.000          NA
#> 72                        YLADKS    17  22    0.167 0.020210452
#> 73                        YLADKS    17  22    1.000 0.014136578
#> 74                        YLADKS    17  22    5.000 0.014218888
#> 75                        YLADKS    17  22   25.000 0.008806022
#> 76                        YLADKS    17  22  150.000 0.010794676
#> 77                        YLADKS    17  22 1440.000 0.021959659
#> 78                       YLADKSY    17  23    0.000          NA
#> 79                       YLADKSY    17  23    0.167 0.035074343
#> 80                       YLADKSY    17  23    1.000 0.047370324
#> 81                       YLADKSY    17  23    5.000 0.027871759
#> 82                       YLADKSY    17  23   25.000 0.043690053
#> 83                       YLADKSY    17  23  150.000 0.015827680
#> 84                       YLADKSY    17  23 1440.000 0.021953043
#> 85             YLADKSYIEGYVPSQAD    17  33    0.000          NA
#> 86             YLADKSYIEGYVPSQAD    17  33    0.167 0.024869885
#> 87             YLADKSYIEGYVPSQAD    17  33    1.000 0.070217597
#> 88             YLADKSYIEGYVPSQAD    17  33    5.000 0.044254661
#> 89             YLADKSYIEGYVPSQAD    17  33   25.000 0.130275459
#> 90             YLADKSYIEGYVPSQAD    17  33  150.000 0.158363412
#> 91             YLADKSYIEGYVPSQAD    17  33 1440.000 0.025931510
#> 92               ADKSYIEGYVPSQAD    19  33    0.000          NA
#> 93               ADKSYIEGYVPSQAD    19  33    0.167 0.075513738
#> 94               ADKSYIEGYVPSQAD    19  33    1.000 0.046895707
#> 95               ADKSYIEGYVPSQAD    19  33    5.000 0.077290653
#> 96               ADKSYIEGYVPSQAD    19  33   25.000 0.240549886
#> 97               ADKSYIEGYVPSQAD    19  33  150.000 0.272782098
#> 98               ADKSYIEGYVPSQAD    19  33 1440.000 0.062994844
#> 99                 KSYIEGYVPSQAD    21  33    0.000          NA
#> 100                KSYIEGYVPSQAD    21  33    0.167 0.038510517
#> 101                KSYIEGYVPSQAD    21  33    1.000 0.009265464
#> 102                KSYIEGYVPSQAD    21  33    5.000 0.001178005
#> 103                KSYIEGYVPSQAD    21  33   25.000 0.041511278
#> 104                KSYIEGYVPSQAD    21  33  150.000 0.084592829
#> 105                KSYIEGYVPSQAD    21  33 1440.000 0.033585746
#> 106                        SYIEG    22  26    0.000          NA
#> 107                        SYIEG    22  26    0.167 0.036066228
#> 108                        SYIEG    22  26    1.000 0.033312698
#> 109                        SYIEG    22  26    5.000 0.021213402
#> 110                        SYIEG    22  26   25.000 0.006680981
#> 111                        SYIEG    22  26  150.000 0.038062727
#> 112                        SYIEG    22  26 1440.000 0.042817563
#> 113                 SYIEGYVPSQAD    22  33    0.000          NA
#> 114                 SYIEGYVPSQAD    22  33    0.167 0.002202557
#> 115                 SYIEGYVPSQAD    22  33    1.000 0.030362658
#> 116                 SYIEGYVPSQAD    22  33    5.000 0.025862117
#> 117                 SYIEGYVPSQAD    22  33   25.000 0.097649846
#> 118                 SYIEGYVPSQAD    22  33  150.000 0.190229329
#> 119                 SYIEGYVPSQAD    22  33 1440.000 0.040301904
#> 120                  YIEGYVPSQAD    23  33    0.000          NA
#> 121                  YIEGYVPSQAD    23  33    0.167 0.035081945
#> 122                  YIEGYVPSQAD    23  33    1.000 0.041916312
#> 123                  YIEGYVPSQAD    23  33    5.000 0.074872205
#> 124                  YIEGYVPSQAD    23  33   25.000 0.014244781
#> 125                  YIEGYVPSQAD    23  33  150.000 0.085141817
#> 126                  YIEGYVPSQAD    23  33 1440.000 0.038905785
#> 127                       IEGYVP    24  29    0.000          NA
#> 128                       IEGYVP    24  29    0.167 0.028087720
#> 129                       IEGYVP    24  29    1.000 0.046649615
#> 130                       IEGYVP    24  29    5.000 0.041731278
#> 131                       IEGYVP    24  29   25.000 0.054563584
#> 132                       IEGYVP    24  29  150.000 0.071706514
#> 133                       IEGYVP    24  29 1440.000 0.025631601
#> 134                    IEGYVPSQA    24  32    0.000          NA
#> 135                    IEGYVPSQA    24  32    0.167 0.118299421
#> 136                    IEGYVPSQA    24  32    1.000 0.014856903
#> 137                    IEGYVPSQA    24  32    5.000 0.136292539
#> 138                    IEGYVPSQA    24  32   25.000 0.058256424
#> 139                    IEGYVPSQA    24  32  150.000 0.064444839
#> 140                    IEGYVPSQA    24  32 1440.000 0.053059532
#> 141                   IEGYVPSQAD    24  33    0.000          NA
#> 142                   IEGYVPSQAD    24  33    0.167 0.017895127
#> 143                   IEGYVPSQAD    24  33    1.000 0.040255001
#> 144                   IEGYVPSQAD    24  33    5.000 0.005643457
#> 145                   IEGYVPSQAD    24  33   25.000 0.010341512
#> 146                   IEGYVPSQAD    24  33  150.000 0.079482032
#> 147                   IEGYVPSQAD    24  33 1440.000 0.075044791
#> 148                      YVPSQAD    27  33    0.000          NA
#> 149                      YVPSQAD    27  33    0.167 0.056430783
#> 150                      YVPSQAD    27  33    1.000 0.036680773
#> 151                      YVPSQAD    27  33    5.000 0.047157292
#> 152                      YVPSQAD    27  33   25.000 0.051612534
#> 153                      YVPSQAD    27  33  150.000 0.040397819
#> 154                      YVPSQAD    27  33 1440.000 0.037930379
#> 155                        VAVFE    34  38    0.000          NA
#> 156                        VAVFE    34  38    0.167 0.029952959
#> 157                        VAVFE    34  38    1.000 0.022151358
#> 158                        VAVFE    34  38    5.000 0.058616625
#> 159                        VAVFE    34  38   25.000 0.019621210
#> 160                        VAVFE    34  38  150.000 0.048383233
#> 161                        VAVFE    34  38 1440.000 0.038887694
#> 162                  FEAVSSPPPAD    37  47    0.000          NA
#> 163                  FEAVSSPPPAD    37  47    0.167 0.072605039
#> 164                  FEAVSSPPPAD    37  47    1.000 0.029235850
#> 165                  FEAVSSPPPAD    37  47    5.000 0.059066051
#> 166                  FEAVSSPPPAD    37  47   25.000 0.286849343
#> 167                  FEAVSSPPPAD    37  47  150.000 0.131247753
#> 168                  FEAVSSPPPAD    37  47 1440.000 0.323132995
#> 169                 FEAVSSPPPADL    37  48    0.000          NA
#> 170                 FEAVSSPPPADL    37  48    0.167 0.001827672
#> 171                 FEAVSSPPPADL    37  48    1.000 0.041334303
#> 172                 FEAVSSPPPADL    37  48    5.000 0.051128537
#> 173                 FEAVSSPPPADL    37  48   25.000 0.048278662
#> 174                 FEAVSSPPPADL    37  48  150.000 0.019477901
#> 175                 FEAVSSPPPADL    37  48 1440.000 0.021762119
#> 176               FEAVSSPPPADLCH    37  50    0.000          NA
#> 177               FEAVSSPPPADLCH    37  50    0.167 0.023687784
#> 178               FEAVSSPPPADLCH    37  50    1.000 0.163561872
#> 179               FEAVSSPPPADLCH    37  50    5.000 0.084783178
#> 180               FEAVSSPPPADLCH    37  50   25.000 0.031796203
#> 181               FEAVSSPPPADLCH    37  50  150.000 0.078642567
#> 182               FEAVSSPPPADLCH    37  50 1440.000 0.052340862
#> 183              FEAVSSPPPADLCHA    37  51    0.000          NA
#> 184              FEAVSSPPPADLCHA    37  51    0.167 0.170792513
#> 185              FEAVSSPPPADLCHA    37  51    1.000 0.080115640
#> 186              FEAVSSPPPADLCHA    37  51    5.000 0.141986094
#> 187              FEAVSSPPPADLCHA    37  51   25.000 0.109084312
#> 188              FEAVSSPPPADLCHA    37  51  150.000 0.197342138
#> 189              FEAVSSPPPADLCHA    37  51 1440.000 0.063604351
#> 190                  EAVSSPPPADL    38  48    0.000          NA
#> 191                  EAVSSPPPADL    38  48    0.167 0.036460163
#> 192                  EAVSSPPPADL    38  48    1.000 0.030927743
#> 193                  EAVSSPPPADL    38  48    5.000 0.044639220
#> 194                  EAVSSPPPADL    38  48   25.000 0.013969435
#> 195                  EAVSSPPPADL    38  48  150.000 0.043086593
#> 196                  EAVSSPPPADL    38  48 1440.000 0.026240160
#> 197                EAVSSPPPADLCH    38  50    0.000          NA
#> 198                EAVSSPPPADLCH    38  50    0.167 0.031257538
#> 199                EAVSSPPPADLCH    38  50    1.000 0.025997166
#> 200                EAVSSPPPADLCH    38  50    5.000 0.018587049
#> 201                EAVSSPPPADLCH    38  50   25.000 0.064847485
#> 202                EAVSSPPPADLCH    38  50  150.000 0.030117944
#> 203                EAVSSPPPADLCH    38  50 1440.000 0.109537964
#> 204               EAVSSPPPADLCHA    38  51    0.000          NA
#> 205               EAVSSPPPADLCHA    38  51    0.167 0.061866552
#> 206               EAVSSPPPADLCHA    38  51    1.000 0.150943111
#> 207               EAVSSPPPADLCHA    38  51    5.000 0.112628475
#> 208               EAVSSPPPADLCHA    38  51   25.000 0.090498372
#> 209               EAVSSPPPADLCHA    38  51  150.000 0.017030991
#> 210               EAVSSPPPADLCHA    38  51 1440.000 0.115011063
#> 211                   AVSSPPPADL    39  48    0.000          NA
#> 212                   AVSSPPPADL    39  48    0.167 0.133248295
#> 213                   AVSSPPPADL    39  48    1.000 0.027596214
#> 214                   AVSSPPPADL    39  48    5.000 0.030332481
#> 215                   AVSSPPPADL    39  48   25.000 0.058049131
#> 216                   AVSSPPPADL    39  48  150.000 0.043875819
#> 217                   AVSSPPPADL    39  48 1440.000 0.109363451
#> 218                    VSSPPPADL    40  48    0.000          NA
#> 219                    VSSPPPADL    40  48    0.167 0.033087631
#> 220                    VSSPPPADL    40  48    1.000 0.085657601
#> 221                    VSSPPPADL    40  48    5.000 0.023967875
#> 222                    VSSPPPADL    40  48   25.000 0.027621467
#> 223                    VSSPPPADL    40  48  150.000 0.023120730
#> 224                    VSSPPPADL    40  48 1440.000 0.102113822
#> 225                  VSSPPPADLCH    40  50    0.000          NA
#> 226                  VSSPPPADLCH    40  50    0.167 0.084987361
#> 227                  VSSPPPADLCH    40  50    1.000 0.046813789
#> 228                  VSSPPPADLCH    40  50    5.000 0.018948192
#> 229                  VSSPPPADLCH    40  50   25.000 0.056346508
#> 230                  VSSPPPADLCH    40  50  150.000 0.005301057
#> 231                  VSSPPPADLCH    40  50 1440.000 0.078926666
#> 232                      ALRWYNH    51  57    0.000          NA
#> 233                      ALRWYNH    51  57    0.167 0.031893041
#> 234                      ALRWYNH    51  57    1.000 0.057142167
#> 235                      ALRWYNH    51  57    5.000 0.042901556
#> 236                      ALRWYNH    51  57   25.000 0.028906059
#> 237                      ALRWYNH    51  57  150.000 0.050853687
#> 238                      ALRWYNH    51  57 1440.000 0.040836822
#> 239                       LRWYNH    52  57    0.000          NA
#> 240                       LRWYNH    52  57    0.167 0.028898586
#> 241                       LRWYNH    52  57    1.000 0.006507050
#> 242                       LRWYNH    52  57    5.000 0.020580860
#> 243                       LRWYNH    52  57   25.000 0.007358824
#> 244                       LRWYNH    52  57  150.000 0.049264727
#> 245                       LRWYNH    52  57 1440.000 0.047937484
#> 246             IKSYEKEKASLPGVKK    58  73    0.000          NA
#> 247             IKSYEKEKASLPGVKK    58  73    0.167 0.054426176
#> 248             IKSYEKEKASLPGVKK    58  73    1.000 0.079015271
#> 249             IKSYEKEKASLPGVKK    58  73    5.000 0.071977738
#> 250             IKSYEKEKASLPGVKK    58  73   25.000 0.054079119
#> 251             IKSYEKEKASLPGVKK    58  73  150.000 0.196342666
#> 252             IKSYEKEKASLPGVKK    58  73 1440.000 0.034257419
#> 253        IKSYEKEKASLPGVKKALGKY    58  78    0.000          NA
#> 254        IKSYEKEKASLPGVKKALGKY    58  78    0.167 0.193488228
#> 255        IKSYEKEKASLPGVKKALGKY    58  78    1.000 0.072963736
#> 256        IKSYEKEKASLPGVKKALGKY    58  78    5.000 0.488698726
#> 257        IKSYEKEKASLPGVKKALGKY    58  78   25.000 0.243096975
#> 258        IKSYEKEKASLPGVKKALGKY    58  78  150.000 0.231713879
#> 259        IKSYEKEKASLPGVKKALGKY    58  78 1440.000 0.102410305
#> 260              EKASLPGVKKALGKY    64  78    0.000          NA
#> 261              EKASLPGVKKALGKY    64  78    0.167 0.063930248
#> 262              EKASLPGVKKALGKY    64  78    1.000 0.320534453
#> 263              EKASLPGVKKALGKY    64  78    5.000 0.022522532
#> 264              EKASLPGVKKALGKY    64  78   25.000 0.057855430
#> 265              EKASLPGVKKALGKY    64  78  150.000 0.045367082
#> 266              EKASLPGVKKALGKY    64  78 1440.000 0.098957337
#> 267          EKASLPGVKKALGKYGPAD    64  82    0.000          NA
#> 268          EKASLPGVKKALGKYGPAD    64  82    0.167 0.107715901
#> 269          EKASLPGVKKALGKYGPAD    64  82    1.000 0.081268025
#> 270          EKASLPGVKKALGKYGPAD    64  82    5.000 0.115992915
#> 271          EKASLPGVKKALGKYGPAD    64  82   25.000 0.039047293
#> 272          EKASLPGVKKALGKYGPAD    64  82  150.000 0.054376684
#> 273          EKASLPGVKKALGKYGPAD    64  82 1440.000 0.175063102
#> 274        EKASLPGVKKALGKYGPADVE    64  84    0.000          NA
#> 275        EKASLPGVKKALGKYGPADVE    64  84    0.167 0.261595194
#> 276        EKASLPGVKKALGKYGPADVE    64  84    1.000 0.042742707
#> 277        EKASLPGVKKALGKYGPADVE    64  84    5.000 0.057381931
#> 278        EKASLPGVKKALGKYGPADVE    64  84   25.000 0.080751022
#> 279        EKASLPGVKKALGKYGPADVE    64  84  150.000 0.163556587
#> 280        EKASLPGVKKALGKYGPADVE    64  84 1440.000 0.072935459
#> 281            ASLPGVKKALGKYGPAD    66  82    0.000          NA
#> 282            ASLPGVKKALGKYGPAD    66  82    0.167 0.054956828
#> 283            ASLPGVKKALGKYGPAD    66  82    1.000 0.091506586
#> 284            ASLPGVKKALGKYGPAD    66  82    5.000 0.087176821
#> 285            ASLPGVKKALGKYGPAD    66  82   25.000 0.077543129
#> 286            ASLPGVKKALGKYGPAD    66  82  150.000 0.048678359
#> 287            ASLPGVKKALGKYGPAD    66  82 1440.000 0.026536989
#> 288                  ALGKYGPADVE    74  84    0.000          NA
#> 289                  ALGKYGPADVE    74  84    0.167 0.030183961
#> 290                  ALGKYGPADVE    74  84    1.000 0.045175869
#> 291                  ALGKYGPADVE    74  84    5.000 0.063811199
#> 292                  ALGKYGPADVE    74  84   25.000 0.042330527
#> 293                  ALGKYGPADVE    74  84  150.000 0.089176645
#> 294                  ALGKYGPADVE    74  84 1440.000 0.030896634
#> 295         VEDTTGSGATDSKDDDDIDL    83 102    0.000          NA
#> 296         VEDTTGSGATDSKDDDDIDL    83 102    0.167 0.063509101
#> 297         VEDTTGSGATDSKDDDDIDL    83 102    1.000 0.054375063
#> 298         VEDTTGSGATDSKDDDDIDL    83 102    5.000 0.134615656
#> 299         VEDTTGSGATDSKDDDDIDL    83 102   25.000 0.137481431
#> 300         VEDTTGSGATDSKDDDDIDL    83 102  150.000 0.175733223
#> 301         VEDTTGSGATDSKDDDDIDL    83 102 1440.000 0.138249119
#> 302           DTTGSGATDSKDDDDIDL    85 102    0.000          NA
#> 303           DTTGSGATDSKDDDDIDL    85 102    0.167 0.049077443
#> 304           DTTGSGATDSKDDDDIDL    85 102    1.000 0.075884537
#> 305           DTTGSGATDSKDDDDIDL    85 102    5.000 0.096951429
#> 306           DTTGSGATDSKDDDDIDL    85 102   25.000 0.138835551
#> 307           DTTGSGATDSKDDDDIDL    85 102  150.000 0.197676075
#> 308           DTTGSGATDSKDDDDIDL    85 102 1440.000 0.138074695
#> 309            TTGSGATDSKDDDDIDL    86 102    0.000          NA
#> 310            TTGSGATDSKDDDDIDL    86 102    0.167 0.028634664
#> 311            TTGSGATDSKDDDDIDL    86 102    1.000 0.093085128
#> 312            TTGSGATDSKDDDDIDL    86 102    5.000 0.081279596
#> 313            TTGSGATDSKDDDDIDL    86 102   25.000 0.076411724
#> 314            TTGSGATDSKDDDDIDL    86 102  150.000 0.155458547
#> 315            TTGSGATDSKDDDDIDL    86 102 1440.000 0.118899440
#> 316                       DDDIDL    97 102    0.000          NA
#> 317                       DDDIDL    97 102    0.167 0.050496681
#> 318                       DDDIDL    97 102    1.000 0.068976783
#> 319                       DDDIDL    97 102    5.000 0.026523807
#> 320                       DDDIDL    97 102   25.000 0.031913731
#> 321                       DDDIDL    97 102  150.000 0.032618571
#> 322                       DDDIDL    97 102 1440.000 0.046947213
#> 323                       LFGSDD   102 107    0.000          NA
#> 324                       LFGSDD   102 107    0.167 0.027543576
#> 325                       LFGSDD   102 107    1.000 0.036862004
#> 326                       LFGSDD   102 107    5.000 0.053364510
#> 327                       LFGSDD   102 107   25.000 0.043062259
#> 328                       LFGSDD   102 107  150.000 0.025202488
#> 329                       LFGSDD   102 107 1440.000 0.045728701
#> 330                        FGSDD   103 107    0.000          NA
#> 331                        FGSDD   103 107    0.167 0.099104204
#> 332                        FGSDD   103 107    1.000 0.054227631
#> 333                        FGSDD   103 107    5.000 0.018298150
#> 334                        FGSDD   103 107   25.000 0.009757323
#> 335                        FGSDD   103 107  150.000 0.046364343
#> 336                        FGSDD   103 107 1440.000 0.059226039
#> 337            FGSDDEEESEEAKRLRE   103 119    0.000          NA
#> 338            FGSDDEEESEEAKRLRE   103 119    0.167 0.100665103
#> 339            FGSDDEEESEEAKRLRE   103 119    1.000 0.087935262
#> 340            FGSDDEEESEEAKRLRE   103 119    5.000 0.063039515
#> 341            FGSDDEEESEEAKRLRE   103 119   25.000 0.131429929
#> 342            FGSDDEEESEEAKRLRE   103 119  150.000 0.131808440
#> 343            FGSDDEEESEEAKRLRE   103 119 1440.000 0.016636706
#> 344           FGSDDEEESEEAKRLREE   103 120    0.000          NA
#> 345           FGSDDEEESEEAKRLREE   103 120    0.167 0.078617481
#> 346           FGSDDEEESEEAKRLREE   103 120    1.000 0.143544716
#> 347           FGSDDEEESEEAKRLREE   103 120    5.000 0.084061848
#> 348           FGSDDEEESEEAKRLREE   103 120   25.000 0.105135790
#> 349           FGSDDEEESEEAKRLREE   103 120  150.000 0.061354396
#> 350           FGSDDEEESEEAKRLREE   103 120 1440.000 0.020586403
#> 351         FGSDDEEESEEAKRLREERL   103 122    0.000          NA
#> 352         FGSDDEEESEEAKRLREERL   103 122    0.167 0.177861683
#> 353         FGSDDEEESEEAKRLREERL   103 122    1.000 0.071027115
#> 354         FGSDDEEESEEAKRLREERL   103 122    5.000 0.073727977
#> 355         FGSDDEEESEEAKRLREERL   103 122   25.000 0.086008687
#> 356         FGSDDEEESEEAKRLREERL   103 122  150.000 0.186283446
#> 357         FGSDDEEESEEAKRLREERL   103 122 1440.000 0.039631162
#> 358           DEEESEEAKRLREERLAQ   107 124    0.000          NA
#> 359           DEEESEEAKRLREERLAQ   107 124    0.167 0.061244508
#> 360           DEEESEEAKRLREERLAQ   107 124    1.000 0.254088805
#> 361           DEEESEEAKRLREERLAQ   107 124    5.000 0.323112969
#> 362           DEEESEEAKRLREERLAQ   107 124   25.000 0.059862499
#> 363           DEEESEEAKRLREERLAQ   107 124  150.000 0.285233733
#> 364           DEEESEEAKRLREERLAQ   107 124 1440.000 0.255233431
#> 365            EEESEEAKRLREERLAQ   108 124    0.000          NA
#> 366            EEESEEAKRLREERLAQ   108 124    0.167 0.034139158
#> 367            EEESEEAKRLREERLAQ   108 124    1.000 0.092069183
#> 368            EEESEEAKRLREERLAQ   108 124    5.000 0.047336577
#> 369            EEESEEAKRLREERLAQ   108 124   25.000 0.073428689
#> 370            EEESEEAKRLREERLAQ   108 124  150.000 0.100011266
#> 371            EEESEEAKRLREERLAQ   108 124 1440.000 0.051577397
#> 372                AQYESKKAKKPAL   123 135    0.000          NA
#> 373                AQYESKKAKKPAL   123 135    0.167 0.065444746
#> 374                AQYESKKAKKPAL   123 135    1.000 0.104394132
#> 375                AQYESKKAKKPAL   123 135    5.000 0.083161364
#> 376                AQYESKKAKKPAL   123 135   25.000 0.037412996
#> 377                AQYESKKAKKPAL   123 135  150.000 0.076119462
#> 378                AQYESKKAKKPAL   123 135 1440.000 0.025845452
#> 379                  YESKKAKKPAL   125 135    0.000          NA
#> 380                  YESKKAKKPAL   125 135    0.167 0.054361717
#> 381                  YESKKAKKPAL   125 135    1.000 0.043536500
#> 382                  YESKKAKKPAL   125 135    5.000 0.093510738
#> 383                  YESKKAKKPAL   125 135   25.000 0.053400630
#> 384                  YESKKAKKPAL   125 135  150.000 0.105698090
#> 385                  YESKKAKKPAL   125 135 1440.000 0.070822922
#> 386                      VAKSSIL   136 142    0.000          NA
#> 387                      VAKSSIL   136 142    0.167 0.017689651
#> 388                      VAKSSIL   136 142    1.000 0.059349162
#> 389                      VAKSSIL   136 142    5.000 0.053659502
#> 390                      VAKSSIL   136 142   25.000 0.079828209
#> 391                      VAKSSIL   136 142  150.000 0.078651063
#> 392                      VAKSSIL   136 142 1440.000 0.051931419
#> 393                     VAKSSILL   136 143    0.000          NA
#> 394                     VAKSSILL   136 143    0.167 0.025600201
#> 395                     VAKSSILL   136 143    1.000 0.069312070
#> 396                     VAKSSILL   136 143    5.000 0.047437523
#> 397                     VAKSSILL   136 143   25.000 0.074585682
#> 398                     VAKSSILL   136 143  150.000 0.026843783
#> 399                     VAKSSILL   136 143 1440.000 0.024891654
#> 400                       AKSSIL   137 142    0.000          NA
#> 401                       AKSSIL   137 142    0.167 0.016800881
#> 402                       AKSSIL   137 142    1.000 0.069075812
#> 403                       AKSSIL   137 142    5.000 0.109466770
#> 404                       AKSSIL   137 142   25.000 0.089145727
#> 405                       AKSSIL   137 142  150.000 0.056615343
#> 406                       AKSSIL   137 142 1440.000 0.070010175
#> 407                     LDVKPWDD   143 150    0.000          NA
#> 408                     LDVKPWDD   143 150    0.167 0.056078749
#> 409                     LDVKPWDD   143 150    1.000 0.039504153
#> 410                     LDVKPWDD   143 150    5.000 0.081729605
#> 411                     LDVKPWDD   143 150   25.000 0.194950108
#> 412                     LDVKPWDD   143 150  150.000 0.078900881
#> 413                     LDVKPWDD   143 150 1440.000 0.058929238
#> 414                  LDVKPWDDETD   143 153    0.000          NA
#> 415                  LDVKPWDDETD   143 153    0.167 0.056813904
#> 416                  LDVKPWDDETD   143 153    1.000 0.011410612
#> 417                  LDVKPWDDETD   143 153    5.000 0.123341446
#> 418                  LDVKPWDDETD   143 153   25.000 0.069772438
#> 419                  LDVKPWDDETD   143 153  150.000 0.086396200
#> 420                  LDVKPWDDETD   143 153 1440.000 0.072787473
#> 421                  DVKPWDDETDM   144 154    0.000          NA
#> 422                  DVKPWDDETDM   144 154    0.167 0.025752550
#> 423                  DVKPWDDETDM   144 154    1.000 0.044105734
#> 424                  DVKPWDDETDM   144 154    5.000 0.034557166
#> 425                  DVKPWDDETDM   144 154   25.000 0.155271936
#> 426                  DVKPWDDETDM   144 154  150.000 0.158700185
#> 427                  DVKPWDDETDM   144 154 1440.000 0.026741586
#> 428                       MAKLEE   154 159    0.000          NA
#> 429                       MAKLEE   154 159    0.167 0.020060656
#> 430                       MAKLEE   154 159    1.000 0.042541692
#> 431                       MAKLEE   154 159    5.000 0.036654956
#> 432                       MAKLEE   154 159   25.000 0.058536414
#> 433                       MAKLEE   154 159  150.000 0.051826388
#> 434                       MAKLEE   154 159 1440.000 0.053613679
#> 435                        AKLEE   155 159    0.000          NA
#> 436                        AKLEE   155 159    0.167 0.010673570
#> 437                        AKLEE   155 159    1.000 0.005377470
#> 438                        AKLEE   155 159    5.000 0.050852284
#> 439                        AKLEE   155 159   25.000 0.108831214
#> 440                        AKLEE   155 159  150.000 0.038824682
#> 441                        AKLEE   155 159 1440.000 0.017121640
#> 442                      CVRSIQA   160 166    0.000          NA
#> 443                      CVRSIQA   160 166    0.167 0.039902301
#> 444                      CVRSIQA   160 166    1.000 0.033954997
#> 445                      CVRSIQA   160 166    5.000 0.083822153
#> 446                      CVRSIQA   160 166   25.000          NA
#> 447                      CVRSIQA   160 166  150.000 0.109316105
#> 448                      CVRSIQA   160 166 1440.000 0.041594984
#> 449                    CVRSIQADG   160 168    0.000          NA
#> 450                    CVRSIQADG   160 168    0.167 0.076406627
#> 451                    CVRSIQADG   160 168    1.000 0.064243183
#> 452                    CVRSIQADG   160 168    5.000 0.169241588
#> 453                    CVRSIQADG   160 168   25.000 0.191865694
#> 454                    CVRSIQADG   160 168  150.000 0.244624152
#> 455                    CVRSIQADG   160 168 1440.000 0.077512154
#> 456                   CVRSIQADGL   160 169    0.000          NA
#> 457                   CVRSIQADGL   160 169    0.167 0.087999461
#> 458                   CVRSIQADGL   160 169    1.000 0.094060643
#> 459                   CVRSIQADGL   160 169    5.000 0.166803246
#> 460                   CVRSIQADGL   160 169   25.000 0.158986614
#> 461                   CVRSIQADGL   160 169  150.000 0.129879835
#> 462                   CVRSIQADGL   160 169 1440.000 0.093246270
#> 463            CVRSIQADGLVWGSSKL   160 176    0.000          NA
#> 464            CVRSIQADGLVWGSSKL   160 176    0.167 0.019163575
#> 465            CVRSIQADGLVWGSSKL   160 176    1.000 0.056656979
#> 466            CVRSIQADGLVWGSSKL   160 176    5.000 0.140078245
#> 467            CVRSIQADGLVWGSSKL   160 176   25.000 0.293302384
#> 468            CVRSIQADGLVWGSSKL   160 176  150.000 0.391065775
#> 469            CVRSIQADGLVWGSSKL   160 176 1440.000 0.085493197
#> 470                    VRSIQADGL   161 169    0.000          NA
#> 471                    VRSIQADGL   161 169    0.167 0.030852690
#> 472                    VRSIQADGL   161 169    1.000 0.051052611
#> 473                    VRSIQADGL   161 169    5.000 0.082699477
#> 474                    VRSIQADGL   161 169   25.000 0.081457909
#> 475                    VRSIQADGL   161 169  150.000 0.083031733
#> 476                    VRSIQADGL   161 169 1440.000 0.055722544
#> 477             VRSIQADGLVWGSSKL   161 176    0.000          NA
#> 478             VRSIQADGLVWGSSKL   161 176    0.167 0.110895107
#> 479             VRSIQADGLVWGSSKL   161 176    1.000 0.093662545
#> 480             VRSIQADGLVWGSSKL   161 176    5.000 0.082037223
#> 481             VRSIQADGLVWGSSKL   161 176   25.000 0.166092460
#> 482             VRSIQADGLVWGSSKL   161 176  150.000 0.206095783
#> 483             VRSIQADGLVWGSSKL   161 176 1440.000 0.162459810
#> 484                      SIQADGL   163 169    0.000          NA
#> 485                      SIQADGL   163 169    0.167 0.058668584
#> 486                      SIQADGL   163 169    1.000 0.025232293
#> 487                      SIQADGL   163 169    5.000 0.036587892
#> 488                      SIQADGL   163 169   25.000 0.047096992
#> 489                      SIQADGL   163 169  150.000 0.122096362
#> 490                      SIQADGL   163 169 1440.000 0.029278111
#> 491               SIQADGLVWGSSKL   163 176    0.000          NA
#> 492               SIQADGLVWGSSKL   163 176    0.167 0.164759129
#> 493               SIQADGLVWGSSKL   163 176    1.000 0.270485797
#> 494               SIQADGLVWGSSKL   163 176    5.000 0.137367362
#> 495               SIQADGLVWGSSKL   163 176   25.000 0.167063959
#> 496               SIQADGLVWGSSKL   163 176  150.000 0.487920178
#> 497               SIQADGLVWGSSKL   163 176 1440.000 0.130318956
#> 498 SIQADGLVWGSSKLVPVGYGIKKLQIQC   163 190    0.000          NA
#> 499 SIQADGLVWGSSKLVPVGYGIKKLQIQC   163 190    0.167 0.086619920
#> 500 SIQADGLVWGSSKLVPVGYGIKKLQIQC   163 190    1.000 0.059736391
#> 501 SIQADGLVWGSSKLVPVGYGIKKLQIQC   163 190    5.000 0.175219571
#> 502 SIQADGLVWGSSKLVPVGYGIKKLQIQC   163 190   25.000 0.322804230
#> 503 SIQADGLVWGSSKLVPVGYGIKKLQIQC   163 190  150.000 0.194275186
#> 504 SIQADGLVWGSSKLVPVGYGIKKLQIQC   163 190 1440.000 0.087211292
#> 505         ADGLVWGSSKLVPVGYGIKK   166 185    0.000          NA
#> 506         ADGLVWGSSKLVPVGYGIKK   166 185    0.167 0.095117502
#> 507         ADGLVWGSSKLVPVGYGIKK   166 185    1.000 0.167510803
#> 508         ADGLVWGSSKLVPVGYGIKK   166 185    5.000 0.233628978
#> 509         ADGLVWGSSKLVPVGYGIKK   166 185   25.000 0.345310685
#> 510         ADGLVWGSSKLVPVGYGIKK   166 185  150.000 0.365671199
#> 511         ADGLVWGSSKLVPVGYGIKK   166 185 1440.000 0.338196297
#> 512                     DGLVWGSS   167 174    0.000          NA
#> 513                     DGLVWGSS   167 174    0.167 0.076788636
#> 514                     DGLVWGSS   167 174    1.000 0.019718762
#> 515                     DGLVWGSS   167 174    5.000 0.049756694
#> 516                     DGLVWGSS   167 174   25.000 0.118152141
#> 517                     DGLVWGSS   167 174  150.000 0.102612389
#> 518                     DGLVWGSS   167 174 1440.000 0.044486076
#> 519                   DGLVWGSSKL   167 176    0.000          NA
#> 520                   DGLVWGSSKL   167 176    0.167 0.071918417
#> 521                   DGLVWGSSKL   167 176    1.000 0.203495677
#> 522                   DGLVWGSSKL   167 176    5.000 0.162512091
#> 523                   DGLVWGSSKL   167 176   25.000 0.124411260
#> 524                   DGLVWGSSKL   167 176  150.000 0.039284244
#> 525                   DGLVWGSSKL   167 176 1440.000 0.057505132
#> 526              DGLVWGSSKLVPVGY   167 181    0.000          NA
#> 527              DGLVWGSSKLVPVGY   167 181    0.167 0.028458660
#> 528              DGLVWGSSKLVPVGY   167 181    1.000 0.100558121
#> 529              DGLVWGSSKLVPVGY   167 181    5.000 0.114790966
#> 530              DGLVWGSSKLVPVGY   167 181   25.000 0.108472540
#> 531              DGLVWGSSKLVPVGY   167 181  150.000 0.088268646
#> 532              DGLVWGSSKLVPVGY   167 181 1440.000 0.134011081
#> 533             DGLVWGSSKLVPVGYG   167 182    0.000          NA
#> 534             DGLVWGSSKLVPVGYG   167 182    0.167 0.049158959
#> 535             DGLVWGSSKLVPVGYG   167 182    1.000 0.151140874
#> 536             DGLVWGSSKLVPVGYG   167 182    5.000 0.158410312
#> 537             DGLVWGSSKLVPVGYG   167 182   25.000 0.168566056
#> 538             DGLVWGSSKLVPVGYG   167 182  150.000 0.256063475
#> 539             DGLVWGSSKLVPVGYG   167 182 1440.000 0.113365619
#> 540                LVWGSSKLVPVGY   169 181    0.000          NA
#> 541                LVWGSSKLVPVGY   169 181    0.167 0.071560176
#> 542                LVWGSSKLVPVGY   169 181    1.000 0.026306137
#> 543                LVWGSSKLVPVGY   169 181    5.000 0.095316112
#> 544                LVWGSSKLVPVGY   169 181   25.000 0.022286914
#> 545                LVWGSSKLVPVGY   169 181  150.000 0.022268340
#> 546                LVWGSSKLVPVGY   169 181 1440.000 0.043847641
#> 547             KLVPVGYGIKKLQIQC   175 190    0.000          NA
#> 548             KLVPVGYGIKKLQIQC   175 190    0.167 0.032276706
#> 549             KLVPVGYGIKKLQIQC   175 190    1.000 0.072611126
#> 550             KLVPVGYGIKKLQIQC   175 190    5.000 0.118354170
#> 551             KLVPVGYGIKKLQIQC   175 190   25.000 0.296985311
#> 552             KLVPVGYGIKKLQIQC   175 190  150.000 0.139917711
#> 553             KLVPVGYGIKKLQIQC   175 190 1440.000 0.062893040
#> 554               VPVGYGIKKLQIQC   177 190    0.000          NA
#> 555               VPVGYGIKKLQIQC   177 190    0.167 0.060513988
#> 556               VPVGYGIKKLQIQC   177 190    1.000 0.084205478
#> 557               VPVGYGIKKLQIQC   177 190    5.000 0.149450172
#> 558               VPVGYGIKKLQIQC   177 190   25.000 0.260239212
#> 559               VPVGYGIKKLQIQC   177 190  150.000 0.084050373
#> 560               VPVGYGIKKLQIQC   177 190 1440.000 0.067127312
#> 561                       KLQIQC   185 190    0.000          NA
#> 562                       KLQIQC   185 190    0.167 0.013259825
#> 563                       KLQIQC   185 190    1.000 0.030843386
#> 564                       KLQIQC   185 190    5.000 0.112467020
#> 565                       KLQIQC   185 190   25.000 0.202311958
#> 566                       KLQIQC   185 190  150.000 0.012219419
#> 567                       KLQIQC   185 190 1440.000 0.034779379
#> 568                 CVVEDDKVGTDM   190 201    0.000          NA
#> 569                 CVVEDDKVGTDM   190 201    0.167 0.034799747
#> 570                 CVVEDDKVGTDM   190 201    1.000 0.092583325
#> 571                 CVVEDDKVGTDM   190 201    5.000 0.061755803
#> 572                 CVVEDDKVGTDM   190 201   25.000 0.124730140
#> 573                 CVVEDDKVGTDM   190 201  150.000 0.173586780
#> 574                 CVVEDDKVGTDM   190 201 1440.000 0.036900245
#> 575                CVVEDDKVGTDML   190 202    0.000          NA
#> 576                CVVEDDKVGTDML   190 202    0.167 0.118464210
#> 577                CVVEDDKVGTDML   190 202    1.000 0.082830695
#> 578                CVVEDDKVGTDML   190 202    5.000 0.034082976
#> 579                CVVEDDKVGTDML   190 202   25.000 0.104095612
#> 580                CVVEDDKVGTDML   190 202  150.000 0.008072486
#> 581                CVVEDDKVGTDML   190 202 1440.000 0.051705338
#> 582                  VVEDDKVGTDM   191 201    0.000          NA
#> 583                  VVEDDKVGTDM   191 201    0.167 0.046341072
#> 584                  VVEDDKVGTDM   191 201    1.000 0.079970481
#> 585                  VVEDDKVGTDM   191 201    5.000 0.041800948
#> 586                  VVEDDKVGTDM   191 201   25.000 0.063265722
#> 587                  VVEDDKVGTDM   191 201  150.000 0.094631662
#> 588                  VVEDDKVGTDM   191 201 1440.000 0.008922125
#> 589                 VVEDDKVGTDML   191 202    0.000          NA
#> 590                 VVEDDKVGTDML   191 202    0.167 0.030789761
#> 591                 VVEDDKVGTDML   191 202    1.000 0.120723114
#> 592                 VVEDDKVGTDML   191 202    5.000 0.051645020
#> 593                 VVEDDKVGTDML   191 202   25.000 0.092256914
#> 594                 VVEDDKVGTDML   191 202  150.000 0.069663591
#> 595                 VVEDDKVGTDML   191 202 1440.000 0.045300847
#> 596                VVEDDKVGTDMLE   191 203    0.000          NA
#> 597                VVEDDKVGTDMLE   191 203    0.167 0.073863746
#> 598                VVEDDKVGTDMLE   191 203    1.000 0.121715334
#> 599                VVEDDKVGTDMLE   191 203    5.000 0.127434354
#> 600                VVEDDKVGTDMLE   191 203   25.000 0.075830676
#> 601                VVEDDKVGTDMLE   191 203  150.000 0.109204936
#> 602                VVEDDKVGTDMLE   191 203 1440.000 0.109238419
#> 603                     DDKVGTDM   194 201    0.000          NA
#> 604                     DDKVGTDM   194 201    0.167 0.027356018
#> 605                     DDKVGTDM   194 201    1.000 0.092010993
#> 606                     DDKVGTDM   194 201    5.000 0.027895076
#> 607                     DDKVGTDM   194 201   25.000 0.123317250
#> 608                     DDKVGTDM   194 201  150.000 0.104691695
#> 609                     DDKVGTDM   194 201 1440.000 0.069359635
#> 610                    DDKVGTDML   194 202    0.000          NA
#> 611                    DDKVGTDML   194 202    0.167 0.024262963
#> 612                    DDKVGTDML   194 202    1.000 0.021195070
#> 613                    DDKVGTDML   194 202    5.000 0.074491925
#> 614                    DDKVGTDML   194 202   25.000 0.057369251
#> 615                    DDKVGTDML   194 202  150.000 0.089152100
#> 616                    DDKVGTDML   194 202 1440.000 0.028379150
#> 617                      DKVGTDM   195 201    0.000          NA
#> 618                      DKVGTDM   195 201    0.167 0.009838963
#> 619                      DKVGTDM   195 201    1.000 0.045600645
#> 620                      DKVGTDM   195 201    5.000 0.046636695
#> 621                      DKVGTDM   195 201   25.000 0.067467588
#> 622                      DKVGTDM   195 201  150.000 0.095413627
#> 623                      DKVGTDM   195 201 1440.000 0.002457390
#> 624                     DKVGTDML   195 202    0.000          NA
#> 625                     DKVGTDML   195 202    0.167 0.026991093
#> 626                     DKVGTDML   195 202    1.000 0.035637427
#> 627                     DKVGTDML   195 202    5.000 0.071279088
#> 628                     DKVGTDML   195 202   25.000 0.014005934
#> 629                     DKVGTDML   195 202  150.000 0.041078552
#> 630                     DKVGTDML   195 202 1440.000 0.008828944
#> 631                       VGTDML   197 202    0.000          NA
#> 632                       VGTDML   197 202    0.167 0.058737637
#> 633                       VGTDML   197 202    1.000 0.059732581
#> 634                       VGTDML   197 202    5.000 0.112357356
#> 635                       VGTDML   197 202   25.000 0.059132047
#> 636                       VGTDML   197 202  150.000 0.025280332
#> 637                       VGTDML   197 202 1440.000 0.025438896
#> 638                       LEEQIT   202 207    0.000          NA
#> 639                       LEEQIT   202 207    0.167 0.011644464
#> 640                       LEEQIT   202 207    1.000 0.043563419
#> 641                       LEEQIT   202 207    5.000 0.063526606
#> 642                       LEEQIT   202 207   25.000 0.092724128
#> 643                       LEEQIT   202 207  150.000 0.023586109
#> 644                       LEEQIT   202 207 1440.000 0.028520757
#> 645                      LEEQITA   202 208    0.000          NA
#> 646                      LEEQITA   202 208    0.167 0.025171639
#> 647                      LEEQITA   202 208    1.000 0.039977061
#> 648                      LEEQITA   202 208    5.000 0.100149246
#> 649                      LEEQITA   202 208   25.000 0.157368400
#> 650                      LEEQITA   202 208  150.000 0.038398565
#> 651                      LEEQITA   202 208 1440.000 0.017833219
#> 652                     LEEQITAF   202 209    0.000          NA
#> 653                     LEEQITAF   202 209    0.167 0.010292275
#> 654                     LEEQITAF   202 209    1.000 0.091745917
#> 655                     LEEQITAF   202 209    5.000 0.105475623
#> 656                     LEEQITAF   202 209   25.000 0.128916062
#> 657                     LEEQITAF   202 209  150.000 0.067306017
#> 658                     LEEQITAF   202 209 1440.000 0.030135747
#> 659                       EEQITA   203 208    0.000          NA
#> 660                       EEQITA   203 208    0.167 0.023299801
#> 661                       EEQITA   203 208    1.000 0.034362194
#> 662                       EEQITA   203 208    5.000 0.093814510
#> 663                       EEQITA   203 208   25.000 0.118828249
#> 664                       EEQITA   203 208  150.000 0.022956657
#> 665                       EEQITA   203 208 1440.000 0.027323313
#> 666                      EEQITAF   203 209    0.000          NA
#> 667                      EEQITAF   203 209    0.167 0.014496628
#> 668                      EEQITAF   203 209    1.000 0.012809780
#> 669                      EEQITAF   203 209    5.000 0.101447216
#> 670                      EEQITAF   203 209   25.000 0.174079957
#> 671                      EEQITAF   203 209  150.000 0.039879622
#> 672                      EEQITAF   203 209 1440.000 0.040923774
#> 673                     EEQITAFE   203 210    0.000          NA
#> 674                     EEQITAFE   203 210    0.167 0.074294178
#> 675                     EEQITAFE   203 210    1.000 0.078790045
#> 676                     EEQITAFE   203 210    5.000 0.170961763
#> 677                     EEQITAFE   203 210   25.000 0.028171285
#> 678                     EEQITAFE   203 210  150.000 0.042420099
#> 679                     EEQITAFE   203 210 1440.000 0.014226182
#> 680                       EQITAF   204 209    0.000          NA
#> 681                       EQITAF   204 209    0.167 0.014516394
#> 682                       EQITAF   204 209    1.000 0.051031554
#> 683                       EQITAF   204 209    5.000 0.066632056
#> 684                       EQITAF   204 209   25.000 0.194812174
#> 685                       EQITAF   204 209  150.000 0.024923733
#> 686                       EQITAF   204 209 1440.000 0.013712932
#> 687                        QITAF   205 209    0.000          NA
#> 688                        QITAF   205 209    0.167 0.020577761
#> 689                        QITAF   205 209    1.000 0.022448467
#> 690                        QITAF   205 209    5.000 0.071992608
#> 691                        QITAF   205 209   25.000 0.100414088
#> 692                        QITAF   205 209  150.000 0.067495298
#> 693                        QITAF   205 209 1440.000 0.013988249
#> 694                     FEDYVQSM   209 216    0.000          NA
#> 695                     FEDYVQSM   209 216    0.167 0.041182668
#> 696                     FEDYVQSM   209 216    1.000 0.041148785
#> 697                     FEDYVQSM   209 216    5.000 0.088548530
#> 698                     FEDYVQSM   209 216   25.000 0.133297100
#> 699                     FEDYVQSM   209 216  150.000 0.035029402
#> 700                     FEDYVQSM   209 216 1440.000 0.010903076
#> 701                      EDYVQSM   210 216    0.000          NA
#> 702                      EDYVQSM   210 216    0.167 0.021740803
#> 703                      EDYVQSM   210 216    1.000 0.039957680
#> 704                      EDYVQSM   210 216    5.000 0.102158172
#> 705                      EDYVQSM   210 216   25.000 0.071841343
#> 706                      EDYVQSM   210 216  150.000 0.021151942
#> 707                      EDYVQSM   210 216 1440.000 0.036142276
#> 708                     EDYVQSMD   210 217    0.000          NA
#> 709                     EDYVQSMD   210 217    0.167 0.107593892
#> 710                     EDYVQSMD   210 217    1.000 0.116922081
#> 711                     EDYVQSMD   210 217    5.000 0.066007859
#> 712                     EDYVQSMD   210 217   25.000 0.211968734
#> 713                     EDYVQSMD   210 217  150.000 0.135087550
#> 714                     EDYVQSMD   210 217 1440.000 0.020350570
#> 715                    EDYVQSMDV   210 218    0.000          NA
#> 716                    EDYVQSMDV   210 218    0.167 0.125416426
#> 717                    EDYVQSMDV   210 218    1.000 0.135030274
#> 718                    EDYVQSMDV   210 218    5.000 0.004321960
#> 719                    EDYVQSMDV   210 218   25.000 0.112518466
#> 720                    EDYVQSMDV   210 218  150.000 0.102171170
#> 721                    EDYVQSMDV   210 218 1440.000 0.017566551
#> 722                       DYVQSM   211 216    0.000          NA
#> 723                       DYVQSM   211 216    0.167 0.011890431
#> 724                       DYVQSM   211 216    1.000 0.041143698
#> 725                       DYVQSM   211 216    5.000 0.061089346
#> 726                       DYVQSM   211 216   25.000 0.060456378
#> 727                       DYVQSM   211 216  150.000 0.062255093
#> 728                       DYVQSM   211 216 1440.000 0.017644512
#> 729                       YVQSMD   212 217    0.000          NA
#> 730                       YVQSMD   212 217    0.167 0.017062025
#> 731                       YVQSMD   212 217    1.000 0.028788529
#> 732                       YVQSMD   212 217    5.000 0.141280444
#> 733                       YVQSMD   212 217   25.000 0.068776020
#> 734                       YVQSMD   212 217  150.000 0.165307952
#> 735                       YVQSMD   212 217 1440.000 0.103268638
#> 736                        DVAAF   217 221    0.000          NA
#> 737                        DVAAF   217 221    0.167 0.011545544
#> 738                        DVAAF   217 221    1.000 0.030987450
#> 739                        DVAAF   217 221    5.000 0.009917812
#> 740                        DVAAF   217 221   25.000 0.050125114
#> 741                        DVAAF   217 221  150.000 0.017205970
#> 742                        DVAAF   217 221 1440.000 0.009153806