Get residue positions

Get residue positions

Usage

get_residue_positions(dat)

Arguments

dat

data imported by the read_hdx function.

Value

a data.frame object.

Details

Prepares data table for high-resolution values. It creates a long data frame with each position in the sequence accompanied by the amino acid placed in this position.

Examples

get_residue_positions(alpha_dat)
#>     position aa
#> 1          1  G
#> 2          2  F
#> 3          3  G
#> 4          4  D
#> 5          5  L
#> 6          6  K
#> 7          7  S
#> 8          8  P
#> 9          9  A
#> 10        10  G
#> 11        11  L
#> 12        12  Q
#> 13        13  V
#> 14        14  L
#> 15        15  N
#> 16        16  D
#> 17        17  Y
#> 18        18  L
#> 19        19  A
#> 20        20  D
#> 21        21  K
#> 22        22  S
#> 23        23  Y
#> 24        24  I
#> 25        25  E
#> 26        26  G
#> 27        27  Y
#> 28        28  V
#> 29        29  P
#> 30        30  S
#> 31        31  Q
#> 32        32  A
#> 33        33  D
#> 34        34  V
#> 35        35  A
#> 36        36  V
#> 37        37  F
#> 38        38  E
#> 39        39  A
#> 40        40  V
#> 41        41  S
#> 42        42  S
#> 43        43  P
#> 44        44  P
#> 45        45  P
#> 46        46  A
#> 47        47  D
#> 48        48  L
#> 49        49  C
#> 50        50  H
#> 51        51  A
#> 52        52  L
#> 53        53  R
#> 54        54  W
#> 55        55  Y
#> 56        56  N
#> 57        57  H
#> 58        58  I
#> 59        59  K
#> 60        60  S
#> 61        61  Y
#> 62        62  E
#> 63        63  K
#> 64        64  E
#> 65        65  K
#> 66        66  A
#> 67        67  S
#> 68        68  L
#> 69        69  P
#> 70        70  G
#> 71        71  V
#> 72        72  K
#> 73        73  K
#> 74        74  A
#> 75        75  L
#> 76        76  G
#> 77        77  K
#> 78        78  Y
#> 79        79  G
#> 80        80  P
#> 81        81  A
#> 82        82  D
#> 83        83  V
#> 84        84  E
#> 85        85  D
#> 86        86  T
#> 87        87  T
#> 88        88  G
#> 89        89  S
#> 90        90  G
#> 91        91  A
#> 92        92  T
#> 93        93  D
#> 94        94  S
#> 95        95  K
#> 96        96  D
#> 97        97  D
#> 98        98  D
#> 99        99  D
#> 100      100  I
#> 101      101  D
#> 102      102  L
#> 103      103  F
#> 104      104  G
#> 105      105  S
#> 106      106  D
#> 107      107  D
#> 108      108  E
#> 109      109  E
#> 110      110  E
#> 111      111  S
#> 112      112  E
#> 113      113  E
#> 114      114  A
#> 115      115  K
#> 116      116  R
#> 117      117  L
#> 118      118  R
#> 119      119  E
#> 120      120  E
#> 121      121  R
#> 122      122  L
#> 123      123  A
#> 124      124  Q
#> 125      125  Y
#> 126      126  E
#> 127      127  S
#> 128      128  K
#> 129      129  K
#> 130      130  A
#> 131      131  K
#> 132      132  K
#> 133      133  P
#> 134      134  A
#> 135      135  L
#> 136      136  V
#> 137      137  A
#> 138      138  K
#> 139      139  S
#> 140      140  S
#> 141      141  I
#> 142      142  L
#> 143      143  L
#> 144      144  D
#> 145      145  V
#> 146      146  K
#> 147      147  P
#> 148      148  W
#> 149      149  D
#> 150      150  D
#> 151      151  E
#> 152      152  T
#> 153      153  D
#> 154      154  M
#> 155      155  A
#> 156      156  K
#> 157      157  L
#> 158      158  E
#> 159      159  E
#> 160      160  C
#> 161      161  V
#> 162      162  R
#> 163      163  S
#> 164      164  I
#> 165      165  Q
#> 166      166  A
#> 167      167  D
#> 168      168  G
#> 169      169  L
#> 170      170  V
#> 171      171  W
#> 172      172  G
#> 173      173  S
#> 174      174  S
#> 175      175  K
#> 176      176  L
#> 177      177  V
#> 178      178  P
#> 179      179  V
#> 180      180  G
#> 181      181  Y
#> 182      182  G
#> 183      183  I
#> 184      184  K
#> 185      185  K
#> 186      186  L
#> 187      187  Q
#> 188      188  I
#> 189      189  Q
#> 190      190  C
#> 191      191  V
#> 192      192  V
#> 193      193  E
#> 194      194  D
#> 195      195  D
#> 196      196  K
#> 197      197  V
#> 198      198  G
#> 199      199  T
#> 200      200  D
#> 201      201  M
#> 202      202  L
#> 203      203  E
#> 204      204  E
#> 205      205  Q
#> 206      206  I
#> 207      207  T
#> 208      208  A
#> 209      209  F
#> 210      210  E
#> 211      211  D
#> 212      212  Y
#> 213      213  V
#> 214      214  Q
#> 215      215  S
#> 216      216  M
#> 217      217  D
#> 218      218  V
#> 219      219  A
#> 220      220  A
#> 221      221  F