Creation of validated hdx_data class

The wrapper function for the constructor of the hdx_data and its validator. Used in read_hdx function.

Usage

hdx_data(dat, source, has_modification, n_rep, msg = "")

Arguments

dat

data.frame, read from the datafile.

source

character, the source of the datafile.

has_modification

logical, indicator if there are modified peptides in the datafile.

Value

hdx_data object.