Plot replicates histogram

Plot histogram on number of replicates per peptide in one or multiple time point of measurement.

plot_replicate_histogram(rep_dat, time_points = F)

Arguments

rep_dat
replicate data, created by create_replicate_dataset function.

time_points
logical, indicator if the histogram should show values aggregated for time points of measurements.

Value

a ggplot object.

Details

The function shows three versions of replicate histogram, based on supplied rep_dat and time_points. If time_points is selected, the histogram shows the number of replicates for time points of measurement, to easly spot if there were troubles with samples for specific time point of measurement. Then, on the X-axis is Exposure (in minutes) and on the Y-axis number of replicates. If time_points is not selected, onn the X-axis there are peptide ID, and on the Y-axis there are numbers of replicates. If rep_dat contains data from one time point of measurement, the histogram colors reflect the number of replicates to highlight the outliers. If rep_dat contains multiple time point of measurement, the colors help to distinguish between them.

See also

Examples

dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

#> [1] ""

rep_dat <- create_replicate_dataset(dat)

#> Error: Problem with `summarise()` column `n`.
#> i `n = n()`.
#> x nie udało się znaleźć funkcji 'n'
#> i The error occurred in group 1: Sequence = "ARSQKSGIRLQGHF", Exposure = 0, Start = 88, End = 101, ID = 38.

plot_replicate_histogram(rep_dat)

#> Error in plot_replicate_histogram(rep_dat): nie znaleziono obiektu 'rep_dat'

rep_dat <- create_replicate_dataset(dat, time_t = 0.167)

#> Error: Problem with `summarise()` column `n`.
#> i `n = n()`.
#> x nie udało się znaleźć funkcji 'n'
#> i The error occurred in group 1: Sequence = "ARSQKSGIRLQGHF", Exposure = 0.167, Start = 88, End = 101, ID = 38.

plot_replicate_histogram(rep_dat)

#> Error in plot_replicate_histogram(rep_dat): nie znaleziono obiektu 'rep_dat'

plot_replicate_histogram(rep_dat, time_points = T)

#> Error in plot_replicate_histogram(rep_dat, time_points = T): nie znaleziono obiektu 'rep_dat'