Show data on peptide overlap

Presents peptide overlap on protein sequence data, based on the supplied parameters.

show_overlap_data(
  dat,
  protein = dat[["Protein"]][1],
  state = dat[["State"]][1],
  start = min(dat[["Start"]]),
  end = max(dat[["End"]])
)

Arguments

dat
data imported by the read_hdx function.

protein
chosen protein.

state
biological state for chosen protein.

start
start position of chosen protein.

end
end position of chosen protein.

Value

a data.frame object.

Details

The data frame presents all the peptides in given state, with its start and end position on the protein sequence. This data is available in the GUI.

See also

Examples

dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

#> [1] ""

show_overlap_data(dat)

#>     Protein                             Sequence Start End
#> 1  db_CD160                      INITSSASQEGTRLN     1  15
#> 2  db_CD160                               LICTVW    16  21
#> 3  db_CD160                       LICTVWHKKEEAEG    16  29
#> 4  db_CD160                        ICTVWHKKEEAEG    17  29
#> 5  db_CD160                                FVVFL    30  34
#> 6  db_CD160                      LCKDRSGDCSPETSL    34  48
#> 7  db_CD160                     LCKDRSGDCSPETSLK    34  49
#> 8  db_CD160                   LCKDRSGDCSPETSLKQL    34  51
#> 9  db_CD160 LCKDRSGDCSPETSLKQLRLKRDPGIDGVGEISSQL    34  69
#> 10 db_CD160                       CKDRSGDCSPETSL    35  48
#> 11 db_CD160                         DRSGDCSPETSL    37  48
#> 12 db_CD160                        DRSGDCSPETSLK    37  49
#> 13 db_CD160                      DRSGDCSPETSLKQL    37  51
#> 14 db_CD160                 KQLRLKRDPGIDGVGEISSQ    49  68
#> 15 db_CD160                  QLRLKRDPGIDGVGEISSQ    50  68
#> 16 db_CD160                 QLRLKRDPGIDGVGEISSQL    50  69
#> 17 db_CD160                QLRLKRDPGIDGVGEISSQLM    50  70
#> 18 db_CD160                        RLKRDPGIDGVGE    52  64
#> 19 db_CD160                    RLKRDPGIDGVGEISSQ    52  68
#> 20 db_CD160                   RLKRDPGIDGVGEISSQL    52  69
#> 21 db_CD160                  RLKRDPGIDGVGEISSQLM    52  70
#> 22 db_CD160                   LKRDPGIDGVGEISSQLM    53  70
#> 23 db_CD160                     KRDPGIDGVGEISSQL    54  69
#> 24 db_CD160                    KRDPGIDGVGEISSQLM    54  70
#> 25 db_CD160                       PGIDGVGEISSQLM    57  70
#> 26 db_CD160                       MFTISQVTPLHSGT    70  83
#> 27 db_CD160                   MFTISQVTPLHSGTYQCC    70  87
#> 28 db_CD160                                FTISQ    71  75
#> 29 db_CD160                        FTISQVTPLHSGT    71  83
#> 30 db_CD160                      FTISQVTPLHSGTYQ    71  85
#> 31 db_CD160                     FTISQVTPLHSGTYQC    71  86
#> 32 db_CD160                    FTISQVTPLHSGTYQCC    71  87
#> 33 db_CD160                         TISQVTPLHSGT    72  83
#> 34 db_CD160                     TISQVTPLHSGTYQCC    72  87
#> 35 db_CD160                      ISQVTPLHSGTYQCC    73  87
#> 36 db_CD160                        QVTPLHSGTYQCC    75  87
#> 37 db_CD160                         VTPLHSGTYQCC    76  87
#> 38 db_CD160                       ARSQKSGIRLQGHF    88 101
#> 39 db_CD160                                LQGHF    97 101
#> 40 db_CD160                                FSILF   102 106
#> 41 db_CD160                             FSHNEGTL   125 132