Show replicate data

Show histogram data on number of replicates per peptide in one or multiple time point of measurement.

Usage

show_replicate_histogram_data(rep_dat)

Arguments

rep_dat

replicate data, created by create_replicate_dataset function.

Value

a data.frame object.

Details

The function shows the information about number of replicates for peptides in one or multiple time point of measurement, depends on supplied data.

See also

Examples

rep_dat <- create_replicate_dataset(alpha_dat)
show_replicate_histogram_data(rep_dat)
#>                         Sequence Exposure Start End  ID n
#> 1                    GFGDLKSPAGL    0.000     1  11   1 1
#> 2                    GFGDLKSPAGL    0.167     1  11   1 3
#> 3                    GFGDLKSPAGL    1.000     1  11   1 3
#> 4                    GFGDLKSPAGL    5.000     1  11   1 3
#> 5                    GFGDLKSPAGL   25.000     1  11   1 3
#> 6                    GFGDLKSPAGL  150.000     1  11   1 3
#> 7                    GFGDLKSPAGL 1440.000     1  11   1 3
#> 8                     FGDLKSPAGL    0.000     2  11   2 1
#> 9                     FGDLKSPAGL    0.167     2  11   2 3
#> 10                    FGDLKSPAGL    1.000     2  11   2 3
#> 11                    FGDLKSPAGL    5.000     2  11   2 3
#> 12                    FGDLKSPAGL   25.000     2  11   2 3
#> 13                    FGDLKSPAGL  150.000     2  11   2 3
#> 14                    FGDLKSPAGL 1440.000     2  11   2 3
#> 15                     GDLKSPAGL    0.000     3  11   3 1
#> 16                     GDLKSPAGL    0.167     3  11   3 3
#> 17                     GDLKSPAGL    1.000     3  11   3 3
#> 18                     GDLKSPAGL    5.000     3  11   3 3
#> 19                     GDLKSPAGL   25.000     3  11   3 3
#> 20                     GDLKSPAGL  150.000     3  11   3 3
#> 21                     GDLKSPAGL 1440.000     3  11   3 3
#> 22                        LKSPAG    0.000     5  10   4 1
#> 23                        LKSPAG    0.167     5  10   4 3
#> 24                        LKSPAG    1.000     5  10   4 3
#> 25                        LKSPAG    5.000     5  10   4 3
#> 26                        LKSPAG   25.000     5  10   4 3
#> 27                        LKSPAG  150.000     5  10   4 3
#> 28                        LKSPAG 1440.000     5  10   4 3
#> 29                       LKSPAGL    0.000     5  11   5 1
#> 30                       LKSPAGL    0.167     5  11   5 3
#> 31                       LKSPAGL    1.000     5  11   5 3
#> 32                       LKSPAGL    5.000     5  11   5 3
#> 33                       LKSPAGL   25.000     5  11   5 3
#> 34                       LKSPAGL  150.000     5  11   5 3
#> 35                       LKSPAGL 1440.000     5  11   5 3
#> 36                        AGLQVL    0.000     9  14   6 1
#> 37                        AGLQVL    0.167     9  14   6 3
#> 38                        AGLQVL    1.000     9  14   6 3
#> 39                        AGLQVL    5.000     9  14   6 3
#> 40                        AGLQVL   25.000     9  14   6 3
#> 41                        AGLQVL  150.000     9  14   6 3
#> 42                        AGLQVL 1440.000     9  14   6 3
#> 43                         QVLND    0.000    12  16   7 1
#> 44                         QVLND    0.167    12  16   7 3
#> 45                         QVLND    1.000    12  16   7 3
#> 46                         QVLND    5.000    12  16   7 3
#> 47                         QVLND   25.000    12  16   7 3
#> 48                         QVLND  150.000    12  16   7 3
#> 49                         QVLND 1440.000    12  16   7 3
#> 50                        QVLNDY    0.000    12  17   8 1
#> 51                        QVLNDY    0.167    12  17   8 3
#> 52                        QVLNDY    1.000    12  17   8 3
#> 53                        QVLNDY    5.000    12  17   8 3
#> 54                        QVLNDY   25.000    12  17   8 3
#> 55                        QVLNDY  150.000    12  17   8 3
#> 56                        QVLNDY 1440.000    12  17   8 3
#> 57                        NDYLAD    0.000    15  20   9 1
#> 58                        NDYLAD    0.167    15  20   9 3
#> 59                        NDYLAD    1.000    15  20   9 3
#> 60                        NDYLAD    5.000    15  20   9 3
#> 61                        NDYLAD   25.000    15  20   9 3
#> 62                        NDYLAD  150.000    15  20   9 3
#> 63                        NDYLAD 1440.000    15  20   9 3
#> 64                       NDYLADK    0.000    15  21  10 1
#> 65                       NDYLADK    0.167    15  21  10 3
#> 66                       NDYLADK    1.000    15  21  10 3
#> 67                       NDYLADK    5.000    15  21  10 3
#> 68                       NDYLADK   25.000    15  21  10 3
#> 69                       NDYLADK  150.000    15  21  10 3
#> 70                       NDYLADK 1440.000    15  21  10 3
#> 71                        YLADKS    0.000    17  22  11 1
#> 72                        YLADKS    0.167    17  22  11 3
#> 73                        YLADKS    1.000    17  22  11 3
#> 74                        YLADKS    5.000    17  22  11 3
#> 75                        YLADKS   25.000    17  22  11 3
#> 76                        YLADKS  150.000    17  22  11 3
#> 77                        YLADKS 1440.000    17  22  11 3
#> 78                       YLADKSY    0.000    17  23  12 1
#> 79                       YLADKSY    0.167    17  23  12 3
#> 80                       YLADKSY    1.000    17  23  12 3
#> 81                       YLADKSY    5.000    17  23  12 3
#> 82                       YLADKSY   25.000    17  23  12 3
#> 83                       YLADKSY  150.000    17  23  12 3
#> 84                       YLADKSY 1440.000    17  23  12 3
#> 85             YLADKSYIEGYVPSQAD    0.000    17  33  13 1
#> 86             YLADKSYIEGYVPSQAD    0.167    17  33  13 3
#> 87             YLADKSYIEGYVPSQAD    1.000    17  33  13 3
#> 88             YLADKSYIEGYVPSQAD    5.000    17  33  13 3
#> 89             YLADKSYIEGYVPSQAD   25.000    17  33  13 3
#> 90             YLADKSYIEGYVPSQAD  150.000    17  33  13 3
#> 91             YLADKSYIEGYVPSQAD 1440.000    17  33  13 3
#> 92               ADKSYIEGYVPSQAD    0.000    19  33  14 1
#> 93               ADKSYIEGYVPSQAD    0.167    19  33  14 3
#> 94               ADKSYIEGYVPSQAD    1.000    19  33  14 3
#> 95               ADKSYIEGYVPSQAD    5.000    19  33  14 3
#> 96               ADKSYIEGYVPSQAD   25.000    19  33  14 3
#> 97               ADKSYIEGYVPSQAD  150.000    19  33  14 3
#> 98               ADKSYIEGYVPSQAD 1440.000    19  33  14 3
#> 99                 KSYIEGYVPSQAD    0.000    21  33  15 1
#> 100                KSYIEGYVPSQAD    0.167    21  33  15 3
#> 101                KSYIEGYVPSQAD    1.000    21  33  15 3
#> 102                KSYIEGYVPSQAD    5.000    21  33  15 3
#> 103                KSYIEGYVPSQAD   25.000    21  33  15 3
#> 104                KSYIEGYVPSQAD  150.000    21  33  15 3
#> 105                KSYIEGYVPSQAD 1440.000    21  33  15 3
#> 106                        SYIEG    0.000    22  26  16 1
#> 107                        SYIEG    0.167    22  26  16 3
#> 108                        SYIEG    1.000    22  26  16 3
#> 109                        SYIEG    5.000    22  26  16 3
#> 110                        SYIEG   25.000    22  26  16 3
#> 111                        SYIEG  150.000    22  26  16 3
#> 112                        SYIEG 1440.000    22  26  16 3
#> 113                 SYIEGYVPSQAD    0.000    22  33  17 1
#> 114                 SYIEGYVPSQAD    0.167    22  33  17 3
#> 115                 SYIEGYVPSQAD    1.000    22  33  17 3
#> 116                 SYIEGYVPSQAD    5.000    22  33  17 3
#> 117                 SYIEGYVPSQAD   25.000    22  33  17 3
#> 118                 SYIEGYVPSQAD  150.000    22  33  17 3
#> 119                 SYIEGYVPSQAD 1440.000    22  33  17 3
#> 120                  YIEGYVPSQAD    0.000    23  33  18 1
#> 121                  YIEGYVPSQAD    0.167    23  33  18 3
#> 122                  YIEGYVPSQAD    1.000    23  33  18 3
#> 123                  YIEGYVPSQAD    5.000    23  33  18 3
#> 124                  YIEGYVPSQAD   25.000    23  33  18 3
#> 125                  YIEGYVPSQAD  150.000    23  33  18 3
#> 126                  YIEGYVPSQAD 1440.000    23  33  18 3
#> 127                       IEGYVP    0.000    24  29  19 1
#> 128                       IEGYVP    0.167    24  29  19 3
#> 129                       IEGYVP    1.000    24  29  19 3
#> 130                       IEGYVP    5.000    24  29  19 3
#> 131                       IEGYVP   25.000    24  29  19 3
#> 132                       IEGYVP  150.000    24  29  19 3
#> 133                       IEGYVP 1440.000    24  29  19 3
#> 134                    IEGYVPSQA    0.000    24  32  20 1
#> 135                    IEGYVPSQA    0.167    24  32  20 3
#> 136                    IEGYVPSQA    1.000    24  32  20 3
#> 137                    IEGYVPSQA    5.000    24  32  20 3
#> 138                    IEGYVPSQA   25.000    24  32  20 3
#> 139                    IEGYVPSQA  150.000    24  32  20 3
#> 140                    IEGYVPSQA 1440.000    24  32  20 3
#> 141                   IEGYVPSQAD    0.000    24  33  21 1
#> 142                   IEGYVPSQAD    0.167    24  33  21 3
#> 143                   IEGYVPSQAD    1.000    24  33  21 3
#> 144                   IEGYVPSQAD    5.000    24  33  21 3
#> 145                   IEGYVPSQAD   25.000    24  33  21 3
#> 146                   IEGYVPSQAD  150.000    24  33  21 3
#> 147                   IEGYVPSQAD 1440.000    24  33  21 3
#> 148                      YVPSQAD    0.000    27  33  22 1
#> 149                      YVPSQAD    0.167    27  33  22 3
#> 150                      YVPSQAD    1.000    27  33  22 3
#> 151                      YVPSQAD    5.000    27  33  22 3
#> 152                      YVPSQAD   25.000    27  33  22 3
#> 153                      YVPSQAD  150.000    27  33  22 3
#> 154                      YVPSQAD 1440.000    27  33  22 3
#> 155                        VAVFE    0.000    34  38  23 1
#> 156                        VAVFE    0.167    34  38  23 3
#> 157                        VAVFE    1.000    34  38  23 3
#> 158                        VAVFE    5.000    34  38  23 3
#> 159                        VAVFE   25.000    34  38  23 3
#> 160                        VAVFE  150.000    34  38  23 3
#> 161                        VAVFE 1440.000    34  38  23 3
#> 162                  FEAVSSPPPAD    0.000    37  47  24 1
#> 163                  FEAVSSPPPAD    0.167    37  47  24 3
#> 164                  FEAVSSPPPAD    1.000    37  47  24 3
#> 165                  FEAVSSPPPAD    5.000    37  47  24 3
#> 166                  FEAVSSPPPAD   25.000    37  47  24 3
#> 167                  FEAVSSPPPAD  150.000    37  47  24 3
#> 168                  FEAVSSPPPAD 1440.000    37  47  24 3
#> 169                 FEAVSSPPPADL    0.000    37  48  25 1
#> 170                 FEAVSSPPPADL    0.167    37  48  25 3
#> 171                 FEAVSSPPPADL    1.000    37  48  25 3
#> 172                 FEAVSSPPPADL    5.000    37  48  25 3
#> 173                 FEAVSSPPPADL   25.000    37  48  25 3
#> 174                 FEAVSSPPPADL  150.000    37  48  25 3
#> 175                 FEAVSSPPPADL 1440.000    37  48  25 3
#> 176               FEAVSSPPPADLCH    0.000    37  50  26 1
#> 177               FEAVSSPPPADLCH    0.167    37  50  26 3
#> 178               FEAVSSPPPADLCH    1.000    37  50  26 3
#> 179               FEAVSSPPPADLCH    5.000    37  50  26 3
#> 180               FEAVSSPPPADLCH   25.000    37  50  26 3
#> 181               FEAVSSPPPADLCH  150.000    37  50  26 3
#> 182               FEAVSSPPPADLCH 1440.000    37  50  26 3
#> 183              FEAVSSPPPADLCHA    0.000    37  51  27 1
#> 184              FEAVSSPPPADLCHA    0.167    37  51  27 3
#> 185              FEAVSSPPPADLCHA    1.000    37  51  27 3
#> 186              FEAVSSPPPADLCHA    5.000    37  51  27 3
#> 187              FEAVSSPPPADLCHA   25.000    37  51  27 3
#> 188              FEAVSSPPPADLCHA  150.000    37  51  27 3
#> 189              FEAVSSPPPADLCHA 1440.000    37  51  27 3
#> 190                  EAVSSPPPADL    0.000    38  48  28 1
#> 191                  EAVSSPPPADL    0.167    38  48  28 3
#> 192                  EAVSSPPPADL    1.000    38  48  28 3
#> 193                  EAVSSPPPADL    5.000    38  48  28 3
#> 194                  EAVSSPPPADL   25.000    38  48  28 3
#> 195                  EAVSSPPPADL  150.000    38  48  28 3
#> 196                  EAVSSPPPADL 1440.000    38  48  28 3
#> 197                EAVSSPPPADLCH    0.000    38  50  29 1
#> 198                EAVSSPPPADLCH    0.167    38  50  29 3
#> 199                EAVSSPPPADLCH    1.000    38  50  29 3
#> 200                EAVSSPPPADLCH    5.000    38  50  29 3
#> 201                EAVSSPPPADLCH   25.000    38  50  29 3
#> 202                EAVSSPPPADLCH  150.000    38  50  29 3
#> 203                EAVSSPPPADLCH 1440.000    38  50  29 3
#> 204               EAVSSPPPADLCHA    0.000    38  51  30 1
#> 205               EAVSSPPPADLCHA    0.167    38  51  30 3
#> 206               EAVSSPPPADLCHA    1.000    38  51  30 3
#> 207               EAVSSPPPADLCHA    5.000    38  51  30 3
#> 208               EAVSSPPPADLCHA   25.000    38  51  30 3
#> 209               EAVSSPPPADLCHA  150.000    38  51  30 3
#> 210               EAVSSPPPADLCHA 1440.000    38  51  30 3
#> 211                   AVSSPPPADL    0.000    39  48  31 1
#> 212                   AVSSPPPADL    0.167    39  48  31 3
#> 213                   AVSSPPPADL    1.000    39  48  31 3
#> 214                   AVSSPPPADL    5.000    39  48  31 3
#> 215                   AVSSPPPADL   25.000    39  48  31 3
#> 216                   AVSSPPPADL  150.000    39  48  31 3
#> 217                   AVSSPPPADL 1440.000    39  48  31 3
#> 218                    VSSPPPADL    0.000    40  48  32 1
#> 219                    VSSPPPADL    0.167    40  48  32 3
#> 220                    VSSPPPADL    1.000    40  48  32 3
#> 221                    VSSPPPADL    5.000    40  48  32 3
#> 222                    VSSPPPADL   25.000    40  48  32 3
#> 223                    VSSPPPADL  150.000    40  48  32 3
#> 224                    VSSPPPADL 1440.000    40  48  32 3
#> 225                  VSSPPPADLCH    0.000    40  50  33 1
#> 226                  VSSPPPADLCH    0.167    40  50  33 3
#> 227                  VSSPPPADLCH    1.000    40  50  33 3
#> 228                  VSSPPPADLCH    5.000    40  50  33 3
#> 229                  VSSPPPADLCH   25.000    40  50  33 3
#> 230                  VSSPPPADLCH  150.000    40  50  33 3
#> 231                  VSSPPPADLCH 1440.000    40  50  33 3
#> 232                      ALRWYNH    0.000    51  57  34 1
#> 233                      ALRWYNH    0.167    51  57  34 3
#> 234                      ALRWYNH    1.000    51  57  34 3
#> 235                      ALRWYNH    5.000    51  57  34 3
#> 236                      ALRWYNH   25.000    51  57  34 3
#> 237                      ALRWYNH  150.000    51  57  34 3
#> 238                      ALRWYNH 1440.000    51  57  34 3
#> 239                       LRWYNH    0.000    52  57  35 1
#> 240                       LRWYNH    0.167    52  57  35 3
#> 241                       LRWYNH    1.000    52  57  35 3
#> 242                       LRWYNH    5.000    52  57  35 3
#> 243                       LRWYNH   25.000    52  57  35 3
#> 244                       LRWYNH  150.000    52  57  35 3
#> 245                       LRWYNH 1440.000    52  57  35 3
#> 246             IKSYEKEKASLPGVKK    0.000    58  73  36 1
#> 247             IKSYEKEKASLPGVKK    0.167    58  73  36 3
#> 248             IKSYEKEKASLPGVKK    1.000    58  73  36 3
#> 249             IKSYEKEKASLPGVKK    5.000    58  73  36 3
#> 250             IKSYEKEKASLPGVKK   25.000    58  73  36 3
#> 251             IKSYEKEKASLPGVKK  150.000    58  73  36 3
#> 252             IKSYEKEKASLPGVKK 1440.000    58  73  36 3
#> 253        IKSYEKEKASLPGVKKALGKY    0.000    58  78  37 1
#> 254        IKSYEKEKASLPGVKKALGKY    0.167    58  78  37 3
#> 255        IKSYEKEKASLPGVKKALGKY    1.000    58  78  37 3
#> 256        IKSYEKEKASLPGVKKALGKY    5.000    58  78  37 3
#> 257        IKSYEKEKASLPGVKKALGKY   25.000    58  78  37 3
#> 258        IKSYEKEKASLPGVKKALGKY  150.000    58  78  37 3
#> 259        IKSYEKEKASLPGVKKALGKY 1440.000    58  78  37 3
#> 260              EKASLPGVKKALGKY    0.000    64  78  38 1
#> 261              EKASLPGVKKALGKY    0.167    64  78  38 3
#> 262              EKASLPGVKKALGKY    1.000    64  78  38 3
#> 263              EKASLPGVKKALGKY    5.000    64  78  38 3
#> 264              EKASLPGVKKALGKY   25.000    64  78  38 3
#> 265              EKASLPGVKKALGKY  150.000    64  78  38 3
#> 266              EKASLPGVKKALGKY 1440.000    64  78  38 3
#> 267          EKASLPGVKKALGKYGPAD    0.000    64  82  39 1
#> 268          EKASLPGVKKALGKYGPAD    0.167    64  82  39 3
#> 269          EKASLPGVKKALGKYGPAD    1.000    64  82  39 3
#> 270          EKASLPGVKKALGKYGPAD    5.000    64  82  39 3
#> 271          EKASLPGVKKALGKYGPAD   25.000    64  82  39 3
#> 272          EKASLPGVKKALGKYGPAD  150.000    64  82  39 3
#> 273          EKASLPGVKKALGKYGPAD 1440.000    64  82  39 3
#> 274        EKASLPGVKKALGKYGPADVE    0.000    64  84  40 1
#> 275        EKASLPGVKKALGKYGPADVE    0.167    64  84  40 3
#> 276        EKASLPGVKKALGKYGPADVE    1.000    64  84  40 3
#> 277        EKASLPGVKKALGKYGPADVE    5.000    64  84  40 3
#> 278        EKASLPGVKKALGKYGPADVE   25.000    64  84  40 3
#> 279        EKASLPGVKKALGKYGPADVE  150.000    64  84  40 3
#> 280        EKASLPGVKKALGKYGPADVE 1440.000    64  84  40 3
#> 281            ASLPGVKKALGKYGPAD    0.000    66  82  41 1
#> 282            ASLPGVKKALGKYGPAD    0.167    66  82  41 3
#> 283            ASLPGVKKALGKYGPAD    1.000    66  82  41 3
#> 284            ASLPGVKKALGKYGPAD    5.000    66  82  41 3
#> 285            ASLPGVKKALGKYGPAD   25.000    66  82  41 2
#> 286            ASLPGVKKALGKYGPAD  150.000    66  82  41 3
#> 287            ASLPGVKKALGKYGPAD 1440.000    66  82  41 3
#> 288                  ALGKYGPADVE    0.000    74  84  42 1
#> 289                  ALGKYGPADVE    0.167    74  84  42 3
#> 290                  ALGKYGPADVE    1.000    74  84  42 3
#> 291                  ALGKYGPADVE    5.000    74  84  42 3
#> 292                  ALGKYGPADVE   25.000    74  84  42 3
#> 293                  ALGKYGPADVE  150.000    74  84  42 3
#> 294                  ALGKYGPADVE 1440.000    74  84  42 3
#> 295         VEDTTGSGATDSKDDDDIDL    0.000    83 102  43 1
#> 296         VEDTTGSGATDSKDDDDIDL    0.167    83 102  43 3
#> 297         VEDTTGSGATDSKDDDDIDL    1.000    83 102  43 3
#> 298         VEDTTGSGATDSKDDDDIDL    5.000    83 102  43 3
#> 299         VEDTTGSGATDSKDDDDIDL   25.000    83 102  43 3
#> 300         VEDTTGSGATDSKDDDDIDL  150.000    83 102  43 3
#> 301         VEDTTGSGATDSKDDDDIDL 1440.000    83 102  43 3
#> 302           DTTGSGATDSKDDDDIDL    0.000    85 102  44 1
#> 303           DTTGSGATDSKDDDDIDL    0.167    85 102  44 3
#> 304           DTTGSGATDSKDDDDIDL    1.000    85 102  44 3
#> 305           DTTGSGATDSKDDDDIDL    5.000    85 102  44 3
#> 306           DTTGSGATDSKDDDDIDL   25.000    85 102  44 3
#> 307           DTTGSGATDSKDDDDIDL  150.000    85 102  44 3
#> 308           DTTGSGATDSKDDDDIDL 1440.000    85 102  44 3
#> 309            TTGSGATDSKDDDDIDL    0.000    86 102  45 1
#> 310            TTGSGATDSKDDDDIDL    0.167    86 102  45 3
#> 311            TTGSGATDSKDDDDIDL    1.000    86 102  45 3
#> 312            TTGSGATDSKDDDDIDL    5.000    86 102  45 3
#> 313            TTGSGATDSKDDDDIDL   25.000    86 102  45 3
#> 314            TTGSGATDSKDDDDIDL  150.000    86 102  45 3
#> 315            TTGSGATDSKDDDDIDL 1440.000    86 102  45 3
#> 316                       DDDIDL    0.000    97 102  46 1
#> 317                       DDDIDL    0.167    97 102  46 3
#> 318                       DDDIDL    1.000    97 102  46 3
#> 319                       DDDIDL    5.000    97 102  46 3
#> 320                       DDDIDL   25.000    97 102  46 3
#> 321                       DDDIDL  150.000    97 102  46 3
#> 322                       DDDIDL 1440.000    97 102  46 3
#> 323                       LFGSDD    0.000   102 107  47 1
#> 324                       LFGSDD    0.167   102 107  47 3
#> 325                       LFGSDD    1.000   102 107  47 3
#> 326                       LFGSDD    5.000   102 107  47 3
#> 327                       LFGSDD   25.000   102 107  47 3
#> 328                       LFGSDD  150.000   102 107  47 3
#> 329                       LFGSDD 1440.000   102 107  47 3
#> 330                        FGSDD    0.000   103 107  48 1
#> 331                        FGSDD    0.167   103 107  48 3
#> 332                        FGSDD    1.000   103 107  48 3
#> 333                        FGSDD    5.000   103 107  48 3
#> 334                        FGSDD   25.000   103 107  48 3
#> 335                        FGSDD  150.000   103 107  48 3
#> 336                        FGSDD 1440.000   103 107  48 3
#> 337            FGSDDEEESEEAKRLRE    0.000   103 119  49 1
#> 338            FGSDDEEESEEAKRLRE    0.167   103 119  49 3
#> 339            FGSDDEEESEEAKRLRE    1.000   103 119  49 3
#> 340            FGSDDEEESEEAKRLRE    5.000   103 119  49 3
#> 341            FGSDDEEESEEAKRLRE   25.000   103 119  49 3
#> 342            FGSDDEEESEEAKRLRE  150.000   103 119  49 3
#> 343            FGSDDEEESEEAKRLRE 1440.000   103 119  49 3
#> 344           FGSDDEEESEEAKRLREE    0.000   103 120  50 1
#> 345           FGSDDEEESEEAKRLREE    0.167   103 120  50 3
#> 346           FGSDDEEESEEAKRLREE    1.000   103 120  50 3
#> 347           FGSDDEEESEEAKRLREE    5.000   103 120  50 3
#> 348           FGSDDEEESEEAKRLREE   25.000   103 120  50 3
#> 349           FGSDDEEESEEAKRLREE  150.000   103 120  50 3
#> 350           FGSDDEEESEEAKRLREE 1440.000   103 120  50 3
#> 351         FGSDDEEESEEAKRLREERL    0.000   103 122  51 1
#> 352         FGSDDEEESEEAKRLREERL    0.167   103 122  51 3
#> 353         FGSDDEEESEEAKRLREERL    1.000   103 122  51 3
#> 354         FGSDDEEESEEAKRLREERL    5.000   103 122  51 3
#> 355         FGSDDEEESEEAKRLREERL   25.000   103 122  51 3
#> 356         FGSDDEEESEEAKRLREERL  150.000   103 122  51 3
#> 357         FGSDDEEESEEAKRLREERL 1440.000   103 122  51 3
#> 358           DEEESEEAKRLREERLAQ    0.000   107 124  52 1
#> 359           DEEESEEAKRLREERLAQ    0.167   107 124  52 3
#> 360           DEEESEEAKRLREERLAQ    1.000   107 124  52 3
#> 361           DEEESEEAKRLREERLAQ    5.000   107 124  52 3
#> 362           DEEESEEAKRLREERLAQ   25.000   107 124  52 3
#> 363           DEEESEEAKRLREERLAQ  150.000   107 124  52 3
#> 364           DEEESEEAKRLREERLAQ 1440.000   107 124  52 3
#> 365            EEESEEAKRLREERLAQ    0.000   108 124  53 1
#> 366            EEESEEAKRLREERLAQ    0.167   108 124  53 3
#> 367            EEESEEAKRLREERLAQ    1.000   108 124  53 3
#> 368            EEESEEAKRLREERLAQ    5.000   108 124  53 3
#> 369            EEESEEAKRLREERLAQ   25.000   108 124  53 3
#> 370            EEESEEAKRLREERLAQ  150.000   108 124  53 3
#> 371            EEESEEAKRLREERLAQ 1440.000   108 124  53 3
#> 372                AQYESKKAKKPAL    0.000   123 135  54 1
#> 373                AQYESKKAKKPAL    0.167   123 135  54 3
#> 374                AQYESKKAKKPAL    1.000   123 135  54 3
#> 375                AQYESKKAKKPAL    5.000   123 135  54 3
#> 376                AQYESKKAKKPAL   25.000   123 135  54 3
#> 377                AQYESKKAKKPAL  150.000   123 135  54 3
#> 378                AQYESKKAKKPAL 1440.000   123 135  54 3
#> 379                  YESKKAKKPAL    0.000   125 135  55 1
#> 380                  YESKKAKKPAL    0.167   125 135  55 3
#> 381                  YESKKAKKPAL    1.000   125 135  55 3
#> 382                  YESKKAKKPAL    5.000   125 135  55 3
#> 383                  YESKKAKKPAL   25.000   125 135  55 3
#> 384                  YESKKAKKPAL  150.000   125 135  55 3
#> 385                  YESKKAKKPAL 1440.000   125 135  55 3
#> 386                      VAKSSIL    0.000   136 142  56 1
#> 387                      VAKSSIL    0.167   136 142  56 3
#> 388                      VAKSSIL    1.000   136 142  56 3
#> 389                      VAKSSIL    5.000   136 142  56 3
#> 390                      VAKSSIL   25.000   136 142  56 3
#> 391                      VAKSSIL  150.000   136 142  56 3
#> 392                      VAKSSIL 1440.000   136 142  56 3
#> 393                     VAKSSILL    0.000   136 143  57 1
#> 394                     VAKSSILL    0.167   136 143  57 3
#> 395                     VAKSSILL    1.000   136 143  57 3
#> 396                     VAKSSILL    5.000   136 143  57 3
#> 397                     VAKSSILL   25.000   136 143  57 3
#> 398                     VAKSSILL  150.000   136 143  57 3
#> 399                     VAKSSILL 1440.000   136 143  57 3
#> 400                       AKSSIL    0.000   137 142  58 1
#> 401                       AKSSIL    0.167   137 142  58 3
#> 402                       AKSSIL    1.000   137 142  58 3
#> 403                       AKSSIL    5.000   137 142  58 3
#> 404                       AKSSIL   25.000   137 142  58 3
#> 405                       AKSSIL  150.000   137 142  58 3
#> 406                       AKSSIL 1440.000   137 142  58 3
#> 407                     LDVKPWDD    0.000   143 150  59 1
#> 408                     LDVKPWDD    0.167   143 150  59 3
#> 409                     LDVKPWDD    1.000   143 150  59 3
#> 410                     LDVKPWDD    5.000   143 150  59 3
#> 411                     LDVKPWDD   25.000   143 150  59 3
#> 412                     LDVKPWDD  150.000   143 150  59 3
#> 413                     LDVKPWDD 1440.000   143 150  59 3
#> 414                  LDVKPWDDETD    0.000   143 153  60 1
#> 415                  LDVKPWDDETD    0.167   143 153  60 3
#> 416                  LDVKPWDDETD    1.000   143 153  60 3
#> 417                  LDVKPWDDETD    5.000   143 153  60 3
#> 418                  LDVKPWDDETD   25.000   143 153  60 3
#> 419                  LDVKPWDDETD  150.000   143 153  60 3
#> 420                  LDVKPWDDETD 1440.000   143 153  60 3
#> 421                  DVKPWDDETDM    0.000   144 154  61 1
#> 422                  DVKPWDDETDM    0.167   144 154  61 3
#> 423                  DVKPWDDETDM    1.000   144 154  61 3
#> 424                  DVKPWDDETDM    5.000   144 154  61 3
#> 425                  DVKPWDDETDM   25.000   144 154  61 3
#> 426                  DVKPWDDETDM  150.000   144 154  61 3
#> 427                  DVKPWDDETDM 1440.000   144 154  61 3
#> 428                       MAKLEE    0.000   154 159  62 1
#> 429                       MAKLEE    0.167   154 159  62 3
#> 430                       MAKLEE    1.000   154 159  62 3
#> 431                       MAKLEE    5.000   154 159  62 3
#> 432                       MAKLEE   25.000   154 159  62 3
#> 433                       MAKLEE  150.000   154 159  62 3
#> 434                       MAKLEE 1440.000   154 159  62 3
#> 435                        AKLEE    0.000   155 159  63 1
#> 436                        AKLEE    0.167   155 159  63 3
#> 437                        AKLEE    1.000   155 159  63 3
#> 438                        AKLEE    5.000   155 159  63 3
#> 439                        AKLEE   25.000   155 159  63 3
#> 440                        AKLEE  150.000   155 159  63 3
#> 441                        AKLEE 1440.000   155 159  63 3
#> 442                      CVRSIQA    0.000   160 166  64 1
#> 443                      CVRSIQA    0.167   160 166  64 3
#> 444                      CVRSIQA    1.000   160 166  64 3
#> 445                      CVRSIQA    5.000   160 166  64 2
#> 446                      CVRSIQA   25.000   160 166  64 1
#> 447                      CVRSIQA  150.000   160 166  64 2
#> 448                      CVRSIQA 1440.000   160 166  64 3
#> 449                    CVRSIQADG    0.000   160 168  65 1
#> 450                    CVRSIQADG    0.167   160 168  65 3
#> 451                    CVRSIQADG    1.000   160 168  65 3
#> 452                    CVRSIQADG    5.000   160 168  65 3
#> 453                    CVRSIQADG   25.000   160 168  65 3
#> 454                    CVRSIQADG  150.000   160 168  65 3
#> 455                    CVRSIQADG 1440.000   160 168  65 3
#> 456                   CVRSIQADGL    0.000   160 169  66 1
#> 457                   CVRSIQADGL    0.167   160 169  66 3
#> 458                   CVRSIQADGL    1.000   160 169  66 3
#> 459                   CVRSIQADGL    5.000   160 169  66 3
#> 460                   CVRSIQADGL   25.000   160 169  66 3
#> 461                   CVRSIQADGL  150.000   160 169  66 3
#> 462                   CVRSIQADGL 1440.000   160 169  66 3
#> 463            CVRSIQADGLVWGSSKL    0.000   160 176  67 1
#> 464            CVRSIQADGLVWGSSKL    0.167   160 176  67 3
#> 465            CVRSIQADGLVWGSSKL    1.000   160 176  67 3
#> 466            CVRSIQADGLVWGSSKL    5.000   160 176  67 3
#> 467            CVRSIQADGLVWGSSKL   25.000   160 176  67 3
#> 468            CVRSIQADGLVWGSSKL  150.000   160 176  67 3
#> 469            CVRSIQADGLVWGSSKL 1440.000   160 176  67 3
#> 470                    VRSIQADGL    0.000   161 169  68 1
#> 471                    VRSIQADGL    0.167   161 169  68 3
#> 472                    VRSIQADGL    1.000   161 169  68 3
#> 473                    VRSIQADGL    5.000   161 169  68 3
#> 474                    VRSIQADGL   25.000   161 169  68 3
#> 475                    VRSIQADGL  150.000   161 169  68 3
#> 476                    VRSIQADGL 1440.000   161 169  68 3
#> 477             VRSIQADGLVWGSSKL    0.000   161 176  69 1
#> 478             VRSIQADGLVWGSSKL    0.167   161 176  69 3
#> 479             VRSIQADGLVWGSSKL    1.000   161 176  69 3
#> 480             VRSIQADGLVWGSSKL    5.000   161 176  69 3
#> 481             VRSIQADGLVWGSSKL   25.000   161 176  69 3
#> 482             VRSIQADGLVWGSSKL  150.000   161 176  69 3
#> 483             VRSIQADGLVWGSSKL 1440.000   161 176  69 3
#> 484                      SIQADGL    0.000   163 169  70 1
#> 485                      SIQADGL    0.167   163 169  70 3
#> 486                      SIQADGL    1.000   163 169  70 3
#> 487                      SIQADGL    5.000   163 169  70 3
#> 488                      SIQADGL   25.000   163 169  70 3
#> 489                      SIQADGL  150.000   163 169  70 3
#> 490                      SIQADGL 1440.000   163 169  70 3
#> 491               SIQADGLVWGSSKL    0.000   163 176  71 1
#> 492               SIQADGLVWGSSKL    0.167   163 176  71 3
#> 493               SIQADGLVWGSSKL    1.000   163 176  71 3
#> 494               SIQADGLVWGSSKL    5.000   163 176  71 3
#> 495               SIQADGLVWGSSKL   25.000   163 176  71 3
#> 496               SIQADGLVWGSSKL  150.000   163 176  71 3
#> 497               SIQADGLVWGSSKL 1440.000   163 176  71 3
#> 498 SIQADGLVWGSSKLVPVGYGIKKLQIQC    0.000   163 190  72 1
#> 499 SIQADGLVWGSSKLVPVGYGIKKLQIQC    0.167   163 190  72 3
#> 500 SIQADGLVWGSSKLVPVGYGIKKLQIQC    1.000   163 190  72 3
#> 501 SIQADGLVWGSSKLVPVGYGIKKLQIQC    5.000   163 190  72 3
#> 502 SIQADGLVWGSSKLVPVGYGIKKLQIQC   25.000   163 190  72 3
#> 503 SIQADGLVWGSSKLVPVGYGIKKLQIQC  150.000   163 190  72 3
#> 504 SIQADGLVWGSSKLVPVGYGIKKLQIQC 1440.000   163 190  72 3
#> 505         ADGLVWGSSKLVPVGYGIKK    0.000   166 185  73 1
#> 506         ADGLVWGSSKLVPVGYGIKK    0.167   166 185  73 3
#> 507         ADGLVWGSSKLVPVGYGIKK    1.000   166 185  73 3
#> 508         ADGLVWGSSKLVPVGYGIKK    5.000   166 185  73 3
#> 509         ADGLVWGSSKLVPVGYGIKK   25.000   166 185  73 3
#> 510         ADGLVWGSSKLVPVGYGIKK  150.000   166 185  73 3
#> 511         ADGLVWGSSKLVPVGYGIKK 1440.000   166 185  73 3
#> 512                     DGLVWGSS    0.000   167 174  74 1
#> 513                     DGLVWGSS    0.167   167 174  74 3
#> 514                     DGLVWGSS    1.000   167 174  74 3
#> 515                     DGLVWGSS    5.000   167 174  74 3
#> 516                     DGLVWGSS   25.000   167 174  74 3
#> 517                     DGLVWGSS  150.000   167 174  74 3
#> 518                     DGLVWGSS 1440.000   167 174  74 3
#> 519                   DGLVWGSSKL    0.000   167 176  75 1
#> 520                   DGLVWGSSKL    0.167   167 176  75 2
#> 521                   DGLVWGSSKL    1.000   167 176  75 3
#> 522                   DGLVWGSSKL    5.000   167 176  75 3
#> 523                   DGLVWGSSKL   25.000   167 176  75 3
#> 524                   DGLVWGSSKL  150.000   167 176  75 3
#> 525                   DGLVWGSSKL 1440.000   167 176  75 3
#> 526              DGLVWGSSKLVPVGY    0.000   167 181  76 1
#> 527              DGLVWGSSKLVPVGY    0.167   167 181  76 3
#> 528              DGLVWGSSKLVPVGY    1.000   167 181  76 3
#> 529              DGLVWGSSKLVPVGY    5.000   167 181  76 3
#> 530              DGLVWGSSKLVPVGY   25.000   167 181  76 3
#> 531              DGLVWGSSKLVPVGY  150.000   167 181  76 3
#> 532              DGLVWGSSKLVPVGY 1440.000   167 181  76 3
#> 533             DGLVWGSSKLVPVGYG    0.000   167 182  77 1
#> 534             DGLVWGSSKLVPVGYG    0.167   167 182  77 3
#> 535             DGLVWGSSKLVPVGYG    1.000   167 182  77 3
#> 536             DGLVWGSSKLVPVGYG    5.000   167 182  77 3
#> 537             DGLVWGSSKLVPVGYG   25.000   167 182  77 3
#> 538             DGLVWGSSKLVPVGYG  150.000   167 182  77 3
#> 539             DGLVWGSSKLVPVGYG 1440.000   167 182  77 3
#> 540                LVWGSSKLVPVGY    0.000   169 181  78 1
#> 541                LVWGSSKLVPVGY    0.167   169 181  78 3
#> 542                LVWGSSKLVPVGY    1.000   169 181  78 3
#> 543                LVWGSSKLVPVGY    5.000   169 181  78 3
#> 544                LVWGSSKLVPVGY   25.000   169 181  78 3
#> 545                LVWGSSKLVPVGY  150.000   169 181  78 3
#> 546                LVWGSSKLVPVGY 1440.000   169 181  78 3
#> 547             KLVPVGYGIKKLQIQC    0.000   175 190  79 1
#> 548             KLVPVGYGIKKLQIQC    0.167   175 190  79 3
#> 549             KLVPVGYGIKKLQIQC    1.000   175 190  79 3
#> 550             KLVPVGYGIKKLQIQC    5.000   175 190  79 3
#> 551             KLVPVGYGIKKLQIQC   25.000   175 190  79 3
#> 552             KLVPVGYGIKKLQIQC  150.000   175 190  79 3
#> 553             KLVPVGYGIKKLQIQC 1440.000   175 190  79 3
#> 554               VPVGYGIKKLQIQC    0.000   177 190  80 1
#> 555               VPVGYGIKKLQIQC    0.167   177 190  80 3
#> 556               VPVGYGIKKLQIQC    1.000   177 190  80 3
#> 557               VPVGYGIKKLQIQC    5.000   177 190  80 3
#> 558               VPVGYGIKKLQIQC   25.000   177 190  80 3
#> 559               VPVGYGIKKLQIQC  150.000   177 190  80 3
#> 560               VPVGYGIKKLQIQC 1440.000   177 190  80 3
#> 561                       KLQIQC    0.000   185 190  81 1
#> 562                       KLQIQC    0.167   185 190  81 3
#> 563                       KLQIQC    1.000   185 190  81 3
#> 564                       KLQIQC    5.000   185 190  81 3
#> 565                       KLQIQC   25.000   185 190  81 3
#> 566                       KLQIQC  150.000   185 190  81 3
#> 567                       KLQIQC 1440.000   185 190  81 3
#> 568                 CVVEDDKVGTDM    0.000   190 201  82 1
#> 569                 CVVEDDKVGTDM    0.167   190 201  82 3
#> 570                 CVVEDDKVGTDM    1.000   190 201  82 3
#> 571                 CVVEDDKVGTDM    5.000   190 201  82 3
#> 572                 CVVEDDKVGTDM   25.000   190 201  82 3
#> 573                 CVVEDDKVGTDM  150.000   190 201  82 3
#> 574                 CVVEDDKVGTDM 1440.000   190 201  82 3
#> 575                CVVEDDKVGTDML    0.000   190 202  83 1
#> 576                CVVEDDKVGTDML    0.167   190 202  83 3
#> 577                CVVEDDKVGTDML    1.000   190 202  83 3
#> 578                CVVEDDKVGTDML    5.000   190 202  83 3
#> 579                CVVEDDKVGTDML   25.000   190 202  83 3
#> 580                CVVEDDKVGTDML  150.000   190 202  83 3
#> 581                CVVEDDKVGTDML 1440.000   190 202  83 3
#> 582                  VVEDDKVGTDM    0.000   191 201  84 1
#> 583                  VVEDDKVGTDM    0.167   191 201  84 3
#> 584                  VVEDDKVGTDM    1.000   191 201  84 3
#> 585                  VVEDDKVGTDM    5.000   191 201  84 3
#> 586                  VVEDDKVGTDM   25.000   191 201  84 3
#> 587                  VVEDDKVGTDM  150.000   191 201  84 3
#> 588                  VVEDDKVGTDM 1440.000   191 201  84 3
#> 589                 VVEDDKVGTDML    0.000   191 202  85 1
#> 590                 VVEDDKVGTDML    0.167   191 202  85 3
#> 591                 VVEDDKVGTDML    1.000   191 202  85 3
#> 592                 VVEDDKVGTDML    5.000   191 202  85 3
#> 593                 VVEDDKVGTDML   25.000   191 202  85 3
#> 594                 VVEDDKVGTDML  150.000   191 202  85 3
#> 595                 VVEDDKVGTDML 1440.000   191 202  85 3
#> 596                VVEDDKVGTDMLE    0.000   191 203  86 1
#> 597                VVEDDKVGTDMLE    0.167   191 203  86 3
#> 598                VVEDDKVGTDMLE    1.000   191 203  86 3
#> 599                VVEDDKVGTDMLE    5.000   191 203  86 3
#> 600                VVEDDKVGTDMLE   25.000   191 203  86 3
#> 601                VVEDDKVGTDMLE  150.000   191 203  86 3
#> 602                VVEDDKVGTDMLE 1440.000   191 203  86 3
#> 603                     DDKVGTDM    0.000   194 201  87 1
#> 604                     DDKVGTDM    0.167   194 201  87 3
#> 605                     DDKVGTDM    1.000   194 201  87 3
#> 606                     DDKVGTDM    5.000   194 201  87 3
#> 607                     DDKVGTDM   25.000   194 201  87 3
#> 608                     DDKVGTDM  150.000   194 201  87 3
#> 609                     DDKVGTDM 1440.000   194 201  87 3
#> 610                    DDKVGTDML    0.000   194 202  88 1
#> 611                    DDKVGTDML    0.167   194 202  88 3
#> 612                    DDKVGTDML    1.000   194 202  88 3
#> 613                    DDKVGTDML    5.000   194 202  88 3
#> 614                    DDKVGTDML   25.000   194 202  88 3
#> 615                    DDKVGTDML  150.000   194 202  88 3
#> 616                    DDKVGTDML 1440.000   194 202  88 3
#> 617                      DKVGTDM    0.000   195 201  89 1
#> 618                      DKVGTDM    0.167   195 201  89 3
#> 619                      DKVGTDM    1.000   195 201  89 3
#> 620                      DKVGTDM    5.000   195 201  89 3
#> 621                      DKVGTDM   25.000   195 201  89 3
#> 622                      DKVGTDM  150.000   195 201  89 3
#> 623                      DKVGTDM 1440.000   195 201  89 3
#> 624                     DKVGTDML    0.000   195 202  90 1
#> 625                     DKVGTDML    0.167   195 202  90 3
#> 626                     DKVGTDML    1.000   195 202  90 3
#> 627                     DKVGTDML    5.000   195 202  90 3
#> 628                     DKVGTDML   25.000   195 202  90 3
#> 629                     DKVGTDML  150.000   195 202  90 3
#> 630                     DKVGTDML 1440.000   195 202  90 3
#> 631                       VGTDML    0.000   197 202  91 1
#> 632                       VGTDML    0.167   197 202  91 3
#> 633                       VGTDML    1.000   197 202  91 3
#> 634                       VGTDML    5.000   197 202  91 3
#> 635                       VGTDML   25.000   197 202  91 3
#> 636                       VGTDML  150.000   197 202  91 3
#> 637                       VGTDML 1440.000   197 202  91 3
#> 638                       LEEQIT    0.000   202 207  92 1
#> 639                       LEEQIT    0.167   202 207  92 3
#> 640                       LEEQIT    1.000   202 207  92 3
#> 641                       LEEQIT    5.000   202 207  92 3
#> 642                       LEEQIT   25.000   202 207  92 3
#> 643                       LEEQIT  150.000   202 207  92 3
#> 644                       LEEQIT 1440.000   202 207  92 3
#> 645                      LEEQITA    0.000   202 208  93 1
#> 646                      LEEQITA    0.167   202 208  93 3
#> 647                      LEEQITA    1.000   202 208  93 3
#> 648                      LEEQITA    5.000   202 208  93 3
#> 649                      LEEQITA   25.000   202 208  93 3
#> 650                      LEEQITA  150.000   202 208  93 3
#> 651                      LEEQITA 1440.000   202 208  93 3
#> 652                     LEEQITAF    0.000   202 209  94 1
#> 653                     LEEQITAF    0.167   202 209  94 3
#> 654                     LEEQITAF    1.000   202 209  94 3
#> 655                     LEEQITAF    5.000   202 209  94 3
#> 656                     LEEQITAF   25.000   202 209  94 3
#> 657                     LEEQITAF  150.000   202 209  94 3
#> 658                     LEEQITAF 1440.000   202 209  94 3
#> 659                       EEQITA    0.000   203 208  95 1
#> 660                       EEQITA    0.167   203 208  95 3
#> 661                       EEQITA    1.000   203 208  95 3
#> 662                       EEQITA    5.000   203 208  95 3
#> 663                       EEQITA   25.000   203 208  95 3
#> 664                       EEQITA  150.000   203 208  95 3
#> 665                       EEQITA 1440.000   203 208  95 3
#> 666                      EEQITAF    0.000   203 209  96 1
#> 667                      EEQITAF    0.167   203 209  96 3
#> 668                      EEQITAF    1.000   203 209  96 3
#> 669                      EEQITAF    5.000   203 209  96 3
#> 670                      EEQITAF   25.000   203 209  96 3
#> 671                      EEQITAF  150.000   203 209  96 3
#> 672                      EEQITAF 1440.000   203 209  96 3
#> 673                     EEQITAFE    0.000   203 210  97 1
#> 674                     EEQITAFE    0.167   203 210  97 3
#> 675                     EEQITAFE    1.000   203 210  97 3
#> 676                     EEQITAFE    5.000   203 210  97 3
#> 677                     EEQITAFE   25.000   203 210  97 3
#> 678                     EEQITAFE  150.000   203 210  97 3
#> 679                     EEQITAFE 1440.000   203 210  97 3
#> 680                       EQITAF    0.000   204 209  98 1
#> 681                       EQITAF    0.167   204 209  98 3
#> 682                       EQITAF    1.000   204 209  98 3
#> 683                       EQITAF    5.000   204 209  98 3
#> 684                       EQITAF   25.000   204 209  98 3
#> 685                       EQITAF  150.000   204 209  98 3
#> 686                       EQITAF 1440.000   204 209  98 3
#> 687                        QITAF    0.000   205 209  99 1
#> 688                        QITAF    0.167   205 209  99 3
#> 689                        QITAF    1.000   205 209  99 3
#> 690                        QITAF    5.000   205 209  99 3
#> 691                        QITAF   25.000   205 209  99 3
#> 692                        QITAF  150.000   205 209  99 3
#> 693                        QITAF 1440.000   205 209  99 3
#> 694                     FEDYVQSM    0.000   209 216 100 1
#> 695                     FEDYVQSM    0.167   209 216 100 3
#> 696                     FEDYVQSM    1.000   209 216 100 3
#> 697                     FEDYVQSM    5.000   209 216 100 3
#> 698                     FEDYVQSM   25.000   209 216 100 3
#> 699                     FEDYVQSM  150.000   209 216 100 3
#> 700                     FEDYVQSM 1440.000   209 216 100 3
#> 701                      EDYVQSM    0.000   210 216 101 1
#> 702                      EDYVQSM    0.167   210 216 101 3
#> 703                      EDYVQSM    1.000   210 216 101 3
#> 704                      EDYVQSM    5.000   210 216 101 3
#> 705                      EDYVQSM   25.000   210 216 101 3
#> 706                      EDYVQSM  150.000   210 216 101 3
#> 707                      EDYVQSM 1440.000   210 216 101 3
#> 708                     EDYVQSMD    0.000   210 217 102 1
#> 709                     EDYVQSMD    0.167   210 217 102 3
#> 710                     EDYVQSMD    1.000   210 217 102 3
#> 711                     EDYVQSMD    5.000   210 217 102 3
#> 712                     EDYVQSMD   25.000   210 217 102 3
#> 713                     EDYVQSMD  150.000   210 217 102 3
#> 714                     EDYVQSMD 1440.000   210 217 102 3
#> 715                    EDYVQSMDV    0.000   210 218 103 1
#> 716                    EDYVQSMDV    0.167   210 218 103 3
#> 717                    EDYVQSMDV    1.000   210 218 103 3
#> 718                    EDYVQSMDV    5.000   210 218 103 3
#> 719                    EDYVQSMDV   25.000   210 218 103 3
#> 720                    EDYVQSMDV  150.000   210 218 103 3
#> 721                    EDYVQSMDV 1440.000   210 218 103 3
#> 722                       DYVQSM    0.000   211 216 104 1
#> 723                       DYVQSM    0.167   211 216 104 3
#> 724                       DYVQSM    1.000   211 216 104 3
#> 725                       DYVQSM    5.000   211 216 104 3
#> 726                       DYVQSM   25.000   211 216 104 3
#> 727                       DYVQSM  150.000   211 216 104 3
#> 728                       DYVQSM 1440.000   211 216 104 3
#> 729                       YVQSMD    0.000   212 217 105 1
#> 730                       YVQSMD    0.167   212 217 105 3
#> 731                       YVQSMD    1.000   212 217 105 3
#> 732                       YVQSMD    5.000   212 217 105 3
#> 733                       YVQSMD   25.000   212 217 105 3
#> 734                       YVQSMD  150.000   212 217 105 3
#> 735                       YVQSMD 1440.000   212 217 105 3
#> 736                        DVAAF    0.000   217 221 106 1
#> 737                        DVAAF    0.167   217 221 106 3
#> 738                        DVAAF    1.000   217 221 106 3
#> 739                        DVAAF    5.000   217 221 106 3
#> 740                        DVAAF   25.000   217 221 106 3
#> 741                        DVAAF  150.000   217 221 106 3
#> 742                        DVAAF 1440.000   217 221 106 3