Calculate deuterium uptake

Calculates deuteration uptake based on supplied parameters.

calculate_state_uptake(
  dat,
  protein = unique(dat[["Protein"]])[1],
  state = unique(dat[["State"]])[1],
  time_0 = min(dat[["Exposure"]]),
  time_t = unique(dat[["Exposure"]])[3],
  time_100 = max(dat[["Exposure"]]),
  deut_part = 0.9
)

Arguments

dat
data as imported by the read_hdx function

protein
protein included in calculations

state
state included in calculations

time_0
minimal exchange control

time_t
chosen time point

time_100
maximal exchange control

deut_part
percentage of deuterium the protein was exposed to, value in range [0, 1]

Value

a data.frame object

Details

The function calculate_state_uptake calculates deuterium uptake (in different forms) for all of the peptides in given protein in given state based on supplied parameters: `time_0`, `time_100` and `time_t`. All four variants (fractiona) are supplied (mean values and uncertainty). Manual correction of percentage of deuterium the protein was exposed to during the exchange in theoretical calculations is provided.

Methods of calculation and uncertainty are profoundly discussed in the vignette.

See also

Examples

dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

#> [1] ""

calculate_state_uptake(dat)

#> # A tibble: 41 x 15
#>    Protein  Sequence Exposure Start   End State frac_deut_uptake err_frac_deut_u~
#>    <chr>    <chr>       <dbl> <int> <int> <chr>            <dbl>            <dbl>
#>  1 db_CD160 INITSSA~    0.167     1    15 CD160             76.9            0.382
#>  2 db_CD160 LICTVW      0.167    16    21 CD160             25.8            1.27 
#>  3 db_CD160 LICTVWH~    0.167    16    29 CD160             63.8            0.692
#>  4 db_CD160 ICTVWHK~    0.167    17    29 CD160             68.2            0.537
#>  5 db_CD160 FVVFL       0.167    30    34 CD160             18.8            0.778
#>  6 db_CD160 LCKDRSG~    0.167    34    48 CD160             87.9            0.289
#>  7 db_CD160 LCKDRSG~    0.167    34    49 CD160             84.0            0.496
#>  8 db_CD160 LCKDRSG~    0.167    34    51 CD160             79.1            0.344
#>  9 db_CD160 LCKDRSG~    0.167    34    69 CD160             81.5            0.406
#> 10 db_CD160 CKDRSGD~    0.167    35    48 CD160             88.3            0.680
#> # ... with 31 more rows, and 7 more variables: deut_uptake <dbl>,
#> #   err_deut_uptake <dbl>, theo_frac_deut_uptake <dbl>,
#> #   err_theo_frac_deut_uptake <dbl>, theo_deut_uptake <dbl>,
#> #   err_theo_deut_uptake <dbl>, Med_Sequence <dbl>

calculate_state_uptake(dat, protein = "db_CD160", state = "CD160",
                            time_0 = 0.001, time_t = 5.000, time_100 = 1440.000, deut_part = 1)

#> # A tibble: 41 x 15
#>    Protein  Sequence Exposure Start   End State frac_deut_uptake err_frac_deut_u~
#>    <chr>    <chr>       <dbl> <int> <int> <chr>            <dbl>            <dbl>
#>  1 db_CD160 INITSSA~        5     1    15 CD160             87.5            0.594
#>  2 db_CD160 LICTVW          5    16    21 CD160             41.7            0.581
#>  3 db_CD160 LICTVWH~        5    16    29 CD160             72.0            1.07 
#>  4 db_CD160 ICTVWHK~        5    17    29 CD160             75.8            0.774
#>  5 db_CD160 FVVFL           5    30    34 CD160             39.3            1.87 
#>  6 db_CD160 LCKDRSG~        5    34    48 CD160             97.6            0.554
#>  7 db_CD160 LCKDRSG~        5    34    49 CD160             97.2            0.498
#>  8 db_CD160 LCKDRSG~        5    34    51 CD160             95.3            0.359
#>  9 db_CD160 LCKDRSG~        5    34    69 CD160             93.6            0.334
#> 10 db_CD160 CKDRSGD~        5    35    48 CD160             97.2            0.691
#> # ... with 31 more rows, and 7 more variables: deut_uptake <dbl>,
#> #   err_deut_uptake <dbl>, theo_frac_deut_uptake <dbl>,
#> #   err_theo_frac_deut_uptake <dbl>, theo_deut_uptake <dbl>,
#> #   err_theo_deut_uptake <dbl>, Med_Sequence <dbl>