Calculate deuterium uptake

Calculates deuteration uptake based on supplied parameters.

Usage

calculate_state_uptake(
  dat,
  protein = unique(dat[["Protein"]])[1],
  state = unique(dat[["State"]])[1],
  time_0 = min(dat[dat[["Exposure"]] > 0, ][["Exposure"]]),
  time_t = unique(dat[["Exposure"]])[3],
  time_100 = max(dat[["Exposure"]]),
  deut_part = 0.9
)

Arguments

dat

data as imported by the read_hdx function

protein

chosen protein.

state

state included in calculations

time_0

minimal exchange control

time_t

chosen time point

time_100

maximal exchange control

deut_part

percentage of deuterium the protein was exposed to, value in range [0, 1]

Value

a data.frame object

Details

The function calculate_state_uptake calculates deuterium uptake (in different forms) for all of the peptides in given protein in given state based on supplied parameters: `time_0`, `time_100` and `time_t`. All four variants (fractiona) are supplied (mean values and uncertainty). Manual correction of percentage of deuterium the protein was exposed to during the exchange in theoretical calculations is provided.

Methods of calculation and uncertainty are profoundly discussed in the vignette.

See also

Examples

calculate_state_uptake(alpha_dat)
#>       Protein                     Sequence Exposure Start End       State
#> 1   db_eEF1Ba                  GFGDLKSPAGL        1     1  11 ALPHA_Gamma
#> 2   db_eEF1Ba                   FGDLKSPAGL        1     2  11 ALPHA_Gamma
#> 3   db_eEF1Ba                    GDLKSPAGL        1     3  11 ALPHA_Gamma
#> 4   db_eEF1Ba                       LKSPAG        1     5  10 ALPHA_Gamma
#> 5   db_eEF1Ba                      LKSPAGL        1     5  11 ALPHA_Gamma
#> 6   db_eEF1Ba                       AGLQVL        1     9  14 ALPHA_Gamma
#> 7   db_eEF1Ba                        QVLND        1    12  16 ALPHA_Gamma
#> 8   db_eEF1Ba                       QVLNDY        1    12  17 ALPHA_Gamma
#> 9   db_eEF1Ba                       NDYLAD        1    15  20 ALPHA_Gamma
#> 10  db_eEF1Ba                      NDYLADK        1    15  21 ALPHA_Gamma
#> 11  db_eEF1Ba                       YLADKS        1    17  22 ALPHA_Gamma
#> 12  db_eEF1Ba                      YLADKSY        1    17  23 ALPHA_Gamma
#> 13  db_eEF1Ba            YLADKSYIEGYVPSQAD        1    17  33 ALPHA_Gamma
#> 14  db_eEF1Ba              ADKSYIEGYVPSQAD        1    19  33 ALPHA_Gamma
#> 15  db_eEF1Ba                KSYIEGYVPSQAD        1    21  33 ALPHA_Gamma
#> 16  db_eEF1Ba                        SYIEG        1    22  26 ALPHA_Gamma
#> 17  db_eEF1Ba                 SYIEGYVPSQAD        1    22  33 ALPHA_Gamma
#> 18  db_eEF1Ba                  YIEGYVPSQAD        1    23  33 ALPHA_Gamma
#> 19  db_eEF1Ba                       IEGYVP        1    24  29 ALPHA_Gamma
#> 20  db_eEF1Ba                    IEGYVPSQA        1    24  32 ALPHA_Gamma
#> 21  db_eEF1Ba                   IEGYVPSQAD        1    24  33 ALPHA_Gamma
#> 22  db_eEF1Ba                      YVPSQAD        1    27  33 ALPHA_Gamma
#> 23  db_eEF1Ba                        VAVFE        1    34  38 ALPHA_Gamma
#> 24  db_eEF1Ba                  FEAVSSPPPAD        1    37  47 ALPHA_Gamma
#> 25  db_eEF1Ba                 FEAVSSPPPADL        1    37  48 ALPHA_Gamma
#> 26  db_eEF1Ba               FEAVSSPPPADLCH        1    37  50 ALPHA_Gamma
#> 27  db_eEF1Ba              FEAVSSPPPADLCHA        1    37  51 ALPHA_Gamma
#> 28  db_eEF1Ba                  EAVSSPPPADL        1    38  48 ALPHA_Gamma
#> 29  db_eEF1Ba                EAVSSPPPADLCH        1    38  50 ALPHA_Gamma
#> 30  db_eEF1Ba               EAVSSPPPADLCHA        1    38  51 ALPHA_Gamma
#> 31  db_eEF1Ba                   AVSSPPPADL        1    39  48 ALPHA_Gamma
#> 32  db_eEF1Ba                    VSSPPPADL        1    40  48 ALPHA_Gamma
#> 33  db_eEF1Ba                  VSSPPPADLCH        1    40  50 ALPHA_Gamma
#> 34  db_eEF1Ba                      ALRWYNH        1    51  57 ALPHA_Gamma
#> 35  db_eEF1Ba                       LRWYNH        1    52  57 ALPHA_Gamma
#> 36  db_eEF1Ba             IKSYEKEKASLPGVKK        1    58  73 ALPHA_Gamma
#> 37  db_eEF1Ba        IKSYEKEKASLPGVKKALGKY        1    58  78 ALPHA_Gamma
#> 38  db_eEF1Ba              EKASLPGVKKALGKY        1    64  78 ALPHA_Gamma
#> 39  db_eEF1Ba          EKASLPGVKKALGKYGPAD        1    64  82 ALPHA_Gamma
#> 40  db_eEF1Ba        EKASLPGVKKALGKYGPADVE        1    64  84 ALPHA_Gamma
#> 41  db_eEF1Ba            ASLPGVKKALGKYGPAD        1    66  82 ALPHA_Gamma
#> 42  db_eEF1Ba                  ALGKYGPADVE        1    74  84 ALPHA_Gamma
#> 43  db_eEF1Ba         VEDTTGSGATDSKDDDDIDL        1    83 102 ALPHA_Gamma
#> 44  db_eEF1Ba           DTTGSGATDSKDDDDIDL        1    85 102 ALPHA_Gamma
#> 45  db_eEF1Ba            TTGSGATDSKDDDDIDL        1    86 102 ALPHA_Gamma
#> 46  db_eEF1Ba                       DDDIDL        1    97 102 ALPHA_Gamma
#> 47  db_eEF1Ba                       LFGSDD        1   102 107 ALPHA_Gamma
#> 48  db_eEF1Ba                        FGSDD        1   103 107 ALPHA_Gamma
#> 49  db_eEF1Ba            FGSDDEEESEEAKRLRE        1   103 119 ALPHA_Gamma
#> 50  db_eEF1Ba           FGSDDEEESEEAKRLREE        1   103 120 ALPHA_Gamma
#> 51  db_eEF1Ba         FGSDDEEESEEAKRLREERL        1   103 122 ALPHA_Gamma
#> 52  db_eEF1Ba           DEEESEEAKRLREERLAQ        1   107 124 ALPHA_Gamma
#> 53  db_eEF1Ba            EEESEEAKRLREERLAQ        1   108 124 ALPHA_Gamma
#> 54  db_eEF1Ba                AQYESKKAKKPAL        1   123 135 ALPHA_Gamma
#> 55  db_eEF1Ba                  YESKKAKKPAL        1   125 135 ALPHA_Gamma
#> 56  db_eEF1Ba                      VAKSSIL        1   136 142 ALPHA_Gamma
#> 57  db_eEF1Ba                     VAKSSILL        1   136 143 ALPHA_Gamma
#> 58  db_eEF1Ba                       AKSSIL        1   137 142 ALPHA_Gamma
#> 59  db_eEF1Ba                     LDVKPWDD        1   143 150 ALPHA_Gamma
#> 60  db_eEF1Ba                  LDVKPWDDETD        1   143 153 ALPHA_Gamma
#> 61  db_eEF1Ba                  DVKPWDDETDM        1   144 154 ALPHA_Gamma
#> 62  db_eEF1Ba                       MAKLEE        1   154 159 ALPHA_Gamma
#> 63  db_eEF1Ba                        AKLEE        1   155 159 ALPHA_Gamma
#> 64  db_eEF1Ba                      CVRSIQA        1   160 166 ALPHA_Gamma
#> 65  db_eEF1Ba                    CVRSIQADG        1   160 168 ALPHA_Gamma
#> 66  db_eEF1Ba                   CVRSIQADGL        1   160 169 ALPHA_Gamma
#> 67  db_eEF1Ba            CVRSIQADGLVWGSSKL        1   160 176 ALPHA_Gamma
#> 68  db_eEF1Ba                    VRSIQADGL        1   161 169 ALPHA_Gamma
#> 69  db_eEF1Ba             VRSIQADGLVWGSSKL        1   161 176 ALPHA_Gamma
#> 70  db_eEF1Ba                      SIQADGL        1   163 169 ALPHA_Gamma
#> 71  db_eEF1Ba               SIQADGLVWGSSKL        1   163 176 ALPHA_Gamma
#> 72  db_eEF1Ba SIQADGLVWGSSKLVPVGYGIKKLQIQC        1   163 190 ALPHA_Gamma
#> 73  db_eEF1Ba         ADGLVWGSSKLVPVGYGIKK        1   166 185 ALPHA_Gamma
#> 74  db_eEF1Ba                     DGLVWGSS        1   167 174 ALPHA_Gamma
#> 75  db_eEF1Ba                   DGLVWGSSKL        1   167 176 ALPHA_Gamma
#> 76  db_eEF1Ba              DGLVWGSSKLVPVGY        1   167 181 ALPHA_Gamma
#> 77  db_eEF1Ba             DGLVWGSSKLVPVGYG        1   167 182 ALPHA_Gamma
#> 78  db_eEF1Ba                LVWGSSKLVPVGY        1   169 181 ALPHA_Gamma
#> 79  db_eEF1Ba             KLVPVGYGIKKLQIQC        1   175 190 ALPHA_Gamma
#> 80  db_eEF1Ba               VPVGYGIKKLQIQC        1   177 190 ALPHA_Gamma
#> 81  db_eEF1Ba                       KLQIQC        1   185 190 ALPHA_Gamma
#> 82  db_eEF1Ba                 CVVEDDKVGTDM        1   190 201 ALPHA_Gamma
#> 83  db_eEF1Ba                CVVEDDKVGTDML        1   190 202 ALPHA_Gamma
#> 84  db_eEF1Ba                  VVEDDKVGTDM        1   191 201 ALPHA_Gamma
#> 85  db_eEF1Ba                 VVEDDKVGTDML        1   191 202 ALPHA_Gamma
#> 86  db_eEF1Ba                VVEDDKVGTDMLE        1   191 203 ALPHA_Gamma
#> 87  db_eEF1Ba                     DDKVGTDM        1   194 201 ALPHA_Gamma
#> 88  db_eEF1Ba                    DDKVGTDML        1   194 202 ALPHA_Gamma
#> 89  db_eEF1Ba                      DKVGTDM        1   195 201 ALPHA_Gamma
#> 90  db_eEF1Ba                     DKVGTDML        1   195 202 ALPHA_Gamma
#> 91  db_eEF1Ba                       VGTDML        1   197 202 ALPHA_Gamma
#> 92  db_eEF1Ba                       LEEQIT        1   202 207 ALPHA_Gamma
#> 93  db_eEF1Ba                      LEEQITA        1   202 208 ALPHA_Gamma
#> 94  db_eEF1Ba                     LEEQITAF        1   202 209 ALPHA_Gamma
#> 95  db_eEF1Ba                       EEQITA        1   203 208 ALPHA_Gamma
#> 96  db_eEF1Ba                      EEQITAF        1   203 209 ALPHA_Gamma
#> 97  db_eEF1Ba                     EEQITAFE        1   203 210 ALPHA_Gamma
#> 98  db_eEF1Ba                       EQITAF        1   204 209 ALPHA_Gamma
#> 99  db_eEF1Ba                        QITAF        1   205 209 ALPHA_Gamma
#> 100 db_eEF1Ba                     FEDYVQSM        1   209 216 ALPHA_Gamma
#> 101 db_eEF1Ba                      EDYVQSM        1   210 216 ALPHA_Gamma
#> 102 db_eEF1Ba                     EDYVQSMD        1   210 217 ALPHA_Gamma
#> 103 db_eEF1Ba                    EDYVQSMDV        1   210 218 ALPHA_Gamma
#> 104 db_eEF1Ba                       DYVQSM        1   211 216 ALPHA_Gamma
#> 105 db_eEF1Ba                       YVQSMD        1   212 217 ALPHA_Gamma
#> 106 db_eEF1Ba                        DVAAF        1   217 221 ALPHA_Gamma
#>     MaxUptake Modification frac_deut_uptake err_frac_deut_uptake deut_uptake
#> 1           9           NA       31.2643389            0.9842724  1.17794140
#> 2           8           NA       55.1613850            1.8959005  1.29287731
#> 3           7           NA       55.0346782            1.6368399  1.32747818
#> 4           4           NA       49.7806260            2.2253434  0.85772467
#> 5           5           NA       49.1100880            1.4982868  0.96223900
#> 6           5           NA       20.4940756            1.7389609  0.66915500
#> 7           4           NA        3.3863046            1.6271906  0.06605100
#> 8           5           NA       -0.6573346            1.7899549 -0.01647167
#> 9           5           NA        4.0259054            1.4813734  0.10537500
#> 10          6           NA       -0.3213617            2.1250358 -0.01049686
#> 11          5           NA       14.8318327            0.4586685  0.42026700
#> 12          6           NA       11.0861929            1.0086615  0.36022042
#> 13         15           NA       10.9479977            0.5963176  0.78101430
#> 14         13           NA       12.9632109            0.7835793  0.77511800
#> 15         11           NA       11.0922191            0.4254848  0.53681142
#> 16          4           NA       -1.7731683            2.0753776 -0.02445400
#> 17         10           NA       11.1277187            0.4472028  0.44182336
#> 18          9           NA       14.9071370            0.8810571  0.50587163
#> 19          4           NA       20.8184911            2.6332837  0.23714600
#> 20          7           NA       25.3791233            2.5538585  0.51938138
#> 21          8           NA       17.1139970            0.8974462  0.49304289
#> 22          5           NA       14.3573237            1.5443779  0.32695933
#> 23          4           NA        0.7649875            0.7887228  0.02075900
#> 24          7           NA       25.2753618            2.8650626  0.52113353
#> 25          8           NA       17.8269860            0.8315175  0.51421536
#> 26         10           NA       21.7004946            2.4847439  0.83198467
#> 27         11           NA       18.9550709            2.1295328  0.82422600
#> 28          7           NA       21.5128057            1.1171849  0.47263550
#> 29          9           NA       24.2960940            0.8089167  0.76458969
#> 30         10           NA       21.5434653            2.4217614  0.82422067
#> 31          6           NA       31.1349206            3.6786304  0.49691875
#> 32          5           NA       37.3996293            4.2759989  0.48500500
#> 33          7           NA       32.4989177            1.9006141  0.77359867
#> 34          6           NA        0.7435766            1.1844531  0.02367716
#> 35          5           NA        2.6885063            0.7184425  0.06242483
#> 36         14           NA       42.0770796            3.4076269  0.61512900
#> 37         19           NA       34.9762367            6.8003347  0.44478709
#> 38         13           NA      209.4437844          132.4122591  0.35439145
#> 39         16           NA       62.3648538           28.8928445  0.17697717
#> 40         18           NA       78.2361052           12.5460553  0.32848517
#> 41         14           NA       68.5909284           38.9767952  0.09640219
#> 42          9           NA      129.2654145           23.5025232  0.19352067
#> 43         19           NA      217.8570044          315.5099281  0.12560612
#> 44         17           NA      279.1276438          374.9110872  0.17302269
#> 45         16           NA      221.7570455          155.1724772  0.23249600
#> 46          5           NA       92.1602096           37.5989256  0.11536400
#> 47          5           NA       19.7137132           32.7110491  0.01528233
#> 48          4           NA       -7.5706783          386.0952083  0.00135500
#> 49         16           NA      103.2512642           13.9759949  0.38241650
#> 50         17           NA       99.7401175           19.6072405  0.42587153
#> 51         19           NA       91.3416127            9.6042497  0.44582806
#> 52         17           NA      110.0994635           50.1093647  0.48065340
#> 53         16           NA      101.0107793           10.8936081  0.56633362
#> 54         11           NA      -73.4151714          256.1318023  0.02571082
#> 55          9           NA     -467.4572701         1351.8327894  0.09075841
#> 56          6           NA       31.8397165            1.4163372  0.81479948
#> 57          7           NA       25.1340026            1.2320334  0.85020531
#> 58          5           NA       34.7610450            1.7663802  0.84137800
#> 59          6           NA       12.0311104            1.5158311  0.29142405
#> 60          9           NA       28.8588355            0.8623717  0.90726410
#> 61          9           NA       40.4137578            1.0292843  1.08659751
#> 62          5           NA        6.3102082            0.8764342  0.19383937
#> 63          4           NA        3.9136956            0.2696633  0.09719500
#> 64          6           NA       25.7978037            0.8830732  0.78111165
#> 65          8           NA       25.3607081            1.7022248  0.75802117
#> 66          9           NA       19.9030511            1.9771431  0.69159500
#> 67         16           NA       26.3624816            0.5600115  1.70103001
#> 68          8           NA       29.8430490            1.2737753  0.78307673
#> 69         15           NA       30.2941953            1.3495986  1.69817349
#> 70          6           NA       51.4344132            1.8849120  0.64483067
#> 71         13           NA       35.6764672            3.9135422  1.54845667
#> 72         26           NA       23.8685083            0.4499614  2.79862947
#> 73         18           NA       31.9074613            1.7559207  2.19814063
#> 74          7           NA       23.3631837            1.4136413  0.60045267
#> 75          9           NA       48.3780011            5.3421358  1.09986249
#> 76         13           NA       25.0038242            1.1856922  1.31639267
#> 77         14           NA       24.9279103            1.6568808  1.37353533
#> 78         11           NA       33.0515424            0.7887827  1.36792600
#> 79         14           NA       16.6974457            0.6149067  1.22502196
#> 80         12           NA       18.4640467            0.8775593  1.19508269
#> 81          5           NA        6.1027176            0.5573114  0.21071233
#> 82         11           NA       27.2842481            1.4898883  1.02035467
#> 83         12           NA       26.3252029            1.5395623  1.19546133
#> 84         10           NA       30.8680417            1.7781210  0.86388942
#> 85         11           NA       28.3408395            2.0074322  1.00577293
#> 86         12           NA       22.4918881            1.8205714  0.97528333
#> 87          7           NA       43.2902834            3.1560428  0.77626340
#> 88          8           NA       30.4157958            0.6560831  0.76119328
#> 89          6           NA       50.8673870            2.1210955  0.63515188
#> 90          7           NA       33.3944630            1.1553022  0.66794015
#> 91          5           NA       17.2345319            4.6710502  0.16432500
#> 92          5           NA        8.3671624            0.8083263  0.26842200
#> 93          6           NA        7.5517677            0.6611047  0.30519500
#> 94          7           NA        7.4730881            1.1596161  0.34329579
#> 95          5           NA        6.1560819            0.7442089  0.19468100
#> 96          6           NA        5.3956536            0.2813607  0.20921162
#> 97          7           NA        8.6601063            1.3355039  0.38933067
#> 98          5           NA        5.5909859            0.9440039  0.18070100
#> 99          4           NA        6.4610022            0.7117242  0.15505733
#> 100         7           NA       20.3357118            0.8428799  0.73371103
#> 101         6           NA       20.0923380            0.7522949  0.68112653
#> 102         7           NA       16.6384879            2.1687750  0.65288600
#> 103         8           NA       23.5303930            2.4150527  0.93196200
#> 104         5           NA       22.9951493            0.8533116  0.65882567
#> 105         5           NA       20.2110261            0.8747460  0.50791400
#> 106         4           NA       37.1741019            1.5457378  0.44440833
#>     err_deut_uptake theo_frac_deut_uptake err_theo_frac_deut_uptake
#> 1       0.035202006            31.7727835                0.42982083
#> 2       0.044252689            44.5169903                0.56944568
#> 3       0.040119708            48.0810128                0.56280453
#> 4       0.038974894            34.9263280                1.04879181
#> 5       0.029799382            37.1101292                0.64319951
#> 6       0.060284573             5.6663511                1.06291837
#> 7       0.031742345            -6.5098572                0.86629048
#> 8       0.044830214            -1.9970641                0.95118050
#> 9       0.040226582            -5.9726322                0.23671199
#> 10      0.069235581             0.1830087                0.58341857
#> 11      0.014239677             0.5497376                0.18006216
#> 12      0.034030169             5.5103023                0.50280928
#> 13      0.043007837            26.4879046                0.29812808
#> 14      0.051320990            30.3347266                0.22974068
#> 15      0.022868528            31.3328808                0.05364419
#> 16      0.028346126             7.4087901                0.53039325
#> 17      0.017575952            35.4656381                0.19336949
#> 18      0.031557990            35.4409479                0.29661201
#> 19      0.031438334            23.6520686                0.74273903
#> 20      0.068836717            41.4464949                0.13516941
#> 21      0.025434233            35.5533986                0.32046310
#> 22      0.038858343            27.1691974                0.46721487
#> 23      0.021508622           -15.7006658                0.35268625
#> 24      0.045189330            37.6570319                0.26599034
#> 25      0.023887689            35.6143846                0.32905524
#> 26      0.095417673            36.2248818                1.04167022
#> 27      0.108916785            31.4381743                0.46384491
#> 28      0.027603555            40.6018445                0.28138333
#> 29      0.023472582            43.8945183                0.18396350
#> 30      0.094182965            34.7517565                0.96130559
#> 31      0.078563476            45.3153978                0.29291825
#> 32      0.053015771            40.0039097                1.09104855
#> 33      0.056018992            43.2182390                0.42591598
#> 34      0.037781801             1.8949155                0.60653188
#> 35      0.017102339            -2.1067524                0.08288239
#> 36      0.055394409            66.2494943                0.35944394
#> 37      0.119389281            70.7826145                0.24456913
#> 38      0.188705602            80.7727894                1.57028882
#> 39      0.077904230            79.0087144                0.32348034
#> 40      0.153034837            80.2886709                0.15122991
#> 41      0.061627127            78.6489250                0.41626749
#> 42      0.031368406            69.5119073                0.31967758
#> 43      0.048270258            56.4062314                0.18226125
#> 44      0.052176170            55.6860885                0.28428408
#> 45      0.056228063            55.4358201                0.37051730
#> 46      0.049354876            34.6861377                0.87857958
#> 47      0.026567248            34.0796176                0.46952326
#> 48      0.065223409            33.7966408                0.86339358
#> 49      0.077170964            68.0396778                0.35001870
#> 50      0.094491258            68.6347419                0.53775748
#> 51      0.110573700            70.7577023                0.23807772
#> 52      0.150899537            83.4209619                0.95188566
#> 53      0.056692788            81.2825077                0.36647342
#> 54      0.071136371            70.6114732                0.60440991
#> 55      0.040210376            66.4687680                0.30807693
#> 56      0.035754938            28.5954007                0.62995789
#> 57      0.042659635            22.7641838                0.63060731
#> 58      0.041043625            26.7060122                0.87984094
#> 59      0.039603889            13.7888414                0.41931430
#> 60      0.033456548            29.3588492                0.08074481
#> 61      0.029487340            36.2621721                0.31210499
#> 62      0.027155266            -3.3665306                0.54186728
#> 63      0.006900296           -13.6721760                0.08561821
#> 64      0.030249713            24.7476916                0.36041314
#> 65      0.057634362            38.7431991                0.51142886
#> 66      0.074366905            34.8411935                0.66560045
#> 67      0.034531415            39.1106147                0.22551821
#> 68      0.034439597            38.1833459                0.40642101
#> 69      0.083806119            40.9750791                0.39767003
#> 70      0.036872173            44.7671580                0.26782656
#> 71      0.182855259            45.4298051                1.32510193
#> 72      0.060749340            41.7049115                0.14632341
#> 73      0.111216318            46.8144127                0.59267757
#> 74      0.045772350            24.9420685                0.17940303
#> 75      0.128021975            48.8782592                1.43999456
#> 76      0.060337470            38.0226716                0.49263127
#> 77      0.091760861            40.7770625                0.68754649
#> 78      0.044018449            42.6715395                0.15230444
#> 79      0.045877233            33.8450768                0.33031121
#> 80      0.059867925            39.2785334                0.44689683
#> 81      0.019383304            -7.2419222                0.39286218
#> 82      0.057104275            34.3061541                0.53602897
#> 83      0.083455963            30.8339071                0.43960056
#> 84      0.053362824            36.3309293                0.50930493
#> 85      0.071930707            33.3256167                0.69894969
#> 86      0.082199910            29.6579393                0.64596982
#> 87      0.055420738            41.3528318                0.83712412
#> 88      0.018600380            37.0996918                0.16873029
#> 89      0.026933400            43.8933565                0.48402513
#> 90      0.025810497            39.1242902                0.32423244
#> 91      0.048366969            36.3082970                0.76083319
#> 92      0.026034368            -6.9528773                0.55488135
#> 93      0.027275000            -4.0713099                0.42433396
#> 94      0.053301795            -1.2471315                0.83471243
#> 95      0.023969711            -8.4678830                0.43768237
#> 96      0.011169045            -6.2428740                0.13596859
#> 97      0.062523318             0.8649369                0.71683876
#> 98      0.030631929            -9.9144608                0.65000539
#> 99      0.017581979           -12.5075026                0.35741673
#> 100     0.033611678            14.6433404                0.37437527
#> 101     0.026263274             8.8173937                0.42412824
#> 102     0.091737340            10.1577887                1.06376739
#> 103     0.106399334            22.3934372                1.07495265
#> 104     0.024726411             7.9689224                0.52406058
#> 105     0.019320905             9.3566837                0.36668879
#> 106     0.019092072            20.8100448                0.49337146
#>     theo_deut_uptake err_theo_deut_uptake Med_Sequence
#> 1        2.592322154          0.035068821          6.0
#> 2        3.228546010          0.041298425          6.5
#> 3        3.051144990          0.035714685          7.0
#> 4        1.266496863          0.038031239          7.5
#> 5        1.682107197          0.029154588          8.0
#> 6        0.256841197          0.048179370         11.5
#> 7       -0.236060137          0.031413385         14.0
#> 8       -0.090521803          0.043114578         14.5
#> 9       -0.270724137          0.010729549         17.5
#> 10       0.009954378          0.031733845         18.0
#> 11       0.024918197          0.008161757         19.5
#> 12       0.299721482          0.027349269         20.0
#> 13       3.601886811          0.040540149         25.0
#> 14       3.574988393          0.027075249         26.0
#> 15       3.124526484          0.005349418         27.0
#> 16       0.268657197          0.019233095         24.0
#> 17       3.215133250          0.017529889         27.5
#> 18       2.891605461          0.024200394         28.0
#> 19       0.857670197          0.026933168         26.5
#> 20       2.630128980          0.008577637         28.0
#> 21       2.578471331          0.023241236         28.5
#> 22       1.231510197          0.021177654         30.0
#> 23      -0.569336803          0.012789092         36.0
#> 24       2.389655659          0.016879326         42.0
#> 25       2.582894273          0.023864371         42.5
#> 26       3.283962393          0.094432491         43.5
#> 27       3.135026393          0.046254786         44.0
#> 28       2.576528802          0.017856141         43.0
#> 29       3.581327150          0.015009471         44.0
#> 30       3.150416393          0.087147046         44.5
#> 31       2.464837230          0.015932682         43.5
#> 32       1.813274863          0.049454439         44.0
#> 33       2.742561060          0.027027954         45.0
#> 34       0.103070005          0.032991046         54.0
#> 35      -0.095493697          0.003756847         54.5
#> 36       8.408176146          0.045619488         65.5
#> 37      12.191900219          0.042125632         68.0
#> 38       9.519182030          0.185060653         71.0
#> 39      11.460041122          0.046920116         73.0
#> 40      13.101408023          0.024677514         74.0
#> 41       9.981872649          0.052831352         74.0
#> 42       5.671434393          0.026082300         79.0
#> 43       9.715650507          0.031393457         92.5
#> 44       8.581966861          0.043811958         93.5
#> 45       8.040844393          0.053742724         94.0
#> 46       1.572233863          0.039823764         99.5
#> 47       1.544741863          0.021282288        104.5
#> 48       1.225532197          0.031308337        105.0
#> 49       9.869006375          0.050769447        111.0
#> 50      10.577526570          0.082875580        111.5
#> 51      12.187609229          0.041007524        112.5
#> 52      12.856279716          0.146698240        115.5
#> 53      11.789849871          0.053156168        116.0
#> 54       7.041402273          0.060271980        129.0
#> 55       5.423146504          0.025135810        130.0
#> 56       1.555387611          0.034265255        139.0
#> 57       1.444579085          0.040017342        139.5
#> 58       1.210515197          0.039880938        139.5
#> 59       0.750015475          0.022807733        146.5
#> 60       2.395370715          0.006587920        148.0
#> 61       2.958608653          0.025464457        149.0
#> 62      -0.152596217          0.024561457        156.5
#> 63      -0.495779803          0.003104684        157.0
#> 64       1.346099441          0.019603927        163.0
#> 65       2.809808122          0.037090819        164.0
#> 66       2.842671863          0.054305938        164.5
#> 67       5.672909080          0.032710922        168.0
#> 68       2.769205381          0.029475239        165.0
#> 69       5.571886455          0.054076095        168.5
#> 70       2.435016863          0.014567871        166.0
#> 71       5.353963727          0.156165048        169.5
#> 72       9.829959991          0.034488821        176.5
#> 73       7.639119125          0.096712407        175.5
#> 74       1.582784197          0.011384633        170.5
#> 75       3.987947550          0.117488284        171.5
#> 76       4.481023060          0.058057258        174.0
#> 77       5.175295727          0.087261224        174.5
#> 78       4.255221727          0.015187855        175.0
#> 79       4.295510035          0.041922053        182.5
#> 80       4.272948989          0.048616056        183.5
#> 81      -0.328257803          0.017807437        187.5
#> 82       3.421022393          0.053453008        195.5
#> 83       3.354293060          0.047822324        196.0
#> 84       3.293576121          0.046170979        196.0
#> 85       3.323242843          0.069699523        196.5
#> 86       3.226364393          0.070272381        197.0
#> 87       2.624185269          0.053122572        197.5
#> 88       2.690614555          0.012236979        198.0
#> 89       2.387488242          0.026327545        198.0
#> 90       2.482765545          0.020575278        198.5
#> 91       1.645762197          0.034486622        199.5
#> 92      -0.315156137          0.025151352        204.5
#> 93      -0.221450470          0.023080767        205.0
#> 94      -0.079140995          0.052969530        205.5
#> 95      -0.383827470          0.019839022        205.5
#> 96      -0.339568203          0.007395730        206.0
#> 97       0.054887530          0.045489453        206.5
#> 98      -0.449397137          0.029463081        206.5
#> 99      -0.453546470          0.012960629        207.0
#> 100      0.929243211          0.023757262        212.5
#> 101      0.479603871          0.023069577        213.0
#> 102      0.644597197          0.067504995        213.5
#> 103      1.624059530          0.077959765        214.0
#> 104      0.361210863          0.023754325        213.5
#> 105      0.424114530          0.016621065        214.5
#> 106      0.754612863          0.017890613        219.0