Show uptake data

Show uptake data

show_uptake_data(
  uptake_dat,
  theoretical = FALSE,
  fractional = FALSE,
  renamed = TRUE
)

Arguments

uptake_dat
data produced by create_uptake_dataset function or create_state_uptake_dataset.

theoretical
logical, determines if values are theoretical.

fractional
logical, determines if values are fractional.

renamed
logical, determines if the names of the columns are renamed to user-friendly ones. Currently FALSE not implemented.

Value

a data.frame object.

Details

This function subsets the dataset based on provided criteria, rounds the numerical values (4 places) and changes the column names to user-friendly ones. This data is available in the GUI.

See also

Examples

dat <- read_hdx(system.file(package = "HaDeX", "HaDeX/data/KD_180110_CD160_HVEM.csv"))

#> [1] ""

uptake_dat <- create_uptake_dataset(dat)
show_uptake_data(uptake_dat)

#> # A tibble: 410 x 9
#>    Protein  Sequence           ID State Start   End Exposure `DU [Da]` `U(DU) [Da]`
#>    <chr>    <chr>           <int> <chr> <int> <int>    <dbl>     <dbl>        <dbl>
#>  1 db_CD160 INITSSASQEGTRLN     1 CD160     1    15    0.167      8.00       0.0283
#>  2 db_CD160 INITSSASQEGTRLN     1 CD160     1    15    1          8.77       0.0557
#>  3 db_CD160 INITSSASQEGTRLN     1 CD160     1    15    5          9.12       0.0529
#>  4 db_CD160 INITSSASQEGTRLN     1 CD160     1    15   25          9.71       0.0496
#>  5 db_CD160 INITSSASQEGTRLN     1 CD160     1    15  120          9.95       0.0794
#>  6 db_CD160 INITSSASQEGTRLN     1 CD16~     1    15    0.167      7.68       0.0363
#>  7 db_CD160 INITSSASQEGTRLN     1 CD16~     1    15    1          8.43       0.0794
#>  8 db_CD160 INITSSASQEGTRLN     1 CD16~     1    15    5          8.57       0.0633
#>  9 db_CD160 INITSSASQEGTRLN     1 CD16~     1    15   25          9.06       0.0235
#> 10 db_CD160 INITSSASQEGTRLN     1 CD16~     1    15  120          9.84       0.0321
#> # ... with 400 more rows